% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLADTest.R
\name{ad.test}
\alias{ad.test}
\title{Anderson-Darling Test}
\usage{
ad.test(x, ...)
}
\arguments{
\item{x:}{a FLVector of data values.}
}
\description{
Performs the Anderson-Darling test for the composite hypothesis of normality, see e.g. Thode (2002, Sec. 5.1.4).
The Anderson-Darling test is an EDF omnibus test for the composite hypothesis of normality. The Anderson-Darling test is the recommended EDF test by Stephens (1986). Compared to the Cramer-von Mises test (as second choice) it gives more weight to the tails of the distribution.
}
\examples{
set.seed(200)
a <- as.FL(rnorm(100, mean = 5, sd = 3))
res <- ad.test(a)

}
