% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{as.data.frame.FLSimpleWideTable}
\alias{as.data.frame.FLSimpleWideTable}
\title{Converts FLVector object to vector in R}
\usage{
\method{as.data.frame}{FLSimpleWideTable}(object, mode = "any")
}
\description{
Converts FLVector object to vector in R
}

