% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.FLMatrix.Matrix}
\alias{as.FLMatrix.Matrix}
\title{Casting to FLMatrix}
\usage{
as.FLMatrix.Matrix(object, connection, sparse = TRUE, ...)
}
\arguments{
\item{object}{matrix,vector,data frame,sparseMatrix,FLVector or FLSparseMatrix which
needs to be casted to FLMatrix}

\item{connection}{ODBC connection object}

\item{sparse}{}

\item{nr}{number of rows in resulting FLMatrix}

\item{nc}{number of columns in resulting FLMatrix.
nr and nc inputs are applicable only in case of vector,FLVector}
}
\value{
FLMatrix object after casting.
}
\description{
Converts input \code{m} to FLMatrix object
In addition, one can specify number of rows and columns
of resulting flmatrix object
}

