% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.FLSparseMatrix}
\alias{as.FLSparseMatrix}
\title{Casting to FLSparseMatrix}
\usage{
as.FLSparseMatrix(m, connection)
}
\arguments{
\item{m}{a sparseMatrix in R of type "dgCMatrix" which
needs to be casted to FLSparseMatrix}

\item{connection}{ODBC connection object}
}
\value{
FLSparseMatrix object after casting.
}
\description{
Converts input \code{m} to FLMatrix object
}

