% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix.FLMatrix}
\alias{as.matrix.FLMatrix}
\title{Converts input FLMatrix object to matrix in R}
\usage{
\method{as.matrix}{FLMatrix}(flmatobj1)
}
\description{
Converts input FLMatrix object to matrix in R
}

