% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix.FLSparseMatrix}
\alias{as.matrix.FLSparseMatrix}
\title{Converts FLSparseMatrix object to a matrix in R}
\usage{
\method{as.matrix}{FLSparseMatrix}(object)
}
\description{
Converts FLSparseMatrix object to a matrix in R
}

