% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix.FLVector}
\alias{as.matrix.FLVector}
\title{Converts FLVector object to a matrix in R}
\usage{
\method{as.matrix}{FLVector}(obj)
}
\description{
Converts FLVector object to a matrix in R
}

