% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix}
\alias{as.matrix}
\title{Converts in-database objects to a matrix in R}
\usage{
as.matrix(x, ...)
}
\arguments{
\item{x}{can be FLTable, FLVector or FLMatrix}
}
\value{
R matrix object
}
\description{
Caution: data is fetched into R session
}

