% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix.sparseMatrix}
\alias{as.matrix.sparseMatrix}
\title{Converts input FLMatrix object to matrix in R}
\usage{
\method{as.matrix}{sparseMatrix}(object, sparse = FALSE)
}
\description{
Converts input FLMatrix object to matrix in R
}

