% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.vector.FLMatrix}
\alias{as.vector.FLMatrix}
\title{Converts FLMatrix object to vector in R}
\usage{
as.vector.FLMatrix(object, mode = "any")
}
\description{
Converts FLMatrix object to vector in R
}

