% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{as.vector.FLSimpleVector}
\alias{as.vector.FLSimpleVector}
\title{Converts FLVector object to vector in R}
\usage{
as.vector.FLSimpleVector(object, mode = "any")
}
\description{
Converts FLVector object to vector in R
}

