% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.vector.FLSkalarAggregate}
\alias{as.vector.FLSkalarAggregate}
\title{Converts FLSkalarAggregate object to vector in R}
\usage{
\method{as.vector}{FLSkalarAggregate}(object, mode = "any")
}
\description{
Converts FLSkalarAggregate object to vector in R
}
