% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.vector.FLSparseMatrix}
\alias{as.vector.FLSparseMatrix}
\title{Converts FLSparseMatrix object to vector in R}
\usage{
as.vector.FLSparseMatrix(object, mode = "any")
}
\description{
Converts FLSparseMatrix object to vector in R
}

