% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.vector.FLVector}
\alias{as.vector.FLVector}
\title{Converts FLVector object to vector in R}
\usage{
as.vector.FLVector(object, mode = "any")
}
\description{
Converts FLVector object to vector in R
}

