% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting.R
\name{bind_tbl}
\alias{bind_tbl}
\title{Create a single session, events, or streams tibble from list of user data}
\usage{
bind_tbl(user_list, tbl_name)
}
\arguments{
\item{user_list}{List of user data created via a method such as
\code{\link{read_all}}()}

\item{tbl_name}{Unquoted name of tibble to extract. Must be `session`,
`events`, or `streams`}
}
\description{
Simulation data read in using \code{\link{read_all}}() is available as a user
list, with each list element being a list with three tibbles.
\code{bind_tbl}() will take the user list and the name of which of the three
tibble to work with. It returns a single tibble which is all users tibbles of
the specific type bound together. A `uid` column is added.
}
\examples{
\dontrun{
# Given a user-list of data `d`...
bind_tbl(d, session)
bind_tbl(d, events)
bind_tbl(d, streams)
}
}
