% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.test.R
\name{binom.test}
\alias{binom.test}
\title{Binomial Test}
\usage{
binom.test(x, n, p = 0.5, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)
}
\arguments{
\item{x}{FLVector}

\item{n}{number of trials}

\item{p}{hypothesized probability of success}

\item{alternative}{indicates the alternative hypothesis and 
must be one of "two.sided", "greater" or "less". 
Additionally "LT","GT","EXACT" are supported for FL objects.}

\item{conf.level}{confidence level for the returned confidence interval.
Not Applicable for FL objects.}
}
\value{
A list with class "htest"
}
\description{
Performs an exact test of a simple null hypothesis
about the probability of success in a Bernoulli experiment.
}
\section{Constraints}{

conf.level is not supported currently for FL objects.
print is not working for result htest object
}
\examples{
flv <- as.FLVector(sample(1:50,10,replace=T))
ResulthtestObject <- binom.test(flv,100,p=0.65)
ResulthtestObject <- binom.test(flv,100,p=0.65,"greater")
ResulthtestObject <- binom.test(flv,100,p=0.65,"LT")
expect_equal(ResulthtestObject[10],binom.test(as.R(flv[10]),100,p=0.65,"LT"))
}

