% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting}
\alias{boosting}
\title{Applies the AdaBoost.M1 and SAMME algorithms to a data set}
\usage{
boosting(formula, data, ...)
}
\arguments{
\item{formula}{formula specifying the independent and dependent variable columns}

\item{data}{FLTable}

\item{mfinal}{an integer, the number of iterations
for which boosting is run or the number of trees to use.
Defaults to mfinal=10 iterations.}

\item{control}{options that control details of the rpart algorithm. See rpart for more details.}
}
\value{
An object of class "FLBoosting" which has details about the trees generated
and the number of votes in bag corresponding to every ObsID.
}
\description{
Fits the AdaBoost.M1 (Freund and Schapire, 1996) and SAMME (Zhu et al., 2009)
algorithms using classification trees as single classifiers.
}
\examples{
flt <- FLTable(getTestTableName("tblBoostDT"),
             "ObsID","VarID","Num_Val")
flobj <- boosting(flt, formula = -1~.,mfinal=10)
pred <- predict(flobj, newdata= flt[1:200, 1:18])
pred$confusion
pred$class
}
\seealso{
\code{\link[adabag]{boosting}} for corresponding R function reference.
}
