% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{calc_anyevent}
\alias{calc_anyevent}
\title{Calculate whether any event is happening}
\usage{
calc_anyevent(user, event_vars = c("fall", "ice", "fog", "block", "animal"))
}
\arguments{
\item{user}{A user list object}

\item{event_vars}{Names of boolean variables indicating whether an event is
happening}
}
\value{
The user object with an any_event column in the events and streams
  tibbles.
}
\description{
Function to calculate whether any event is happening at a given time point.
This is a cenvenience instead of constantly writing things like `if (is | fog
| ...)`
}
