% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate distance travelled since last time point}
\usage{
calc_distance(user)
}
\arguments{
\item{user}{A user list object}
}
\value{
The user object with a distance column in the streams tibble if a
  position column is present.
}
\description{
Take a user object and use the position data to calculate the distance in the
sim world they travelled since the last recorded time point.
}
