% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculating.R
\name{calc_speed}
\alias{calc_speed}
\title{Calculate instantaneous speed}
\usage{
calc_speed(user)
}
\arguments{
\item{user}{A user list object}
}
\value{
The user object with a speed column in the streams tibble if a
  velocity column is present.
}
\description{
Take a user object and calculate instantaneous speed from the Unity-produced
Vector 3 of the driver's velocity contained in the streams tibble.
}
