% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCbind.R
\name{cbind.FLMatrix}
\alias{cbind.FLMatrix}
\title{Combine objects by columns.}
\usage{
\method{cbind}{FLMatrix}(object, ...)
}
\arguments{
\item{object...}{can be a sequence of vector, FLVector, matrix, FLMatrix or data frames}
}
\value{
\code{cbind} returns a FLMatrix object which is the column-wise combination of input arguments.
}
\description{
\code{cbind} combines input objects by columns and forms a FLMatrix.
}
\details{
\code{cbind} takes a sequence of vector, FLVector, matrix, FLMatrix or data frames arguments,
combines them by columns and makes a FLMatrix.
}
\section{Constraints}{

Input matrices, FLMatrices and data frames should have same number of rows.
}

