% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCbind.R
\name{cbind}
\alias{cbind}
\title{Combine objects by columns.}
\usage{
cbind(x, ...)
}
\arguments{
\item{x...}{can be a sequence of vector, FLVector, matrix, FLMatrix or data frames}
}
\value{
\code{cbind} returns a FLMatrix object which is the column-wise combination of input arguments.
}
\description{
\code{cbind} combines input objects by columns and forms a FLMatrix.
}
\details{
\code{cbind} takes a sequence of vector, FLVector, matrix, FLMatrix or data frames arguments,
combines them by columns and makes a FLMatrix.
}
\section{Constraints}{

Input matrices, FLMatrices and data frames should have same number of rows.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- cbind(flmatrix,1:5,flmatrix)
}

