% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check whether file exists and return TRUE or determined value}
\usage{
check_file_exists(..., warning_if_not = FALSE)
}
\arguments{
\item{...}{Character strings that will be concantenated to create a file path}

\item{warning_if_not}{Boolean indicating whether a warning should be printed
if the file does not exist}
}
\description{
Take various strings that combine to make a file path and check whether that
file exists. If it does, return TRUE. If not, return the value return_if_not.
}
