% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{checkQueryLimits}
\alias{checkQueryLimits}
\title{Check if sql query limits have been violated}
\usage{
checkQueryLimits(pObj1)
}
\arguments{
\item{pObj1}{input FL object}
}
\value{
logical TRUE if limits violated
}
\description{
Limits can be on length of query to be parsed
or number of nested queries
}
