% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{checkSameDims}
\alias{checkSameDims}
\title{Compare Matrix Dimensions}
\usage{
checkSameDims(object1, object2)
}
\arguments{
\item{object1}{FLMatrix or R Matrix}

\item{object2}{FLMatrix or R Matrix}
}
\value{
logical
}
\description{
Takes two matrices in-database or R, and returns true if they have same dimensions
}
\author{
Phani Srikar <phanisrikar93ume@gmail.com>
}

