% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test.R
\name{chisq.test}
\alias{chisq.test}
\title{Chi-squared Test}
\usage{
chisq.test(x, ...)
}
\arguments{
\item{x}{FLMatrix}

\item{pear}{The values is 1 if wanna perform pearson chi-square test. Otherwise, 
by default it's 0 and perform chi- square test.}
}
\value{
A list with class "htest" outputting the corresponding expected and P Values.
}
\description{
Produces the expected and the chi-square value for each cell in the contingency
table.
}
\examples{
flx<-as.FLVector(rnorm(100))
fly<-as.FLVector(rnorm(100))
z.test(flx)
    ## Two- sided one sample z-test.
    ## mean for 'x' is zero. The alternative hypothesis states 
    ## that it is either greater or less than zero.

z.test(flx,fly,2)
    ## Two- sided two sample z-test.
    ## The null hypothesis is that the population mean for 'x'
    ## less that for 'y' is 2. 
    ## The alternative hypothesis is that this difference is not 2.

## z- test for population proportion.
fla<-as.FLVector(sample(c(0,1),replace=T,size=100))
flb<-as.FLVector(sample(c(0,1),replace=T,size=100))
z.test(fla,prob=1)
z.test(fla,flb,prob=1)
}
