% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCholeskyDecomp.R
\name{chol.FLMatrix}
\alias{chol.FLMatrix}
\title{Cholesky Decomposition.}
\usage{
\method{chol}{FLMatrix}(object)
}
\arguments{
\item{table}{an object of class FLMatrix}
}
\value{
\code{chol} returns the upper triangular factor of the Cholesky decomposition
}
\description{
\code{chol} computes the Cholesky factorization of FLMatrix object.\cr
The Cholesky decomposition is a decomposition of a positive definite matrix into the product of a lower triangular matrix
and its conjugate transpose.
}
\details{
The wrapper overloads chol and implicitly calls FLCholeskyDecompUdt.
}
\section{Constraints}{

Input can only be a Hermitian, positive definite square matrix (n x n)
with maximum dimension limitations of (1000 x 1000)
}
\examples{
connection<-odbcConnect("Gandalf")
table<-FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
chol(table)
}

