% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCholeskyDecomp.R
\name{chol}
\alias{chol}
\title{Cholesky Decomposition.}
\usage{
chol(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{chol} returns FLMatrix which is the 
upper triangular factor of the Cholesky decomposition
}
\description{
\code{chol} computes the Cholesky factorization of FLMatrix object.\cr
The Cholesky decomposition is a decomposition of a positive definite matrix 
into the product of a lower triangular matrix and its conjugate transpose.
}
\section{Constraints}{

Input can only be a Hermitian, positive definite square matrix (n x n)
with maximum dimension limitations of (1000 x 1000)
}
\examples{
connection<-RODBC::odbcConnect("Gandalf")
flmatrix<-FLMatrix("FL_DEMO", 
"tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- chol(flmatrix)
}

