% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_laps}
\alias{clean_laps}
\title{Clean lap information}
\usage{
clean_laps(user, n_laps = 5)
}
\arguments{
\item{user}{A user list object created by reading functions like
\code{\link{read_logs}}}

\item{n_laps}{The number of laps that the participant was supposed to do in
the simulation.}
}
\description{
This function will add a `lap` number column to the events and streams
tibbles. It will also remove all data that precedes the start of the first
lap, and comes after the end of the last lap.
}

