% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochran.qtest.R
\name{cochran.qtest}
\alias{cochran.qtest}
\title{Cochran's Q test}
\usage{
cochran.qtest(formula, data = NULL, alpha = 0.05, p.method = "fdr", ...)
}
\arguments{
\item{formula}{a formula of the form a ~ b | c, where a, b and c 
give the data values and corresponding groups and blocks, respectively.}

\item{data}{Wide FLTable or FLTableMD objects}

\item{alpha}{significance level to compute pairwise comparisons. Always 0.05.}

\item{p.method}{method for p-values correction. Not used currently.}

\item{...}{The additional arguments used by FL function:
\code{whereconditions} WhereConditions to subset data
\code{GroupBy} Column names defining the different groups in data, if any.}
}
\value{
A list with class \code{RVtest}.
A list of \code{RVtest} objects if the input is a FLTableMD object.
}
\description{
Performs the Cochran's Q test for unreplicated 
randomized block design experiments with a binary 
response variable and paired data.
}
\section{Constraints}{

p.value.multcomp is NA in result
}
\examples{
response <- c(0,1,1,0,0,1,0,1,1,1,1,1,0,0,1,1,0,1,0,1,1,0,0,1,0,1,1,0,0,1)
fact <- gl(3,1,30,labels=LETTERS[1:3])
block <- gl(10,3,labels=letters[1:10])
Rdata <- data.frame(response=response,block=block,fact=fact)
dropFLTestTable()
flt <- as.FLTable(Rdata,tableName="ARBaseTestTempTable")
RVtestObject <- cochran.qtest(response~fact|block,data=flt)
print(RVtestObject)
}

