% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLColMeans.R
\name{colMeans.FLMatrix}
\alias{colMeans.FLMatrix}
\title{column means of a FLMatrix.}
\usage{
\method{colMeans}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix.}
}
\value{
\code{colMeans} returns a FLVector object representing the column-wise Means.
}
\description{
\code{colMeans} computes the column-wise average of FLMatrix objects.
}
\details{
The wrapper overloads colMeans and extends it to FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLVector <- colMeans(flmatrix)
}

