% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLColSums.R
\name{colSums.FLMatrix}
\alias{colSums.FLMatrix}
\title{column sums of a FLMatrix.}
\usage{
\method{colSums}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix.}
}
\value{
\code{colSums} returns a FLVector object representing the col-wise sums.
}
\description{
\code{colSums} computes the column-wise sums of FLMatrix objects.
}
\details{
The wrapper overloads colSums and extends it to FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLVector <- colSums(flmatrix)
}

