% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{constructSelect}
\alias{constructSelect}
\title{constructs a sql statement returning the
deep table representation of the object.}
\usage{
constructSelect(object, ...)
}
\arguments{
\item{object}{the object to query}

\item{...}{arguments passed on to SQL generation. see joinNames}
}
\value{
a character SQL representation
}
\description{
constructs a sql statement returning the
deep table representation of the object.
}
