% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLCorrel.R
\name{cor.FLMatrix}
\alias{cor.FLMatrix}
\title{Correlation.}
\usage{
\method{cor}{FLVector}(x, y = x)
}
\arguments{
\item{x}{A numeric vector,matrix or data frame}

\item{y}{A vector,matrix or data frame with compatible dimensions to x}
}
\value{
\code{cor} returns correlation of x and y.
}
\description{
\code{cor} computes correlation of FLVectors: x and y.
}
\details{
The wrapper overloads cor and implicitly calls FLCorrel.
}
\section{Constraints}{

The number of non-null pairs must be greater than or equal to 2.
If number of non-null pairs is less than 2, FLCorrel returns a NULL.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
table <- FLTable(connection, "FL_TRAIN", "tblAbaloneWide", "ObsID")
cor(table,table)
}

