% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCorrel.R
\name{cor}
\alias{cor}
\title{Correlation.}
\usage{
cor(x, y, ...)
}
\arguments{
\item{x}{FLMatrix, FLVector or FLTable object or any R object}

\item{y}{FLMatrix, FLVector or FLTable object or any R object}
}
\value{
\code{cor} returns FLMatrix object representing correlation of x and y.
}
\description{
\code{cor} computes correlation of in-database Objects
}
\section{Constraints}{

The number of non-null pairs must be greater than or equal to 2.
If number of non-null pairs is less than 2, FLCorrel returns a NULL.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
deeptable <- FLTable(connection, "FL_DEMO", "tblUSArrests", "ObsID","VarID","Num_Val")
widetable <- FLTable(connection,"FL_DEMO","tblAbaloneWide","ObsID")
cor(deeptable,deeptable)
cor(widetable,widetable)
}
\author{
Gregor Kappler <g.kappler@gmx.net>, phani srikar <phanisrikar93ume@gmail.com>
}

