% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCorrel.R
\name{cov.wtGeneric}
\alias{cov.wtGeneric}
\title{Weighted Covariance.}
\usage{
\method{cov}{wtGeneric}(x, wt = rep(1/nrow(x), nrow(x)), cor = FALSE,
  center = TRUE, method = 1, rowIdColumn = "rowIdColumn",
  colIdColumn = "colIdColumn", valueColumn = "valueColumn")
}
\arguments{
\item{x}{FLMatrix, FLVector or FLTable object or any R object}

\item{y}{FLMatrix, FLVector or FLTable object or any R object}

\item{...}{any additional arguments}
}
\value{
\code{cor} returns FLMatrix object representing correlation of x and y.
}
\description{
\code{cov.wt} computes correlation of in-database Objects
}
\section{Constraints}{

The \code{method} input can be 1,2,3 as in DB-Lytix manual.
"ML" method is equivalent to method=1."unbiased" is not supported.
The number of non-null pairs must be greater than or equal to 2.
If number of non-null pairs is less than 2, FLCorrel returns a NULL.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
deeptable <- FLTable( "FL_DEMO", 
"tblUSArrests", "ObsID","VarID","Num_Val")
widetable <- FLTable("FL_DEMO","tblAbaloneWide","ObsID")
cor(deeptable,deeptable)
cor(widetable,widetable)
}

