% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAbstractSQL.R
\name{createView}
\alias{createView}
\title{Create View}
\usage{
createView(pViewName, pSelect, pDatabase = getOption("ResultDatabaseFL"), ...)
}
\arguments{
\item{pViewName}{Name of view}

\item{pSelect}{SELECT clause for view creation}

\item{pDatabase}{Name of the database}
}
\value{
Name of view if operation is successful
}
\description{
Create an in-database view from a SELECT clause
}
\examples{
vres <- createView("myview120","SELECT * FROM tblmatrixmulti")
}
