% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croston.R
\name{croston}
\alias{croston}
\title{Forecasts for intermittent demand using Croston's method}
\usage{
croston(object, ...)
}
\arguments{
\item{object}{FLVector}

\item{alpha}{Smoothing parameter. Default value = 0.5 .}

\item{h}{Number of periods for forecasting.}
}
\value{
An object containing details of the fitted model and forecasted values.
}
\description{
Returns forecasts and other information for Croston's forecasts applied to object.
}
\details{
Based on Croston's (1972) method for intermittent demand forecasting,
also described in Shenstone and Hyndman (2005). Croston's method 
involves using simple exponential smoothing (SES) on the non-zero elements of the time series
and a separate application of SES to the times between non-zero elements of the time series.
The smoothing parameters are denoted by alpha.
}
\examples{
x<-rnorm(1000)
flv<-as.FLVector(x)
flobj<-croston(object = flv, h=9)
}
