% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLDiag.R
\name{diag}
\alias{diag}
\title{Matrix Diagonals}
\usage{
diag(x, ...)
}
\arguments{
\item{x}{is an object of class FLMatrix or FLVector}
}
\value{
If x is a FLMatrix then diag(x) returns the diagonal of x as FLVector object.
  If x is FLVector, the value is a diagonal square FLMatrix with diagonal elements as given in FLVector.
}
\description{
Extract or replace the diagonal of a matrix, or construct a diagonal matrix.
}
\details{
diag has three distinct usages:
x is a FLMatrix, when it extracts the diagonal.
x is a scalar (length-one FLVector) and the only argument, it returns a square identity matrix of size given by the scalar.
x is a FLVector, either of length at least 2. This returns a square matrix with the given diagonal entries.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
resultFLVector <- diag(flmatrix)
DeepTable <- FLTable(connection, "FL_TRAIN", "tblVectorDeep","vector_id","vector_key","vector_value")
flvectorDeep <- FLVector(DeepTable,"vector_value",1)
resultFLMatrix <- diag(flvectorDeep)
}

