% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLDims.R
\name{dim}
\alias{dim}
\title{Returns the dimensions of the object}
\usage{
dim(object)
}
\arguments{
\item{object}{FLMatrix, FLVector or FLTable object}
}
\value{
R vector giving dimensions of input object
}
\description{
Returns the dimensions of the object
}

