% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStatFunctions.R
\docType{methods}
\name{dist,FLMatrix-method}
\alias{dist,FLMatrix-method}
\title{computes distance between}
\usage{
\S4method{dist}{FLMatrix}(x, method = "euclidean", diag = FALSE,
  upper = FALSE, p = 2)
}
\description{
computes distance between
}
\examples{
flmatrix <- FLMatrix("FL_DEMO",
"tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
dist(flmatrix)
dist(flmatrix,diag=TRUE)
dist(flmatrix,upper=TRUE)
dist(flmatrix,diag=TRUE,upper=TRUE)
dist(flmatrix,"manhattan",TRUE,TRUE)
}

