% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{drop.FLTable}
\alias{drop.FLTable}
\title{drop a table}
\usage{
drop.FLTable(object)
}
\arguments{
\item{object}{FLTable object}
}
\value{
message if the table is dropped
}
\description{
drop a table
}
\author{
Phani Srikar <phani.srikar@fuzzyl.com>
}

