% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLEigen.R
\name{eigen.FLMatrix}
\alias{eigen.FLMatrix}
\title{Spectral Decomposition of a Matrix.}
\usage{
\method{eigen}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{eigen} returns a list of FLMatrix object containing the eigen vectors and
a FLVector object containing eigen values which replicates the equivalent R output.
}
\description{
\code{eigen} Computes eigenvalues and eigenvectors of FLMatrices.
}
\details{
The wrapper overloads eigen and implicitly calls FLEigenValueUdt and FLEigenVectorUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultList <- eigen(flmatrix)
resultList$valueColumns
resultList$vectors
}

