% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{end_with_slash}
\alias{end_with_slash}
\title{Check/append slash to string}
\usage{
end_with_slash(string)
}
\arguments{
\item{string}{Character string to check for, and potentially append, a slash
on the end.}
}
\value{
String of path with "/" on end
}
\description{
Check if a character string has a slash ("/") on the end. If not, append one
and return.
}
