% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIdentical.R
\name{==.FLMatrix}
\alias{==.FLMatrix}
\title{Equality of in-database objects.}
\usage{
\method{==}{FLMatrix}(pObj1, pObj2)
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{==} returns a logical TRUE or FALSE.
}
\description{
\code{==} checks the equality of in-database objects.
}
\details{
The equality of in-database objects mimics the normal addition of R data types.
One can check equality of FLMatrices, FLMatrix - R matrices, FLVectors and
FLVector - RVector.
}
\section{Constraints}{

Currently only \code{dgCMatrix},\code{dgeMatrix},\code{dsCMatrix},
\code{dgTMatrix},\code{matrix},\code{Matrix},\code{vector} R types
are supported. Comparision of FLMatrix with FLVector is not currently Supported.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
Rvector <- 1:5
Result <- flmatrix == flmatrix
Result <- Rvector == as.FLVector(Rvector,connection)
}

