% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIdentical.R
\name{==}
\alias{==}
\title{Equality of in-database objects.}
\usage{
"=="(pObj1, pObj2)
}
\arguments{
\item{pObj1}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{pObj2}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{==} returns a logical TRUE or FALSE matrix similar to R output
}
\description{
\code{==} checks the equality of in-database objects.
}
\details{
The equality of in-database objects mimics the normal addition of R data types.
One can check equality of FLMatrices, FLMatrix - R matrices, FLVectors and
FLVector - RVector.
}
\section{Constraints}{

Currently only \code{dgCMatrix},\code{dgeMatrix},\code{dsCMatrix},
\code{dgTMatrix},\code{matrix},\code{Matrix},\code{vector} R types
are supported. Comparision of FLMatrix with FLVector is not currently Supported.
In case of FLVector and Rvector comparision use FLVector==RVector in place of 
RVector==FLVector
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO.tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
flvector <- as.FLVector(1:5)
Result <- flmatrix == flmatrix
Result <- flvector==flvector
Result <- flvector==1:5
}

