% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{eval_expect_equal}
\alias{eval_expect_equal}
\title{Evaluates and benchmarks the expression e in an R and an FL environment.
tests all your new variable names for equality in R and FL environments.
TODO: The results of both expressions will be returned together with benchmarking statistics}
\usage{
eval_expect_equal(e, Renv, FLenv = as.FL(Renv), description = NULL,
  runs = 1, noexpectation = c(), ...)
}
\arguments{
\item{e}{the expression that will be evaluated in both environments}

\item{description}{if not supplied will default to deparse of the expression}

\item{runs}{if runs>1 the expressions are evaluated several times.  Make sure you do not re-assign the variables in environments that are evaluated on.}

\item{noexpectation}{You can exclude names from}

\item{...}{arguments passed to FLexpect_equal}
}
\value{
a data frame with the description
}
\description{
Created objects will be in both environments.
}
\author{
Gregor Kappler <gregor.kappler@fuzzylogix.com>
}

