% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLExpLog.R
\docType{methods}
\name{exp,FLMatrix-method}
\alias{exp,FLMatrix-method}
\title{Logarithms and Exponentials of in-database objects.}
\usage{
\S4method{exp}{FLMatrix}(x)
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLVector or
a normal R object}

\item{base}{a positive number with respect to which logs are computed}
}
\value{
returns an in-database object if there is atleast one in-database object 
as input.Otherwise, the default behavior of R is preserved
}
\description{
Element-wise Logarithms and Exponentials of in-database objects.
}
\details{
The \code{exp} computes the exponential function
The \code{log} computes the logrithm to base as specified(default e)
The \code{logb} computes the logrithm to base as specified(default e)
The \code{log1p} computes the log10(1+x)
The \code{log10} computes the logarithm to base 10
The \code{log2} computes the logarithm to base 2
The \code{expm1} computes exp(x)-1
All types of operands are possible just like in R 
and the result is an in-database object.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO", 
"tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
ResultFLmatrix <- exp(flmatrix)
ResultFLmatrix <- logb(flmatrix,3)
ResultFLmatrix <- log10(flmatrix)
ResultFLmatrix <- log1p(flmatrix)
ResultFLmatrix <- log(flmatrix)
ResultFLmatrix <- log2(flmatrix)
deeptable <- FLTable("FL_DEMO","tblUSArrests","ObsID","VarID","Num_Val")
flvector <- deeptable[1:5,1]
resultFLVector <- exp(flvector)
resultFLVector <- log(flvector,4)
}

