% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLUDecomp.R
\docType{class}
\name{expandFLLU-class}
\alias{expandFLLU-class}
\title{An S4 class to represent L,U and P factors as a list of matrices}
\description{
An S4 class to represent L,U and P factors as a list of matrices
}
\section{Slots}{

\describe{
\item{\code{luobject}}{object of class FLLU}
}}

