% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{expect_equal_RMatrix_FLMatrix}
\alias{expect_equal_RMatrix_FLMatrix}
\title{tests if a R matrix is correctly stored and
represented when casting the R matrix into FLMatrix
and correctly recieved back, when cast to a vector.
checking dimnames, checking for subsetting.
For an optical check, both matrices are printed.}
\usage{
expect_equal_RMatrix_FLMatrix(a)
}
\arguments{
\item{a}{an R Matrix}
}
\description{
tests if a R matrix is correctly stored and
represented when casting the R matrix into FLMatrix
and correctly recieved back, when cast to a vector.
checking dimnames, checking for subsetting.
For an optical check, both matrices are printed.
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

