% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{expect_eval_equal}
\alias{expect_eval_equal}
\title{DEPRECATED: use eval_expect_equal}
\usage{
expect_eval_equal(initF, FLcomputationF, RcomputationF, ...)
}
\description{
DEPRECATED: use eval_expect_equal
}
