% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFKMeans.R
\name{fanny.FLTable}
\alias{fanny.FLTable}
\title{FuzzyKMeans Clustering.}
\usage{
\method{fanny}{FLTable}(x, k, diss = FALSE, memb.exp = 2,
  metric = "euclidean", Stand = FALSE, iniMem.p = NULL,
  cluster.only = FALSE, keep.diss = (!diss && !cluster.only),
  keep.data = (!diss && !cluster.only), maxit = 500, tol = 1e-06,
  trace.lev = 0, excludeCols = "", classSpec = list(),
  whereconditions = "", distTable = "")
}
\arguments{
\item{x}{an object of class FLTable, can be wide or deep table}

\item{k}{the number of clusters}

\item{diss}{logical if \code{x} is dissimilarity matrix.
currently not used.}

\item{memb.exp}{degree of fuzziness or membership coefficient}

\item{metric}{only "euclidean" distance supported currently}

\item{Stand}{logical indicating if standardization
should be done before calculating diss matrix}

\item{iniMem.p}{inital membership matrix. Currently not used}

\item{cluster.only}{logical if only clustering vector is 
needed as output}

\item{keep.diss}{logicals indicating if the 
dissimilarities and/or input data x should be kept in the result}

\item{keep.data}{logicals indicating if the 
dissimilarities and/or input data x should be kept in the result}

\item{maxit}{maximum number of iterations}

\item{tol}{tolerance used for convergence. Currently 0.000001}

\item{trace.lev}{integer specifying a trace level for 
printing diagnostics during the build and swap phase of the algorithm.
currently always 0}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}

\item{distTable}{name of the in-database table having dissimilarity
matrix or distance table}
}
\value{
\code{fanny} returns a list and replicates equivalent R output
from \code{fanny} in cluster package
}
\description{
\code{fanny} performs fuzzy analysis on FLTable objects.
}
\section{Constraints}{

Plotting for large datasets takes longer time to fetch data.
If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
widetable  <- FLTable(connection, "FL_DEMO", "iris", "rownames")
fkmeansobject <- fanny(widetable,2,memb.exp=2)
print(fkmeansobject)
plot(fkmeansobject)
One can specify ClassSpec and transform categorical variables 
before clustering. This increases the number of variables in the plot
because categorical variable is split into binary numerical variables.
The clusters may not be well-defined as is observed in the case below:-
widetable  <- FLTable( "FL_DEMO", "iris", "rownames")
fannyobjectnew <- fanny(widetable,3,classSpec=list("Species(setosa)"))
plot(fannyobjectnew)
}

