% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLDA.R
\name{fda}
\alias{fda}
\title{Flexible Discriminant Analysis}
\usage{
fda(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable or FLTableMD.}

\item{MaxMARSMaximum}{number of basis functions to be used for multivariate nonparametric
regression step where multivariate adaptive regression splines is used.}

\item{MinRsq}{Terminating condition for multivariate regression step based on change in Rsq upon addition of new basis functions.}
}
\value{
\code{fda} returns an object of class \code{FLLDA}
}
\description{
\code{fda} performs Flexible Discriminant Analysis on FLTable objects.
}
\details{
The DB Lytix function called is FLFlexDiscriminant. Performs Flexible Discriminant Analysis and stores the results in predefined tables.
}
\examples{
#'deeptbl <- FLTable("tblIrisDeep", "ObsID", "VarID", "Num_Val")
flmod <- fda(a~., data = deeptbl)
predict(flmod); cof <-coefficients(flmod);confusion(flmod)
Performs flexible discriminant analysis and stores the results in predefined tables. It
involves the use of multivariate adaptive regression splines for obtaining a basis
transformation of the independent variables and performing the multivariate
non-parametric regression step in the Flexible Discriminant Analysis procedure.

couldnt be implemented: plot, values, precent.explained
(lack of data of discriminant space).
}
\seealso{
\code{\link[stats]{fda}} for R reference implementation.
}

