% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{FLamendDimnames}
\alias{FLamendDimnames}
\title{Amends dimension names to a matrix if a mapping exists.}
\usage{
FLamendDimnames(flm, map_table)
}
\arguments{
\item{flm}{FLMatrix}

\item{map_table}{name of the mapping table if already exists}
}
\value{
the FLMatrix object, with slot dimnames re set
}
\description{
If no mapping exists uses the unique row and column names.
Dimnames 1:n are set to NULL in order to adhere to R conventions.
}

