% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLAncova.R
\name{FLAncova}
\alias{FLAncova}
\title{Analysis of Co-Variance}
\usage{
FLAncova(table, response, variable, control, where_clause = "")
}
\arguments{
\item{table}{An object of class \code{FLTable}}

\item{response}{column name for the data values}

\item{variable}{column name for the categorical independent variable}

\item{control}{column name for continuous independent variable (covariate)}

\item{where_clause}{condition to filter out data from the table}
}
\value{
A \code{data.frame} with the following columns:
\item{Df}{Degrees of Freedom}
\item{Sum Sq}{Squared sum from mean}
\item{Mean Sq}{Mean squared sum}
\item{F value}{Value of the F-Statistic}
\item{Pr(>F)}{P-Value}
There are two rows in the result, one for the variable term in the model,
plus one for \code{Residuals} if there are any.

If all the values in all the categories are the same then the Sum Sq and
Mean Sq for \code{Residuals} are zero and the F Value and P Value are \code{NA}
}
\description{
Performs analysis of co-variance(ANCOVA).
}
\details{
ANCOVA is used to test if population means of a dependent variable
are equal across levels of a categorical independent variable, while
controlling for the effects of other continuous variables (known as
covariates) that are not of primary interest.
}
\examples{
\dontrun{
connection <- odbcConnect("Gandalf")
db_name    <- "FL_R_WRAP"
table_name <- "tblAutoMpg"
# Create FLTable object
table      <-  FLTable(connection, db_name, table_name)
ancovaResult <- FLAncova(table, response = "MPG", variable = "CarName", control = "Weight")
# Select American brands only
ancovaResult <- FLAncova(table, response = "MPG", variable = "CarName", control = "Weight", where_clause = "CarName IN ('Chrysler', 'Dodge', 'Ford')" )
}
}

