% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCApply.R
\name{FLCApply}
\alias{FLCApply}
\title{Apply a function to a subset of data}
\usage{
FLCApply(data, FUN, column)
}
\arguments{
\item{data}{FLTable object}

\item{FUN}{function to apply on each subset}

\item{column}{character giving column to partition by
or integer index of the column}
}
\value{
list of results from each subset
}
\description{
Partition data based on given column
and apply given function on each partition
}
\examples{
irisfl <- FLTable("iris","rownames")
resultList <- FLCApply(irisfl,function(x)kmeans(x,3),"Species")
print(resultList$setosa)
plot(resultList$virginica)
print(resultList$versicolor)
}
