% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{FLClose}
\alias{FLClose}
\title{Close Session and Drop temp Tables}
\usage{
FLClose(connection)
}
\arguments{
\item{connection}{ODBC/JDBC connection object}
}
\description{
Strongly recommended to run before quitting current R session
}
