% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMedoids.R
\docType{class}
\name{FLClustering-class}
\alias{FLClustering-class}
\title{An S4 class to represent FLKMedoids}
\arguments{
\item{object}{retrieves the clustering vector}

\item{object}{returns matrix of the medoids or representative objects of the clusters. If a dissimilarity matrix was given as 
input to pam, then a vector of numbers or labels of observations is given, else medoids is a matrix with in 
each row the coordinates of one medoid.}

\item{object}{returns integer vector of indices giving the medoid observation numbers.}

\item{object}{the objective function after the first and second step of the pam algorithm.}

\item{object}{returns vector with length equal to the number of clusters, specifying which clusters are isolated clusters 
(L- or L*-clusters) and which clusters are not isolated.}

\item{object}{returns matrix, each row gives numerical information for one cluster. These are the cardinality of the cluster
(number of observations), the maximal and average dissimilarity between the observations in the cluster and the cluster's
medoid, the diameter of the cluster (maximal dissimilarity between two observations of the cluster), and the separation 
of the cluster (minimal dissimilarity between an observation of the cluster and an observation of another cluster).}

\item{object}{returns list with silhouette width information.}

\item{object}{dissimilarity (maybe NULL).}

\item{object}{function generating call.}

\item{object}{returns a matrix containing the original or standardized data. This might be missing to save memory or when a 
dissimilarity matrix was given as input structure to the clustering method.}

\item{object}{prints the results of pam on FL objects.}

\item{object}{plots the results of pam on FL objects.}

\item{object}{gives the mapping data.frame which is used in execution.}
}
\description{
An S4 class to represent FLKMedoids
}
\section{Slots}{

\describe{
\item{\code{centers}}{A numeric vector containing the number of clusters, say k}

\item{\code{AnalysisID}}{A character output used to retrieve the results of analysis}

\item{\code{wideToDeepAnalysisID}}{A character string denoting the intermediate identifier
during widetable to deeptable conversion.}

\item{\code{diss}}{logical TRUE if dissimilarity matrix is supplied to \code{pam}}

\item{\code{table}}{FLTable object given as input on which analysis is performed}

\item{\code{results}}{A list of all fetched components}

\item{\code{deeptable}}{A character vector containing a deeptable(either conversion from a 
widetable or input deeptable)}

\item{\code{temptables}}{A list of temporary table names used across the results}

\item{\code{mapTable}}{A character string name for the mapping table in-database if input is wide-table.}

\item{\code{distTable}}{Name of the distance matrix. DistTable should
contain a N x N distance matrix between all ObsID in an input FLTable.}

\item{\code{maxit}}{maximal number of iterations for the FANNY algorithm.}
}}

