% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLConcatString}
\alias{FLConcatString}
\title{Concatenate elements of vector}
\usage{
FLConcatString(object, delimiter)
}
\arguments{
\item{object}{FLVector of characters}

\item{delimiter}{character}
}
\value{
FLVector of length 1 with result string
}
\description{
Concatenate elements of FLVector with
a delimiter as collapse value
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLConcatString(flv,",")
}

