% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{FLConnection}
\alias{FLConnection}
\title{A FLConnection object stores either a JDBC or a ODBC connection
as well as the platform that is connected to.}
\usage{
FLConnection(connection, platform)
}
\arguments{
\item{connection}{ODBC/JDBC connection class for connectivity for R}

\item{platform}{character, either TD, TDAster, or Hadoop}
}
\description{
A FLConnection object stores either a JDBC or a ODBC connection
as well as the platform that is connected to.
}

