% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCoxph.R
\docType{class}
\name{FLCoxPH-class}
\alias{FLCoxPH-class}
\title{An S4 class to represent output from coxph on in-database Objects}
\description{
An S4 class to represent output from coxph on in-database Objects
}
\section{Slots}{

\describe{
\item{\code{timeValCol}}{column name representing time variable}

\item{\code{statusCol}}{column name representing Status variable}

\item{\code{vfcalls}}{information about system tables
and in-database procedures called during execution}
}}

