% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateDistributionsCreators.R
\name{FLcreateUnivariateMethodsFile}
\alias{FLcreateUnivariateMethodsFile}
\title{Function to generate S4 methods for univariate distributions code.}
\usage{
FLcreateUnivariateMethodsFile(definitions = "data/univariate.RFL",
  genFile = "AdapteR/R/univariate.R")
}
\description{
Univariate distribution functions are defined in data/univariate.RFL.
Uses R-like syntax to define function and argument mappings efficiently.
This code is parsed programmatically:
function definitions like are created programmatically.
}
\details{
stats::pweibull            <- FLCDFWeibull(q,location=0,scale,shape)

for executing the R function (left of " <- ") with
in-dabase with DB Lytix function (right of " <- ").
}

