% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLDataMining-class}
\alias{FLDataMining-class}
\title{An S4 class to represent objects returned by Data-Mining functions}
\description{
An S4 class to represent objects returned by Data-Mining functions
}
\section{Slots}{

\describe{
\item{\code{deeptable}}{A character vector containing 
the deeptable on conversion from a widetable}

\item{\code{AnalysisID}}{An output character ID from call to data-mining function}

\item{\code{wideToDeepAnalysisId}}{An output character ID from FLRegrDataPrep}

\item{\code{mapTable}}{name of the mapping table generated by FLRegrDataPrep}

\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

