% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLdifference.R
\name{FLdiff.FLVector}
\alias{FLdiff.FLVector}
\title{rv <- sqlQuery(connection, "SELECT NUM_VAL FROM tblTimeSeriesW1")
flv <- as.FL(rv$NUM_VAL)
FLdiff.FLVector(flv)}
\usage{
\method{FLdiff}{FLVector}(data, degree = 2, ...)
}
\description{
rv <- sqlQuery(connection, "SELECT NUM_VAL FROM tblTimeSeriesW1")
flv <- as.FL(rv$NUM_VAL)
FLdiff.FLVector(flv)
}
