% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLdifference.R
\name{FLdiff}
\alias{FLdiff}
\title{\code{FLdiff} performs difference operator on FLVector objects.}
\usage{
FLdiff(data = list(), degree, ...)
}
\arguments{
\item{data}{An object of class FLVector.}

\item{degree}{The degree of the difference operator.}
}
\value{
\code{FLdiff} returns a data.frame
}
\description{
The DB Lytix function called is FLDifferenceUdt. Performs Difference operator and 
stores the results in data frame.
}
\examples{
rv <- sqlQuery(connection, "SELECT NUM_VAL FROM tblTimeSeriesW1")
flv <- as.FL(rv$NUM_VAL)
FLdiff(flv)
}
