% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLdifference.R
\name{FLEWMA.FLVector}
\alias{FLEWMA.FLVector}
\title{rv <- sqlQuery(connection, "SELECT stockreturn FROM finstockreturns WHERE TICKERID =3")
flv <- as.FL(rv$STOCKRETURN)
FLEWMA(flv)}
\usage{
\method{FLEWMA}{FLVector}(data, ValueType = "R", ...)
}
\description{
rv <- sqlQuery(connection, "SELECT stockreturn FROM finstockreturns WHERE TICKERID =3")
flv <- as.FL(rv$STOCKRETURN)
FLEWMA(flv)
}
