% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLdifference.R
\name{FLEWMA}
\alias{FLEWMA}
\title{\code{FLEWMA} performs Exponentially Weighted Moving Average on FLVector objects.}
\usage{
FLEWMA(data = list(), ValueType = "R", ...)
}
\arguments{
\item{data}{An object of class FLVector.}

\item{ValueType}{Is the input actual time series values or changes
(deltas/Returns)?The degree of the difference operator.}
}
\value{
\code{FLEWMA} returns a data.frame
}
\description{
The DB Lytix function called is FLEWMAUdt. Performs volatility based on most recent volatility and the most recent return (change) in a times series.
stores the results in data frame.
}
\examples{
rv <- sqlQuery(connection, "SELECT stockreturn as stockreturn FROM finstockreturns WHERE TICKERID =3")
flv <- as.FL(rv$stockreturn)
FLEWMA(flv)
}
