% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLExtractStr}
\alias{FLExtractStr}
\title{Extract parts of strings}
\usage{
FLExtractStr(object, delimiter, stringpos)
}
\arguments{
\item{object}{FLVector of characters}

\item{delimiter}{character}

\item{stringpos}{identifier to reference the
sub-string given by its position}
}
\value{
FLVector
}
\description{
Extract sub-strings separated by a
delimiter and identified by their position
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLExtractStr(flv,"A",1)
}

