% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLFetch}
\alias{FLFetch}
\title{Fetch Data Mining results from Database}
\usage{
FLFetch(object)
}
\arguments{
\item{object}{an object inherited from class \code{FLDataMiningAnalysis}}
}
\value{
a fitted model object of the relevant type with the parameters
filled in
}
\description{
\code{FLFetch} fetches the slots of a fitted data mining model object from
the database. For details see the documentation for the respective model
fitting function
}
\examples{
\dontrun{
		KMeansAnalysis <- FLFetch(KMeansAnalysis)
}
}

