% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLFitDistribution.R
\name{FLFitDistrObject}
\alias{FLFitDistrObject}
\title{Constructor function for FLFitDistr}
\usage{
FLFitDistrObject(connection, db_name, table_name, distribution_type,
  value = "NumVal", num_success = "NumOfSuccess",
  num_trials = "NumOfTrials")
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{db_name}{name of the database in \code{Teradata} which contains the table}

\item{table_name}{name of the  table which contains the points for distribution
fitting}

\item{distribution_type}{type of distribution which is to be fitted. It has
to be "Cont" for continuous distribution or "Disc" for discrete distribution.}

\item{value}{name of the column which has the value of the input data points.
This is required only if the distribution is a continuous one.}

\item{num_success}{name of the column which has the number of successes for
a binomial distribution}

\item{num_trials}{name of the column which has the number of trials for a
binomial distribution}
}
\value{
\code{FLFitDistrObject} returns an object of class \code{FLFitContDistr}
or or \code{FLFitDiscDistr} which is mapped to a table in Teradata. An object
of any of the above class has the following common components:
\describe{
\item{odbc_connection}{}
\item{db_name}{}
\item{table_name}{}
}
In addition to the above components, an object of class \code{FLFitContDistr}
contains the component \code{value}. \cr
An object of class \code{FLFitDiscDistr} contains the components
\code{num_success} and \code{num_trials}.
}
\description{
Constructor function for FLFitDistr
}
\details{
\code{FLFitDistrObject} constructs an object of class \code{FLFitContDistr}
or \code{FLFitDiscDistr} depending on whether a continuous or a discrete
distribution is to be fitted. This object is used as the input for fit
distribution functions.
}
\examples{
\dontrun{
   connection  <- odbcConnect("Gandalf")
		database    <- "FL_R_WRAP"
		table_name  <- "RWrapperFitNormalDistrTest"
  	# Create distribution object
  	data        <-  FLFitDistrObject( connection,
                                     db_name,
                                     table_name,
                                     distribution_type = "Cont")
		# Fit Distribution
  	result      <- FLFitDistr(Data, "Normal", "MDE"))
}
}

