% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFKMeans.R
\docType{class}
\name{FLFKMeans-class}
\alias{FLFKMeans-class}
\title{An S4 class to represent FLFKMeans}
\arguments{
\item{object}{returns the clustering vector of the nearest crisp clustering.}

\item{object}{returns matrix containing the memberships for each pair consisting of an observation and a cluster.}

\item{object}{returns vector with Dunn's partition coefficient F(k) of the clustering, where k is the number of clusters. F(k) is
the sum of all squared membership coefficients, divided by the number of observations. Its value is between 1/k and 1.
The normalized form of the coefficient is also given. It is defined as (F(k) - 1/k) / (1 - 1/k), and ranges between 0 and 1.
A low value of Dunn's coefficient indicates a very fuzzy clustering, whereas a value close to 1 indicates a near-crisp clustering.}

\item{object}{returns named vector containing the minimal value of the objective function reached by the FANNY algorithm and 
the relative convergence tolerance tol used.}

\item{object}{returns integer (<= k) giving the number of crisp clusters; can be less than k, where it's 
recommended to decrease memb.exp.}

\item{object}{returns named vector with iterations, the number of iterations needed and converged indicating if the algorithm 
converged (in maxit iterations within convergence tolerance tol).}

\item{object}{returns list with silhouette information of the nearest crisp clustering.}

\item{object}{generating call}

\item{object}{prints results of agglomerative clustering.}

\item{object}{plots results of agglomerative clustering.}
}
\description{
An S4 class to represent FLFKMeans
}
\section{Slots}{

\describe{
\item{\code{centers}}{A numeric vector containing the number of clusters, say k}

\item{\code{AnalysisID}}{A character output used to retrieve the results of analysis}

\item{\code{wideToDeepAnalysisID}}{A character string denoting the intermediate identifier
during widetable to deeptable conversion.}

\item{\code{diss}}{logical TRUE if dissimilarity matrix is supplied to \code{fanny}}

\item{\code{table}}{FLTable object given as input on which analysis is performed}

\item{\code{results}}{A list of all fetched components}

\item{\code{deeptable}}{A character vector containing a deeptable(either conversion from a 
widetable or input deeptable)}

\item{\code{temptables}}{A list of temporary table names used across the results}

\item{\code{mapTable}}{A character string name for the mapping table in-database if input is wide-table.}

\item{\code{memb.exp}}{A number r strictly larger than 1 specifying the membership exponent 
used in the fit criterion.Default: 2 which is hardwired inside FANNY.}

\item{\code{maxit}}{maximal number of iterations for the FANNY algorithm.}
}}

