% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLHessenDecomp.R
\name{FLHessen}
\alias{FLHessen}
\title{Hessenberg Decomposition of a Matrix.}
\usage{
FLHessen(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{FLHessen} returns a list of two components:
      \item{P}{FLMatrix representing P matrix obtained from Hessenberg decomposition}
      \item{H}{FLMatrix representing H matrix obtained from Hessenberg decomposition}
}
\description{
\code{FLHessen} computes the Hessenberg decomposition for FLMatrix objects.
}
\section{Constraints}{

Input can only be square matrix with maximum dimension limitations of (700 x 700).
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO.tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- FLHessen(flmatrix)
resultList$P
resultList$H
}

