% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\docType{class}
\name{FLIndexedValues-class}
\alias{FLIndexedValues-class}
\title{Models sparse data objects.}
\description{
Sparse Indexed values objects are table queries with a value column and
one (vectors), two (matrices) or several (arrays) index columns.
}
\details{
The name of the values column
}

