% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLInsertMatrix.R
\name{FLInsertMatrix}
\alias{FLInsertMatrix}
\title{Insert R matrix object in database}
\usage{
FLInsertMatrix(rmatrix, flmatrix)
}
\arguments{
\item{rmatrix}{object of class \code{matrix}}

\item{rmatrix}{object of class \code{FLMatrix}}
}
\description{
This function takes an R matrix object and inserts into an existing database
table with the appropriate schema
}

