% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLJordanDecomp.R
\name{FLJordan}
\alias{FLJordan}
\title{Jordan Decomposition of a Matrix.}
\usage{
FLJordan(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}
}
\value{
\code{FLJordan} returns a list of two components:
      \item{J}{FLVector representing J vector obtained from Jordan decomposition}
      \item{P}{FLMatrix representing P matrix obtained from Jordan decomposition}
      \item{PInv}{FLMatrix representing PInv matrix obtained from Jordan decomposition}
}
\description{
\code{FLJordan} computes the Jordan decomposition for FLMatrix objects.
}
\section{Constraints}{

Input can only be square matrix with maximum dimension limitations of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- FLJordan(flmatrix)
resultList$J
resultList$P
resultList$PInv
}

