% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLKMeans.R
\docType{class}
\name{FLKMeans-class}
\alias{FLKMeans-class}
\title{An S4 class to represent FLKMeans}
\arguments{
\item{object}{retrieves the cluster vector}

\item{object}{retrieves the coordinates of the centroids}

\item{object}{overloads the print function}

\item{object}{total within sum of squares}

\item{object}{within sum of squares}

\item{object}{between sum of squares}

\item{object}{total sum of squares}

\item{object}{size vector}
}
\description{
An S4 class to represent FLKMeans
}
\section{Slots}{

\describe{
\item{\code{no_of_centers}}{A numeric vector containing the number of clusters, say k}

\item{\code{AnalysisID}}{A character output used to retrieve the results of analysis}

\item{\code{odbc_connection}}{ODBC connectivity for R}

\item{\code{table_name}}{A character}

\item{\code{clusterfetched}}{A logical vector either TRUE or FALSE}

\item{\code{centerfetched}}{A logical vector either TRUE or FALSE}

\item{\code{cluster}}{A vector data type}

\item{\code{centers}}{A matrix data type}

\item{\code{deeptablename}}{A character vector containing a deeptable(either conversion from a widetable or IsDeep=TRUE)}
}}

