% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{FLKMeans}
\alias{FLKMeans}
\title{K-Means Clustering}
\usage{
FLKMeans(table, primary_key, centers, max_iter = 10, nstart = 1,
  exclude = as.character(c()), class_spec = list(), where_clause = "",
  note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{An object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{centers}{the number of clusters}

\item{max_iter}{the maximum number of iterations allowed}

\item{nstart}{Number of starting hypotheses}

\item{exclude}{vector of names of columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used a reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{note}
}
\value{
an object of class \code{FLKMeans} whose components can be
pulled to R by running FLFetch. It has the following slots:
\item{centers}{ a \code{data.frame} with columns HypothesisID, Level, ClusterID, VarID, Centroid }
\item{cluster}{ a \code{data.frame} with columns HypothesisID, ObsID, ClusterID}
}
\description{
Performs K-Means clustering of input data.
}
\details{
Clustering algorithms can output slightly different centroid locations
and the associated cluster membership of each data point because the
initialization centroid locations are random. In general, k-means has two
steps: assigning data points to the nearest cluster and then moving each
cluster's centroid to the center of the members of the cluster. Occasionally
during this iterative process, a centroid can get so far away from the
"central mass" of the data points (relative to other centroids) such that
the cluster has no members. In this case, this centroid actually drops off
of the dendrogram. This is a known characteristic of such clustering algorithms,
and in fact k-means is generally run multiple times to take into account this
randomness, and which is why we built in the ability to run multiple hypothesis
into the XSP to run in parallel and take advantage of in-database parallel
processing capabilities.
}
\examples{
\dontrun{

#Simple example with only numerical variables
FLKMeans(KMeansTbl, primary_key = "ID", centers = 2, max_iter = 20, nstart = 1)

}
}

