% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLDA.R
\docType{class}
\name{FLLDA-class}
\alias{FLLDA-class}
\title{An S4 class to represent output from Discriminant Analysis on in-database Objects}
\description{
An S4 class to represent output from Discriminant Analysis on in-database Objects
}
\section{Slots}{

\describe{
\item{\code{offset}}{column name used as offset}
}}

