% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLLinRegr-class}
\alias{FLLinRegr-class}
\title{An S4 class to represent FLLinRegr}
\arguments{
\item{object}{prints results of FLLinRegr on FL objects}

\item{object}{returns coefficient vector of the object}

\item{object}{the residuals, that is response minus fitted values.}

\item{object}{returns the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.}

\item{object}{returns the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.}

\item{object}{plots the results of FLLinRegr on FL objects.}

\item{object}{returns prediction results as FLVector.}
}
\description{
An S4 class to represent FLLinRegr
}
\section{Slots}{

\describe{
\item{\code{formula}}{an object of class 'formula': Model Formula}

\item{\code{deeptable}}{A character vector containing a deeptable(either conversion from a 
widetable or input deeptable)}

\item{\code{AnalysisID}}{An output character ID from CALL FLLinRegr}

\item{\code{wideToDeepAnalysisID}}{An output character ID from FLRegrDataPrep}

\item{\code{mapTable}}{A character string name for the mapping table in-database if input is wide-table.}

\item{\code{scoreTable}}{name of the scoring table}

\item{\code{modelID}}{id of the model with best fit}

\item{\code{table}}{input FLTable object}

\item{\code{results}}{cache list of results computed}

\item{\code{vfcalls}}{contains names of tables}

\item{\code{offset}}{this can be used to specify a priori known component to be included in the 
linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases}
}}

