% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLLinRegrStep.R
\name{FLLinRegrStep}
\alias{FLLinRegrStep}
\title{Stepwise Linear Regression}
\usage{
FLLinRegrStep(table, primary_key, response, type, p_allow = 0.05,
  p_allow_initial = 0.3, p_allow_final = 0.05, stepwise_decrease = 0.05,
  top_n = 5, include = c(), exclude = c(), class_spec = list(),
  where_clause = "", note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column}

\item{type}{one of \{ "BW", "FB", "UFB", "SW" \} standing for Backward,
Fast Backward, Ultra-Fast Backward and Stepwise algorithm respectively}

\item{p_allow}{(For "BW" and "SW") see details}

\item{p_allow_initial}{(For "FB" and "UFB") see details}

\item{p_allow_final}{(For "FB" and "UFB") see details}

\item{stepwise_decrease}{(For "UFB") see details}

\item{top_n}{(For "SW") see details}

\item{include}{vector of names of the columns which are not to be dropped}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{free form string that will be stored with the results, typically
used to document the purpose of the analysis}
}
\value{
an object of class \code{FLLinRegr}. For details see \code{\link{FLLinRegr}}
}
\description{
Chooses a Linear Regression model by eliminating predictors by PValue with
a Stepwise Algorithm
}
\details{
Each of the algorithm works as follows:
\describe{
\item{BW}{In Backward Elimination variables are eliminated one at a time till
all the p-Values are lower than value specified by the parameter \code{p_allow}}
\item{FB}{In Fast Backward Elimination, all the variables that exceed the
value specified by the parameter \code{p_allow_initial} are dropped in one go.
Thereafter, only one variable is dropped at a time till all the p-Values are
below \code{p_allow_final} }
\item{UFB}{In Ultra Fast Backward Elimination, the \code{p_allow_initial} is
revised downward by \code{stepwise_decrease} after each step and all variables
that exceed \code{p_allow_initial} are dropped till all the p-Values are
below \code{p_allow_final} }
\item{SW}{Stepwise Linear regression starts with 1-variable models built
individually with all the variables and then it shortlists the top N number
of models based on their respective Gini coefficient estimates. Then it adds
an additional variable to each of the top N models and filters the top N
models again based on their Gini coefficient estimates. This process continues
till no additional variables can be added because the p-Value of the model
exceeds \code{p_allow}.}
}
}
\examples{
\dontrun{

}
}

