% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLLogRegr.R
\name{FLLogRegr}
\alias{FLLogRegr}
\title{Logistic Regression}
\usage{
FLLogRegr(table, primary_key, response, max_iter, threshold,
  exclude = as.character(c()), class_spec = list(), where_clause = "",
  note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column}

\item{max_iter}{Maximum number of iterations done before terminating the algorithm}

\item{threshold}{threshold for False positive value used to calculate
the false positives and false negatives.}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{note (a string)}
}
\value{
returns an object of class \code{FLLogRegr} whose components can be
pulled to R by running FLFetch. The class will then have 2 slots:

\code{stats}:
\item{Concordant}{Concordant Pairs}
\item{Discordant}{Discordant Pairs}
\item{Tied}{Tied Pairs}
\item{TotalPairs}{Total Pairs}
\item{GiniCoeff}{(Concordant -Discordant)/TotalPairs}
\item{CStatistic}{(GiniCoeff +1)/2}
\item{Gamma}{(Concordant - Discordant)/(Concordant + Discordant)}
\item{HighestPValue}{Highest P-Value}
\item{Events}{Number of Events (the 1's)}
\item{NonEvents}{Number of Non-Events (the 0's)}
\item{FalsePositive}{Number of false positives}
\item{FalseNegative}{Number of False negatives}

\code{coeffs}:
\item{COEFFID}{Coefficient ID}
\item{VAR_TYPE}{Variable Type}
\item{COLUMN_NAME}{Variable Name}
\item{CATVALUE}{Category Name represented by dummy variable}
\item{CoeffValue}{They are coefficients, one for each explanatory variable, that represent the strength and type of relationship the explanatory variable has to the dependent variable.}
\item{StdErr}{Standard Error(an estimate of the standard deviation of the coefficient)}
\item{CHISQ}{Chi-Square estimate of the StdErr}
\item{PValue}{P-Value for the Chi-Square estimate}
}
\description{
Performs Logistic Regression
}
\details{
The function will perform iterative calculations until convergence
is achieved or the maximum number of iterations as specified by \code{max_iter} is crossed
}
\examples{
\dontrun{

	LogRegModel <- FLLogRegr(TblIris, 'ObsID', 'species', max_iter = 20, threshold = 0.8, note = "RWrappers Example")
	LogRegModel <- FLFetch(LogRegModel)
 }
}

