% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLogRegrMN.R
\docType{class}
\name{FLLogRegrMN-class}
\alias{FLLogRegrMN-class}
\title{An S4 class to represent output from glm when family is multinomial}
\description{
An S4 class to represent output from glm when family is multinomial
}
\section{Slots}{

\describe{
\item{\code{offset}}{column name used as offset}

\item{\code{vfcalls}}{contains names of tables}
}}

