% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{FLMatrix}
\alias{FLMatrix}
\title{Constructor function for FLMatrix.}
\usage{
FLMatrix(connection, database, matrix_table, matrix_id_value,
  matrix_id_colname = "MATRIX_ID", row_id_colname = "ROW_ID",
  col_id_colname = "COL_ID", cell_val_colname = "CELL_VAL")
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{database}{name of the database}

\item{matrix_table}{name of the matrix table}

\item{matrix_id_value}{identifier for the input matrix}

\item{matrix_id_colname}{matrix id value in \code{matrix_table}}

\item{row_id_colname}{row number in \code{matrix_table}}

\item{col_id_colname}{column number in \code{matrix_table}}

\item{cell_val_colname}{value of the matrix element in \code{matrix_table}}
}
\value{
\code{FLMatrix} returns an object of class FLMatrix mapped
to a matrix.
}
\description{
\code{FLMatrix} constructs an object of class \code{FLMatrix}.
}
\details{
This object is used as input for matrix operation functions.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
table <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
}

