% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixArithmetic.R
\name{FLMatrixArithmetic}
\alias{FLMatrixArithmetic}
\title{Arithmetic operations on in-database objects.}
\usage{
FLMatrixArithmetic(pObj1, pObj2, pOperator)
}
\description{
\code{-} does the subtraction of in-database objects.
}
\details{
All the operators work the same way with in-database objects as they do with R data types.
The result is an in-database object.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
Rvector <- 1:5
ResultFLmatrix <- flmatrix - Rvector
}

