% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixBind.R
\docType{class}
\name{FLMatrixBind-class}
\alias{FLMatrixBind-class}
\title{An S4 class to represent FLMatrix}
\description{
An S4 class to represent FLMatrix
}
\section{Slots}{

\describe{
\item{\code{connection}}{ODBC connectivity for R}

\item{\code{by}}{character either rows or cols}
}}

