% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLMatrixInv}
\alias{FLMatrixInv}
\title{Inverse of a matrix}
\usage{
FLMatrixInv(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix; \code{matrix} should be and
non-singular.}
}
\value{
\code{FLMatrixInv} returns an object of class FLMatrix mapped
to the inverse of the  input matrix,which can be pulled to R using
\code{FLFetchMatrix}.
}
\description{
\code{FLMatrixInv} computes the inverse of a non-singular square matrix.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
\dontrun{
connection      <- odbcConnect("Gandalf")
database        <- "FL_R_WRAP"
matrix_table    <- "tblMatrixMulti"
matrix_id_value <- 5
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
out_matrix      <- FLMatrixInv(in_matrix)
}
}

