% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixNorm.R
\name{FLMatrixNorm}
\alias{FLMatrixNorm}
\title{Norm of a Matrix.}
\usage{
FLMatrixNorm(object, NormMethod)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{NormMethod}{is an integer from 1-4 representing the type of norm that
should be computed.}
}
\value{
\code{FLMatrixNorm} returns a R vector object which is the Norm of input
FLMatrix object calculated using method specified by NormMethod input.
There are 4 types of norms of a matrix:
\item{1-Norm}{Maximum of the sum of the absolute values for the columns}
\item{2-Norm}{Maximum of the sum of the absolute values for the rows}
\item{Frobenius Norm}{Square root of the trace of (t(A)A)}
\item{Infinity Norm}{Square root of the maximum of the magnitudes of the Eigenvalues of (t(A)A)}
}
\description{
\code{FLMatrixNorm} gives the value of Norm for FLMatrix objects.
}
\section{Constraints}{

Input can only be with maximum dimension limitations of (700 x 700).
}
\examples{
connection <- RODBC::odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", 
"tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- FLMatrixNorm(flmatrix,4)
}

