% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixREF.R
\name{FLMatrixREF.FLMatrix}
\alias{FLMatrixREF.FLMatrix}
\title{Row Echelon form of a Matrix.}
\usage{
\method{FLMatrixREF}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{FLMatrixREF} returns a FLMatrix object which is the Row Echelon form of input FLMatrix.
}
\description{
\code{FLMatrixREF} gives the Row Echelon form of FLMatrix objects.
}
\details{
The wrapper overloads FLMatrixREF and implicitly calls FLMatrixREFUdt.
}
\section{Constraints}{

Input can only be a square FLMatrix with maximum dimension limitations of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- FLMatrixREF(flmatrix)
}

