% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixREF.R
\name{FLMatrixREF}
\alias{FLMatrixREF}
\title{Row Echelon form of a Matrix.}
\usage{
FLMatrixREF(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{FLMatrixREF} returns a FLMatrix object which is the Row Echelon form of input FLMatrix.
}
\description{
\code{FLMatrixREF} gives the Row Echelon form of FLMatrix objects.
}
\section{Constraints}{

Input can only be a square FLMatrix with maximum dimension limitations of (1000 x 1000).
}

\examples{
flmatrix <- FLMatrix("tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- FLMatrixREF(flmatrix)
}
