% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLMatrixRREF.R
\name{FLMatrixRREF.FLMatrix}
\alias{FLMatrixRREF.FLMatrix}
\title{Reduced Row Echelon form of a Matrix.}
\usage{
\method{FLMatrixRREF}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{FLMatrixRREF} returns a FLMatrix object which is the Reduced Row Echelon form of input FLMatrix.
}
\description{
\code{FLMatrixRREF} gives the Reduced Row Echelon form of FLMatrix objects.
}
\details{
The wrapper overloads FLMatrixRREF and implicitly calls FLMatrixRREFUdt.
}
\section{Constraints}{

Input can only be a square FLMatrix with maximum dimension limitations of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- FLMatrixRREF(flmatrix)
}

