% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixRREF.R
\name{FLMatrixRREF}
\alias{FLMatrixRREF}
\title{Reduced Row Echelon form of a Matrix.}
\usage{
FLMatrixRREF(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}
}
\value{
\code{FLMatrixRREF} returns a FLMatrix object which is the Reduced Row Echelon form of input FLMatrix.
}
\description{
\code{FLMatrixRREF} gives the Reduced Row Echelon form of FLMatrix objects.
}
\section{Constraints}{

Input can only be a square FLMatrix with maximum dimension limitations of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- FLMatrixRREF(flmatrix)
}

