% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLNaiveBayes.R
\name{FLNaiveBayes}
\alias{FLNaiveBayes}
\title{Naive Bayes Classifier}
\usage{
FLNaiveBayes(table, primary_key, response, laplace = 0, exclude = c(),
  class_spec = list(), where_clause = "",
  note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column}

\item{laplace}{indicates whether Laplacian Correction is to be used (1 for
true and 0 for false)}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{free form string that will be stored with the results, typically
used to document the purpose of the analysis}
}
\description{
Naive Bayes is a simple probabilistic classifier that applies the Bayes'
theorem to compute conditional a-posterior probabilities of a categorical
class variable under the independence assumption -- the presence
(or absence) of a particular feature of a class is unrelated to the
presence (or absence) of any other feature.
}
\details{
Laplacian Correction is used to avoid the issue of zero probability
for a given attribute by adding 1 to the numerator. In order to compensate
for this addition, the denominator is also incremented by the total number
of discrete values for the attribute
}
\examples{
\dontrun{
}
}

