% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLNeedleManWunschDist}
\alias{FLNeedleManWunschDist}
\title{FLNeedleManWunschDist}
\usage{
FLNeedleManWunschDist(a, b, matchWeight = 1, mismatchWeight = -1,
  gapPenalty = -1, caseFlag = 0, ...)
}
\arguments{
\item{a}{character or FLVector of characters}

\item{b}{character or FLVector of characters}

\item{matchWeight}{integer weight for having
matching sequential characters between
the strings}

\item{mismatchWeight}{integer Weight
for having nonmatching or non-sequential characters
between the strings}

\item{gapPenalty}{integer penality for gaps}

\item{caseFLag}{logical or 0/1 indicating 
if comparision should be case sensitive}
}
\value{
FLVector if any \code{a} or \code{b}
is R character of length 1. Otherwise returns a FLMatrix.
}
\description{
compute NeedleManWunsch distance between strings.
}
\details{
The DB Lytix function called is FLNeedlemanWunschDist.
This function performs global sequence alignment between two
sequences and finds out structural and functional similarity between them
and returns a score which indicates the best alignment between two  
sequences by searching the highest scores in the similarity matrix.
}
\section{Constraints}{

row vectors are not supported currently.
Refer to \code{@return} section.
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflvector <- FLNeedleManWunschDist("xyz",flv)
resultflvector <- FLNeedleManWunschDist("xyz",flv,method="lv",caseFLag=1)
resultflvector <- FLNeedleManWunschDist("xyz",flv,method="hamming",vlength=4)
resultflmatrix <- FLNeedleManWunschDist(flv,flv,method="jw",p=1)
resultflmatrix <- FLNeedleManWunschDist(c("xyz","juio"),flv,method="jw")
}
\seealso{
\code{\link[Biostrings]{pairwiseAlignment}} for R function 
reference implementation in BioConductor.
}

