% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FLodbcClose}
\alias{FLodbcClose}
\title{Close Session and Drop temp Tables}
\usage{
FLodbcClose(connection)
}
\arguments{
\item{connection}{ODBC/JDBC connection object}
}
\description{
Strongly recommended to run before quitting current R session
}

