% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLParseXML}
\alias{FLParseXML}
\title{Parse XML files}
\usage{
FLParseXML(object)
}
\arguments{
\item{object}{FLVector of characters}
}
\value{
dataframe with parsed XML
}
\description{
Parse XML files stored in-database
as elements of FLVector.
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
wtd <- FLTable("FL_DEMO","tblXMLTest","GroupID")
flv <- wtd[,"pXML"]
resultdataframe <- FLParseXML(flv)
}

