% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLPredict.R
\name{FLPredict}
\alias{FLPredict}
\title{Scoring function}
\usage{
FLPredict(analysis, table, ...)
}
\arguments{
\item{analysis}{a fitted model object of the relevant type with the parameters
filled in. The predict function currently scores the dataset for the analysis
objects mentioned in the Details section.}

\item{table}{an \code{FLTable} object mapped to the dataset which is to be
scored}

\item{...}{other parameters depending on the type of analysis}
}
\value{
\code{FLPredict} returns a \code{data.frame} having the scored value of the dependent variable
}
\description{
\code{FLPredict} scores a new dataset using an existing data mining analysis
model.
}
\details{
\code{FLPredict} predicts the value of the dependent variable using
an existing model for the analysis objects of the following classes.

\describe{
\item{FLLinRegr}{}
\item{FLLogRegr}{}
\item{FLNaiveBayes}{}
\item{FLDecisionTree}{}
}
}
\examples{
\dontrun{

}
}

