% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLRegr-class}
\alias{FLRegr-class}
\title{An S4 class to represent objects returned by Regression Functions}
\description{
An S4 class to represent objects returned by Regression Functions
}
\section{Slots}{

\describe{
\item{\code{formula}}{Formula object given as input}

\item{\code{scoreTable}}{Name of the in-database table where scoring results are stored}
}}

