% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTable.R
\name{FLReshape}
\alias{FLReshape}
\title{Data Preparation}
\usage{
FLReshape(data, formula, value.var, subset = NULL,
  outTable = paste0("ARBase", data, "Reshape"), deepOutput = TRUE, ...)
}
\arguments{
\item{data}{name of table in the database}

\item{formula}{should be of form column1~column2}

\item{value.var}{column containing the values}

\item{outTable}{name of output table}

\item{deepOutput}{TRUE if output has to be a deep table}
}
\value{
\code{list} of \code{table} FLTable object, \code{Dimnames} Mappings
}
\description{
Process deep table to have consistent 
obs and var ids
}
\examples{
resultList <- FLReshape(data="medEconomicData",
                        formula=CountryName ~ IndicatorCode,
                        value.var="TimeSeriesVal",
                        subset="IndicatorCode in ('NY.GDP.MKTP.KD.ZG','FP.CPI.TOTL.ZG') and Years=2010",
                        outTable="tbl1",
                        drop=TRUE)
}
