% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSolveExcl.R
\name{FLSolveExcl}
\alias{FLSolveExcl}
\title{Inverse of a Matrix excluding a dimension.}
\usage{
FLSolveExcl(x, ExclIdx, ...)
}
\arguments{
\item{x}{is of class FLMatrix}

\item{ExclIdx}{is a positive integer specifying row or column id to be excluded.}
}
\value{
\code{solveExcl} returns a FLMatrix object which is the inverse of input FLMatrix object
after excluding given dimension.
}
\description{
\code{FLSolveExcl} computes the inverse for FLMatrix objects by excluding 
the specified row and column from the matrix.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- solveExcl(flmatrix,3)
}

