% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSparseMatrix.R
\docType{class}
\name{FLSparseMatrix-class}
\alias{FLSparseMatrix-class}
\title{An S4 class to represent FLSparseMatrix}
\description{
An S4 class to represent FLSparseMatrix
}
\section{Slots}{

\describe{
\item{\code{connection}}{ODBC connectivity for R}

\item{\code{database}}{character}

\item{\code{table_name}}{character}

\item{\code{matrix_id_value}}{numeric id value}

\item{\code{matrix_id_colname}}{character}

\item{\code{row_id_colname}}{character}

\item{\code{col_id_colname}}{character}

\item{\code{cell_val_colname}}{character}

\item{\code{nrow}}{numeric number of rows of FLSparseMatrix object}

\item{\code{ncol}}{numeric number of cols of FLSparseMatrix object}

\item{\code{dimnames}}{list dimension names of FLSparseMatrix object}
}}

