% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSparseMatrix.R
\name{FLSparseMatrix}
\alias{FLSparseMatrix}
\title{Constructor function for FLSparseMatrix.}
\usage{
FLSparseMatrix(connection, database, table_name, matrix_id_value,
  matrix_id_colname = "MATRIX_ID", row_id_colname = "rowIdColumn",
  col_id_colname = "colIdColumn", cell_val_colname = "valueColumn",
  dimnames = list(c(), c()))
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{database}{name of the database}

\item{table_name}{name of the SparseMatrix table}

\item{matrix_id_value}{identifier for the input matrix}

\item{matrix_id_colname}{matrix id value in \code{table_name}}

\item{row_id_colname}{column name in \code{table_name} where row numbers are stored}

\item{col_id_colname}{column name in \code{table_name} where column numbers are stored}

\item{cell_val_colname}{column name in \code{table_name} where matrix elements are stored}

\item{dimnames}{dimension names of \code{FLSparseMatrix}}
}
\value{
\code{FLSparseMatrix} returns an object of class FLSparseMatrix mapped
to a sparseMatrix.
}
\description{
\code{FLSparseMatrix} constructs an object of class \code{FLSparseMatrix}.
}
\details{
\code{FLSparseMatrix} object is an in-database equivalent to sparseMatrix object.
This object is used as input for matrix operation functions.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flsparsematrix <- FLSparseMatrix(connection, "FL_TRAIN", "tblMatrixMultiSparse", 2)
}

