% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLSqueezeSpace}
\alias{FLSqueezeSpace}
\title{Remove extra spaces in strings}
\usage{
FLSqueezeSpace(object)
}
\arguments{
\item{object}{FLVector of characters}
}
\value{
FLVector with extra spaces removed
}
\description{
Removes extra spaces from elements
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLSqueezeSpace(flv)
}

