% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSV.R
\name{FLSV.FLMatrix}
\alias{FLSV.FLMatrix}
\title{Singular Values of a FLMatrix.}
\usage{
\method{FLSV}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{FLSV} returns a FLVector object representing the singular values.
}
\description{
\code{FLSV} computes the singular values for FLMatrix objects.
}
\details{
The wrapper overloads FLSV and implicitly calls FLSVUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLVector <- FLSV(flmatrix)
}

