% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLSVD}
\alias{FLSVD}
\title{Singular Value Decomposition of a matrix}
\usage{
FLSVD(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix.}
}
\value{
\code{FLLUDecomp} returns an object of class FLSVD. An object
of class FLSVD is a list containing three components of class FLMatrix
which are described as follows.
\item{u_matrix}{the U matrix}
\item{s_matrix}{the \eqn{\sum} matrix}
\item{v_matrix}{the V matrix}
All the three matrices of class FLSVD can be pulled in R using
\code{FLFetchMatrix}.
}
\description{
Formally, the singular value decomposition of an m x n matrix M is a
factorization of the form \deqn{M = U \sum V'}
where U is an m x m unitary matrix, \eqn{\sum} is an m x n rectangular diagonal
matrix with non-negative real numbers on the diagonal, and \eqn{V'} (the
conjugate transpose of V) is an n x n real or complex unitary matrix.
The diagonal entries \eqn{\sum _{i,i}} of \eqn{\sum} are known as the singular
values of M. The m columns of U and the n columns of V are called the
left-singular vectors and right-singular vectors of M, respectively.
}
\section{Constraints}{

Input matrix has a maximum dimension limitations of (550 x 550).
}
\examples{
\dontrun{
connection      <- odbcConnect("Gandalf")
database        <- "FL_R_WRAP"
matrix_table    <- "tblMatrixMulti"
matrix_id_value <- 3
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
out_svd         <- FLSVD(in_matrix)
}
}

