% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSVM.R
\docType{class}
\name{FLSVM-class}
\alias{FLSVM-class}
\title{An S4 class to represent output from Support Vector Machine(svm) on in-database Objects}
\description{
An S4 class to represent output from Support Vector Machine(svm) on in-database Objects
}
\section{Slots}{

\describe{
\item{\code{offset}}{column name used as offset}

\item{\code{vfcalls}}{information about system tables}
}}

