% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLTable.R
\docType{class}
\name{FLTable-class}
\alias{FLTable-class}
\title{An S4 class to represent FLTable}
\arguments{
\item{object}{retrieves the column names of FLTable object}
}
\description{
An S4 class to represent FLTable
}
\section{Slots}{

\describe{
\item{\code{odbc_connection}}{ODBC connectivity for R}

\item{\code{db_name}}{A character}

\item{\code{table_name}}{A character}

\item{\code{primary_key}}{A character}
}}

