% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\docType{class}
\name{FLTableFunctionQuery-class}
\alias{FLTableFunctionQuery-class}
\title{A TableFunctionQuery models a select from an arbitrary query}
\description{
A TableFunctionQuery models a select from an arbitrary query
}
\section{Slots}{

\describe{
\item{\code{SQLquery}}{character The free SQL query returning a table.}
}}

