% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLTableMD-class}
\alias{FLTableMD-class}
\title{An S4 class to represent FLTableMD, an in-database data.frame.}
\description{
An S4 class to represent FLTableMD, an in-database data.frame.
}
\section{Slots}{

\describe{
\item{\code{select}}{FLTableQuery the select statement for the table.}

\item{\code{dimnames}}{the observation id and column names}

\item{\code{isDeep}}{logical (currently ignored)}

\item{\code{mapSelect}}{\code{FLSelectFrom} object which contains the 
mapping information if any}
}}

