% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTableMD.R
\name{FLTableMD}
\alias{FLTableMD}
\title{Constructor function for FLTableMD.}
\usage{
FLTableMD(table, group_id_colname, obs_id_colname,
  var_id_colnames = character(0), cell_val_colname = character(0),
  connection = getFLConnection(), group_id = c(), fetchIDs = TRUE, ...)
}
\arguments{
\item{table}{name of the table}

\item{group_id_colname}{column name identifying the datasets}

\item{obs_id_colname}{column name set as primary key}

\item{cell_val_colname}{column name where cell values are stored if \code{FLTableMD} is deep}

\item{group_id}{vector of dataset IDs'to be considered.
Default is all those contained in the table}

\item{database}{name of the database}

\item{var_id_colname}{column name where variable id's are stored if \code{FLTableMD} is deep}
}
\value{
\code{FLTableMD} returns an object of class FLTableMD mapped to a table
in database
}
\description{
\code{FLTableMD} constructs an object of class \code{FLTableMD}.
}
\details{
\code{FLTableMD} refers to an in-database table with multiple data sets.
This object is commonly used as input for some data mining functions
which operate parallelly on all the datasets
}
\examples{
widetableMD <- FLTableMD(table="tblAutoMPGMD",
                      group_id_colname="GroupID",
                      obs_id_colname="ObsID",
                      group_id = c(2,4))
deeptableMD <- FLTableMD(table="LinRegrMultiMD",
                      group_id_colname="DatasetID",
                      obs_id_colname="ObsID",
                      var_id_colname="VarID",
                      cell_val_colname="Num_Val")
head(widetableMD)
head(deeptableMD)
}
