% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTableMDS.R
\name{FLTableMDS}
\alias{FLTableMDS}
\title{Constructor function for FLTableMDS.}
\usage{
FLTableMDS(table, group_id_colname, obs_id_colname,
  var_id_colnames = character(0), cell_val_colname = character(0),
  whereconditions = character(0), connection = getFLConnection(),
  dims = list(0, 0, 0), sparse = TRUE, dimnames = list(list(NULL),
  list(NULL), list(NULL)), ...)
}
\arguments{
\item{table}{name of the table}

\item{group_id_colname}{column name identifying the datasets}

\item{obs_id_colname}{column name set as primary key}

\item{cell_val_colname}{column name where cell values are stored if \code{FLTableMDS} is deep}

\item{database}{name of the database}

\item{var_id_colname}{column name where variable id's are stored if \code{FLTableMDS} is deep}

\item{group_id}{vector of dataset IDs'to be considered.
Default is all those contained in the table}
}
\value{
\code{FLTableMDS} returns an object of class FLTableMDS mapped to a table
in database
}
\description{
\code{FLTableMDS} constructs an object of class \code{FLTableMDS}.
}
\details{
\code{FLTableMDS} refers to an in-database table with multiple data sets.
This object is commonly used as input for some data mining functions
which operate parallelly on all the datasets
}
\examples{
widetableMD <- FLTableMDS(table="tblAutoMPGMD",
                      group_id_colname="GroupID",
                      obs_id_colname="ObsID")
deeptableMD <- FLTableMDS(table="LinRegrMultiMD",
                      group_id_colname="DatasetID",
                      obs_id_colname="ObsID",
                      var_id_colname="VarID",
                      cell_val_colname="Num_Val")
head(widetableMD)
head(deeptableMD)
}
