% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLTableQuery-class}
\alias{FLTableQuery-class}
\title{A table query models a select or a table result of a sql statement.}
\description{
A table query models a select or a table result of a sql statement.
}
\section{Slots}{

\describe{
\item{\code{connectionName}}{character variable not used right now, will be used for future support of working with several connections (to different platforms possibly)}

\item{\code{variables}}{list Named list of variables for the table query: values are rownames, names (keys) are result column names.}

\item{\code{whereconditions}}{character vector of strings restricting the select query (if any)}

\item{\code{order}}{character ordering statements (if any)}
}}

