% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTriDiag.R
\name{FLTriDiag.FLMatrix}
\alias{FLTriDiag.FLMatrix}
\title{TriDiagonal or Upper Hessenberg matrix of a FLMatrix.}
\usage{
\method{FLTriDiag}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{FLTriDiag} returns a FLMatrix object representing the upper Hessenberg or TriDiagonal matrix.
}
\description{
\code{FLTriDiag} computes the TriDiagonal or Upper Hessenberg matrix of FLMatrix object.
}
\details{
The wrapper overloads FLTriDiag and implicitly calls FLTriDiagUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- FLTriDiag(flmatrix)
}

