% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLtTest.R
\name{FLtTest}
\alias{FLtTest}
\title{Student's t-Test}
\usage{
FLtTest(table, primary_key = "ObsID", input1, input2 = NULL, mu = 0,
  var.equal = FALSE, num_tails = 2)
}
\arguments{
\item{table}{An object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to FLTable}

\item{input1}{column name for first sample input vector}

\item{input2}{(only in case of two-sample test) column name for second sample input vector}

\item{mu}{(only in case of one-sample test) reference mean}

\item{var.equal}{(only in case of two-sample test) a logical variable indicating whether to treat the two variances as being equal}

\item{num_tails}{number of tails in test distribution. This decides whether
one-tailed test or two-tailed test is performed \cr Ref: \url{http://www.ats.ucla.edu/stat/mult_pkg/faq/general/tail_tests.htm}}
}
\value{
returns the value of t-test statistic and probability (P-Value) of test statistic

If the input vector contains zero non-null observations, the function returns a NULL. Also, if the standard
deviation of values in the input vector is zero, the function returns a NULL value.
}
\description{
Performs one and two sample z-tests on vectors of data.
}
\details{
One-sample Z-test is used to test if population mean of the sample is
significantly different from a given reference value.
Two-sample Z-test is used to determine if the population means from the two
samples are equal.
}
\examples{
\dontrun{

FLtTest(	table       = tbl,
			primary_key = "ObsID",
			input1      = "Var1",
			mu          = 0.45,
			num_tails   = 2)

FLtTest(	table       = tbl,
			primary_key = "ObsID",
			input1      = "Var1",
			input2      = "Var2",
			var.equal   = FALSE,
			num_tails   = 2)

}
}

