% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLUnionMatrix-class}
\alias{FLUnionMatrix-class}
\title{An S4 class to represent FLMatrix}
\description{
An S4 class to represent FLMatrix
}
\section{Slots}{

\describe{
\item{\code{odbc_connection}}{ODBC connectivity for R}

\item{\code{by}}{character either rows or cols}
}}

