% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVarCluster.R
\name{FLVarCluster}
\alias{FLVarCluster}
\title{Variable Clustering.}
\usage{
FLVarCluster(x, ...)
}
\arguments{
\item{x}{an object of class FLTable, wide or deep}

\item{contrib}{Level of contribution expected in the
output clusters. Value between 0 and 1}

\item{matrixType}{whether a correlation matrix or a
covariance matrix should be used for Eigenvalue decomposition.
Allowed values c("COVAR","CORREL")}

\item{groupBy}{Comma separated column names identifying
each data set. Currently not used and NULL always.}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{FLVarCluster} returns a R vector if data can be fetched
or a FLVector of cluster to which each variable or column is assigned.
}
\description{
\code{FLVarCluster} performs variable clustering on FLTable objects
using Principal Component Analysis
}
\details{
The DB Lytix function called is FLVarCluster. Uses a principal component analysis for dimensionality 
reduction in order to cluster a given set of input variables into a smaller representative set.
The number of output clusters depend on the contribution level specified.
}
\section{Constraints}{

If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}

\examples{
deeptable  <- FLTable(getTestTableName("tblLogRegr"), "ObsID","VarID",
                      "Num_Val", whereconditions= "ObsID<101")                        
clustervector <- FLVarCluster(deeptable,0.75,"COVAR",whereconditions=" VarID>0 ")
print(clustervector)
}
\seealso{
\code{\link[ClustOfVar]{ClustOfVar}} package for R reference implementation.
}
