% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCorrel.R
\name{FLVarP}
\alias{FLVarP}
\title{population variance.}
\usage{
FLVarP(x, y = NULL, ...)
}
\arguments{
\item{x}{FLMatrix, FLVector or FLTable object or any R object}

\item{y}{FLMatrix, FLVector or FLTable object or any R object}

\item{...}{any additional arguments}
}
\value{
\code{cov} returns FLMatrix object representing correlation of x and y.
}
\description{
\code{cov} computes correlation of in-database Objects
}
\section{Constraints}{

The number of non-null pairs must be greater than or equal to 2.
If number of non-null pairs is less than 2, FLCorrel returns a NULL.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
deeptable <- FLTable( "FL_DEMO", 
"tblUSArrests", "ObsID","VarID","Num_Val")
widetable <- FLTable("FL_DEMO","tblAbaloneWide","ObsID")
FLCovarP(deeptable,deeptable)
FLCovarP(widetable,widetable)
}

