% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVector.R
\docType{class}
\name{FLVector-class}
\alias{FLVector-class}
\title{An S4 class to represent FLVector}
\description{
An S4 class to represent FLVector
}
\section{Slots}{

\describe{
\item{\code{table}}{object of class FLTable}

\item{\code{col_name}}{A character column name}

\item{\code{vector_id_value}}{numeric}

\item{\code{size}}{numeric}
}}

