% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVector.R
\name{FLVector}
\alias{FLVector}
\title{Constructor function for FLVector, representing a vector in database}
\usage{
FLVector(table, val_col_name = character(), val_row_name = character(),
  whereconditions = character())
}
\value{
\code{FLVector} returns an object of class FLVector mapped to an in-database vector.
}
\description{
Please use subsetting of FLTable to create FLVector object
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
WideTable <- FLTable(connection, "FL_DEMO", "tblAbaloneWide","ObsID")
flvectorColumn <- FLTable[,"Diameter"]
flvectorRow <- FLTable[3,]
flvectorRow
flvectorColumn
}
\seealso{
\code{\link{FLTable}}
}

