% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{FLVectorForColumnWhere}
\alias{FLVectorForColumnWhere}
\title{Creates FLVector from a column from a database table after applying WHERE conditions.}
\usage{
FLVectorForColumnWhere(data, column, where)
}
\arguments{
\item{data}{FLTable to select from}

\item{column}{Column to select}

\item{where}{additional where conditions for restriction}
}
\description{
Creates FLVector from a column from a database table after applying WHERE conditions.
}

