% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman.test.R
\name{friedman.test.FLVector}
\alias{friedman.test.FLVector}
\title{Friedman Rank Sum Test}
\usage{
friedman.test.FLVector(y, groups, blocks, ...)
}
\arguments{
\item{y}{FLVector with data values or data FLMatrix}

\item{groups}{FLVector giving the group for the 
corresponding elements of y if y is a FLVector; ignored if y is FLMatrix.}

\item{blocks}{FLVector giving the block for the 
corresponding elements of y if y is a FLVector; ignored if y is FLMatrix.}

\item{...}{The additional arguments used by FL function:
\code{whereconditions} WhereConditions to subset data
\code{GroupBy} Column names defining the different groups in data, if any.}

\item{formula}{a formula of the form a ~ b | c, where a, b and c 
give the data values and corresponding groups and blocks, respectively.
Not applicable if FLVector or FLMatrix is input.}

\item{data}{FLTable or FLTableMD objects.}

\item{subset}{Not currently used.}

\item{na.action}{na values omitted always.}
}
\value{
A list with class \code{htest}.
A list of \code{htest} objects if the input is a FLTableMD object.
}
\description{
Performs a Friedman rank sum test with unreplicated blocked data.
}
\examples{
RoundingTimes <-
matrix(c(5.40, 5.50, 5.55,
       5.85, 5.70, 5.75,
      5.20, 5.60, 5.50,
      5.55, 5.50, 5.40,
      5.90, 5.85, 5.70,
      5.45, 5.55, 5.60,
      5.40, 5.40, 5.35,
      5.45, 5.50, 5.35,
      5.25, 5.15, 5.00,
      5.85, 5.80, 5.70,
      5.25, 5.20, 5.10,
      5.65, 5.55, 5.45,
      5.60, 5.35, 5.45,
      5.05, 5.00, 4.95,
      5.50, 5.50, 5.40,
      5.45, 5.55, 5.50,
      5.55, 5.55, 5.35,
      5.45, 5.50, 5.55,
      5.50, 5.45, 5.25,
      5.65, 5.60, 5.40,
      5.70, 5.65, 5.55,
      6.30, 6.30, 6.25),
    nrow = 22,
    byrow = TRUE,
    dimnames = list(1 : 22,
                    c("Round Out", "Narrow Angle", "Wide Angle")))
FLMatrixObj <- as.FLMatrix(RoundingTimes)
result1 <- friedman.test(FLMatrixObj)
print(result1)
wb <- aggregate(warpbreaks$breaks,
             by = list(w = warpbreaks$wool,
                       t = warpbreaks$tension),
             FUN = mean)
FLTableObj <- as.FLTable(wb,tableName="ARBaseTestTempTable",drop=TRUE)
result2 <- friedman.test(FLTableObj$x, FLTableObj$w, FLTableObj$t)
print(result2)
result3 <- friedman.test(x~w|t, data = FLTableObj)
print(result3)
fltMD <- FLTableMD("tblFriedmanTest","datasetid","obsid","groupid","num_val")
result4 <- friedman.test(x~w|t, data = fltMD)
print(result4)
}

