% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLGAM.R
\name{gam}
\alias{gam}
\title{Generalized Additive Models}
\usage{
gam(formula, family = stats::gaussian, data, ...)
}
\arguments{
\item{formula}{An R formula object with smooth terms \code{s},
\code{te}. In these smooth terms, \code{m} and \code{k} are specified which
represent splineDegree and number of knots respectively.In
addition, NumOfPredPts is specified through \code{xt} and byVarName
through \code{by}.One Can club Spline terms with same set of
specifications by passing more than one variables/colnames in \code{s}.}

\item{family}{currently only stats::poisson is supported}

\item{data}{an object of class FLTable, must be a wide table}

\item{...}{other arguments as in mgcv::gam may be passed
but not currently used}

\item{offset}{column name indicating the offset if any}

\item{maxiter}{maximum num of iterations}
}
\value{
\code{gam} returns \code{FLGAM} object
}
\description{
\code{gam} Fits a generalized additive model
 to data given by FLTable object.
}
\section{Constraints}{

Plotting is not available. The result Set may not include all
results and methods as in mgcv::gam.
}
\examples{
widetable <- FLTable("FL_DEMO","tblGAMSimData","ObsID")
myformula <- yVal~x0Val+s(x1Val,m=3,k=10)+te(x1Val,x2Val,m=3,k=5)+s(x2Val,x1Val)
gamobject <- gam(myformula,data=widetable,offset="x2Val")
predictedValues <- predict(gamobject,widetable)
gamobject$coefficients
gamobject$fitted.values
gamobject$residuals
gamobject$df.null
print(gamobject)
}

