% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLGarch.R
\name{garch.FLVector}
\alias{garch.FLVector}
\title{rv <- sqlQuery(connection, "SELECT stockreturn FROM tblbac_return")
flv <- as.FL(rv$stockreturn)
garch(flv)}
\usage{
\method{garch}{FLVector}(data, degree = 2, ...)
}
\description{
rv <- sqlQuery(connection, "SELECT stockreturn FROM tblbac_return")
flv <- as.FL(rv$stockreturn)
garch(flv)
}
