% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMaxMatrixId}
\alias{getMaxMatrixId}
\title{Get Max Matrix ID+1 from result Matrix table}
\usage{
getMaxMatrixId(vconnection = getOption("connectionFL"),
  vtable = getOption("ResultMatrixTableFL"), ...)
}
\arguments{
\item{vconnection}{ODBC/JDBC connection object}
}
\description{
used to know ID of next entry in table
}

