% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMaxValue}
\alias{getMaxValue}
\title{Get Max ID from given table}
\usage{
getMaxValue(vdatabase = getOption("ResultDatabaseFL"),
  vtable = getOption("ResultVectorTableFL"), vcolName = "vectorIdColumn",
  vconnection = vconnection)
}
\arguments{
\item{vdatabase}{name of the database of table}

\item{vtable}{name of the table}

\item{vcolName}{name of the primary index column in table}

\item{vconnection}{ODBC/JDBC connection object}
}
\description{
used to know ID of last entry in table
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

