% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMaxVectorId}
\alias{getMaxVectorId}
\title{Get Max Vector ID+1 from result Vector table}
\usage{
getMaxVectorId(vconnection = getFLConnection(),
  vtable = getOption("ResultVectorTableFL"), ...)
}
\arguments{
\item{vconnection}{ODBC/JDBC connection object}
}
\description{
used to know ID of next entry in table
}

