% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLGinv.R
\name{ginv.FLMatrix}
\alias{ginv.FLMatrix}
\title{Generalized Inverse of a Matrix.}
\usage{
\method{ginv}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{ginv} returns a FLMatrix object which is the pseudo-inverse 
of input FLMatrix object and replicates the equivalent R output.
}
\description{
\code{ginv} computes the pseudo-inverse for FLMatrix objects.
}
\details{
The wrapper overloads ginv and implicitly calls FLMatrixPseudoInvUdt.
}
\section{Constraints}{

Input can only be with maximum dimension limitations of (500 x 500).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
resultFLMatrix <- ginv(flmatrix)
}

