% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLGinv.R
\name{ginv}
\alias{ginv}
\title{Generalized Inverse of a Matrix.}
\usage{
ginv(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{ginv} returns a FLMatrix object which is the pseudo-inverse 
of input FLMatrix object and replicates the equivalent R output.
}
\description{
\code{ginv} computes the pseudo-inverse for FLMatrix objects.
}
\section{Constraints}{

Input can only be with maximum dimension limitations of (500 x 500).
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- ginv(flmatrix)
}

