% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLogRegr.R, R/FLPoissonRegr.R
\name{glm.FLTable}
\alias{glm.FLTable}
\title{Linear Regression.}
\usage{
\method{glm}{FLTable}(formula, family = stats::poisson, data, offset = "",
  maxiter = 25, catToDummy = 0, performNorm = 0, performVarReduc = 0,
  makeDataSparse = 0, minStdDev = 0, maxCorrel = 1, classSpec = list(),
  whereconditions = "", ...)

\method{glm}{FLTable}(formula, family = stats::poisson, data, offset = "",
  maxiter = 25, catToDummy = 0, performNorm = 0, performVarReduc = 0,
  makeDataSparse = 0, minStdDev = 0, maxCorrel = 1, classSpec = list(),
  whereconditions = "", ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable}
}
\value{
\code{glm} performs linear regression and replicates equivalent R output.
}
\description{
\code{glm} performs linear regression on FLTable objects.
}
\details{
The wrapper overloads glm and implicitly calls FLRegrDataPrep and FLLogRegr.
}
\section{Constraints}{

None
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable( "FL_REV4546", "tblAbaloneWide", "ObsID")
glmfit <- glm(Rings~Height+Diameter,widetable)
library(RODBC)
connection <- flConnect(odbcSource = "Gandalf",database = "FL_DEV")
widetable  <- FLTable("FL_DEV", "siemenswidetoday1", "ObsID")
options(debugSQL=T)
pfitToffA <- glm(event ~ meanTemp, family=poisson, data=widetable,offset="age")
predData <- FLTable("FL_DEV","preddata1","ObsID")
mu <- predict(pfitToffA,newdata=predData)
}

