% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixArithmetic.R
\name{\%\%}
\alias{\%\%}
\title{remainder of division on in-database objects.}
\usage{
pObj1 \%\% pObj2
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{\%\%} returns an in-database object if there is atleast one in-database object 
as input.Otherwise, the default behavior of R is preserved
}
\description{
\code{\%\%} calculates the remainder of in-database object division.
}
\details{
The remainder of in-database objects mimics the normal remainder of R data types.
All combinations of operands are possible just like in R and the result is an in-database object.
}
\section{Constraints}{
 division by 0 gives inf in R, but is not supported for 
in-database objects
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
Rvector <- 1:5
ResultFLmatrix <- flmatrix \%\% Rvector
}
\author{
phani srikar <phanisrikar93ume@gmail.com>
}

