% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{grepl}
\alias{grepl}
\title{Pattern Matching}
\usage{
grepl(pattern, x, ignore.case = FALSE, perl = FALSE, fixed = FALSE,
  useBytes = FALSE)
}
\arguments{
\item{pattern}{string to search for}

\item{x}{FLVector of characters or R vector
where matches are sought}

\item{ignore.case}{logical indicating case-sensitivity.
Currently always FALSE for FLVectors}

\item{perl}{logical. Should perl-compatible regexps be used?
Always FALSE for FLVectors}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.
For FLVectors, regualar expressions are not supported}

\item{useBytes}{If TRUE the matching is done byte-by-byte
rather than character-by-character. Always FALSE for FLVector}
}
\value{
for FLVector input as x, FLVector with 1 for found
and 0 for no match is returned.Else R Vector as in base::grepl
is returned.
}
\description{
Match \code{pattern} in each element of \code{x}
}
\section{Constraints}{

row FLVectors are not supported currently.
Output slightly differs from base::grepl. See \code{return}
}
\examples{
widetable  <- FLTable("tblstringID", "stringID")
flv <- widetable[1:6,"string"]
flvector <- grepl("A",flv)
}
\seealso{
\code{\link[base]{grepl}} for R function reference implementation.
}

