% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{gsub}
\alias{gsub}
\title{Pattern Matching and Replacement}
\usage{
gsub(pattern, replacement, x, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, useBytes = FALSE)
}
\arguments{
\item{pattern}{string to search for}

\item{x}{FLVector of characters or R vector
where matches are sought}

\item{ignore.case}{logical indicating case-sensitivity.
Currently always FALSE for FLVectors}

\item{perl}{logical. Should perl-compatible regexps be used?
Always FALSE for FLVectors}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.
For FLVectors, regualar expressions are not supported}

\item{useBytes}{If TRUE the matching is done byte-by-byte
rather than character-by-character. Always FALSE for FLVector}

\item{the}{replacement character}
}
\value{
FLVector or R vector after replacement
}
\description{
Replace \code{pattern} in each element of \code{x}
with \code{replacement}
}
\details{
The DB Lytix function called is FLReplaceChar.The replace character function is a
scaler that replaces a character in a string with a another character.
}
\section{Constraints}{

row FLVectors are not supported currently.
Currently only one character is used for replacement.
}

\examples{
widetable  <- FLTable("tblstringID", "stringID")
flv <- widetable[1:6,"string"]
flvector <- gsub("A","X",flv)
}
\seealso{
\code{\link[base]{gsub}} for R function reference implementation.
}
