% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{hamming.distance}
\alias{hamming.distance}
\title{hamming.distance}
\usage{
hamming.distance(xsource, targets, caseFlag = 0, vlength = 3, ...)
}
\arguments{
\item{xsource}{character or FLVector of characters}

\item{targets}{character or FLVector of characters}

\item{vlength}{optional, length of strings to compare}

\item{caseFLag}{logical or 0/1 indicating
if comparision should be case sensitive}
}
\value{
FLMatrix if both \code{xsource} and \code{targets}
are FLVectors. Otherwise returns a FLVector
}
\description{
computes the hamming distance between strings
}
\section{Constraints}{

row vectors are not supported currently
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflvector <- hamming.distance("xyz",flv)
resultflmatrix <- hamming.distance(flv,flv,caseFLag=1)
resultflvector <- hamming.distance(flv,"xyz",caseFLag=1)
}

