% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLHClust.R
\name{hclust.FLTable}
\alias{hclust.FLTable}
\title{Hierarchial Clustering}
\usage{
\method{hclust}{FLTable}(d, method = "average", maxit = 500,
  excludeCols = "", classSpec = list(), whereconditions = "", ...)
}
\arguments{
\item{d}{an object of class FLTable, can be wide or deep table}

\item{method}{character. Allowed methods are "average",
"single", "complete", "centroid"}

\item{maxit}{maximum number of iterations}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{hclust} returns a list and replicates equivalent R output
from \code{hclust} in stats package.The mapping table can be viewed
using \code{object$mapping} if input is wide table.
}
\description{
\code{hclust} computes hierarchial 
clustering on FLTable objects.
}
\section{Constraints}{

Error is thrown if results cannot be fetched. maxit should be more than
no.of. observations for algorithm to reach completion.
Error is thrown if algorithm does not reach completion or more than one
cluster is formed at any step.


If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}

\examples{
deeptable  <- FLTable("tblUSArrests", "ObsID","VarID","Num_Val")
hclustobject <- hclust(deeptable,maxit=50)
print(hclustobject)
plot(hclustobject)
}
