% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLHeadTail.R
\name{head.FLTable}
\alias{head.FLTable}
\title{Return the First or Last Part of an in-database Object}
\usage{
head.FLTable(x, n = 6L, ...)
}
\arguments{
\item{x}{an in-database object(FLVector, FLMatrix or FLTable)}

\item{n}{a single integer. If positive, size for the resulting object: number of elements for a vector 
(including lists), rows for a matrix or data frame or lines for a function. If negative, all but the
n last/first number of elements of x.}
}
\value{
returns an object (usually) like x i.e. an in-database object but generally smaller.
}
\description{
\code{head()} and \code{tail()} functions return the first or last parts of a FLVector, FLMatrix orFLTable.
}
\section{Usage}{
 \code{tail.FLMatrix(x, n=6L, ...)}
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO.tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
headflmatrix <- head(flmatrix, n=6L, ...)
print(headflmatrix)
}
\seealso{
\code{\link[utils]{head}} , \code{\link[utils]{tail}} for R reference implementation.
}

