% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLHessenDecomp.R
\name{hessen.FLMatrix}
\alias{hessen.FLMatrix}
\title{Hessenberg Decomposition of a Matrix.}
\usage{
\method{hessen}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{hessen} returns a list of two components:
      \item{P}{FLMatrix representing P matrix obtained from Hessenberg decomposition}
      \item{H}{FLMatrix representing H matrix obtained from Hessenberg decomposition}
}
\description{
\code{hessen} computes the Hessenberg decomposition for FLMatrix objects.
}
\details{
The wrapper overloads hessen and implicitly calls FLHessenbergDecompUdt.
}
\section{Constraints}{

Input can only be square matrix with maximum dimension limitations of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultList <- hessen(flmatrix)
resultList$P
resultList$H
}

