% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLHKMeans.R
\name{hkmeans}
\alias{hkmeans}
\title{Hierarchial K-Means Clustering.}
\usage{
hkmeans(x, ...)
}
\arguments{
\item{x}{an object of class FLTable, wide or deep}

\item{centers}{the number of clusters}

\item{levels}{no.of.levels in the hierarchy}

\item{iter.max}{the maximum number of iterations allowed}

\item{nstart}{the initial number of random sets}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{hkmeans} returns a list which replicates equivalent R output
from \code{hkmeans} in stats package.The mapping table can be viewed
using \code{object$mapping} if input is wide table.
}
\description{
\code{hkmeans} performs Hierarchial k-means clustering on FLTable objects.
}
\details{
The DB Lytix function called is FLHKMeans.Hierarchical K-Means clusters the training data.  
The relationship of observations to clusters has hard edges. It re-clusters the training data in 
each cluster until the desired hierarchical level is reached.
}
\section{Constraints}{

If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}
\examples{
widetable  <- FLTable("tblAbaloneWide", "ObsID")
hkmeansobject <- hkmeans(widetable,3,2,20,1,"Rings,SEX")
print(hkmeansobject)
plot(hkmeansobject)
One can specify ClassSpec and transform categorical variables 
before clustering. This increases the number of variables in the plot
because categorical variable is split into binary numerical variables.
The clusters may not be well-defined as is observed in the case below:-
hkmeansobjectnew <- hkmeans(widetable,3,2,20,1,"Rings,SEX",list("DummyCat(D)","SEX(M)"))
plot(hkmeansobjectnew)
}

