% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoltWinters.R
\name{HoltWinters}
\alias{HoltWinters}
\title{Holt-Winters Filtering}
\usage{
HoltWinters(object, ...)
}
\arguments{
\item{object}{FLVector}

\item{alpha}{alpha parameter of Holt-Winters Filter.}

\item{beta}{beta parameter of Holt-Winters Filter. If set to FALSE, the function will do exponential smoothing.}

\item{gamma}{gamma parameter used for the seasonal component. If set to FALSE, an non-seasonal model is fitted.}

\item{periodicity}{The period across which the data shows a seasonal trend.}

\item{forecastperiod}{The period for which the forecasted values are to be calculated.}

\item{normalization}{If set to 1, the mean is added to values to normalize the data.}
}
\value{
An object containing details of the fitted model and forecasted values.

A list with class "htest" containing the following components:

\describe{
\item{fitted}{the forecasted values}
\item{x}{the data}
\item{alpha}{the alpha value used for forecasting}
\item{beta}{the beta value used for forecasting}
\item{gamma}{the gamma value used for forecasting}
\item{seasonal}{the type of seasonal model used, additive or multiplicative}
\item{call}{The call used}
}
}
\description{
Computes Holt-Winters Filtering of a given time series. Unknown parameters are determined by minimizing the squared prediction error.
}
\details{
\description{ The additive Holt-Winters prediction function (for time series with period length p) is
Yhat[t+h] = a[t] + h * b[t] + s[t - p + 1 + (h - 1) mod p],
where a[t], b[t] and s[t] are given by
a[t] = α (Y[t] - s[t-p]) + (1-α) (a[t-1] + b[t-1])
b[t] = β (a[t] - a[t-1]) + (1-β) b[t-1]
s[t] = γ (Y[t] - a[t]) + (1-γ) s[t-p]
The multiplicative Holt-Winters prediction function (for time series with period length p) is
Yhat[t+h] = (a[t] + h * b[t]) * s[t - p + 1 + (h - 1) mod p],
where a[t], b[t] and s[t] are given by
a[t] = α (Y[t] / s[t-p]) + (1-α) (a[t-1] + b[t-1])
b[t] = β (a[t] - a[t-1]) + (1-β) b[t-1]
s[t] = γ (Y[t] / a[t]) + (1-γ) s[t-p]
The data in x are required to be non-zero for a multiplicative model, but it makes most sense if they are all positive.}
}
\examples{
x<-rnorm(1000)
flv<-as.FLVector(x)
flobj<-HoltWinters(object = flv, alpha = 0.5)
}

