% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIdentical.R
\name{identical}
\alias{identical}
\title{Equality of in-database objects.}
\usage{
identical(x, y)
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{identical} returns a logical TRUE or FALSE.
}
\description{
\code{identical} checks the equality of in-database objects.
}
\details{
The equality of in-database objects mimics the normal addition of R data types.
One can check equality of FLMatrices, FLMatrix - R matrices, FLVectors and
FLVector - RVector.
}
\section{Constraints}{

Currently only \code{dgCMatrix},\code{dgeMatrix},\code{dsCMatrix},
\code{dgTMatrix},\code{matrix},\code{Matrix},\code{vector} R types
are supported.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
Rvector <- 1:5
Result <- identical(flmatrix,flmatrix)
Result <- identical(Rvector,as.FLVector(Rvector,connection))
}

