% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{initFdefault}
\alias{initFdefault}
\title{initF.default helps to return a list of list.
Can be used for comparing results of R and FL functions which require two objects.}
\usage{
initFdefault(specs = list(c(n = 5, isSquare = TRUE), c(n = 5, isRowVec =
  FALSE)), classes = c("FLMatrix", "FLVector"))
}
\description{
initF.default helps to return a list of list.
Can be used for comparing results of R and FL functions which require two objects.
}
