% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{initFgeneric}
\alias{initFgeneric}
\title{initF.default helps to return a list of list.
Can be used for comparing results of R and FL functions which require two objects.}
\usage{
initFgeneric(specs = list(numberattribute = 5, featureattribute = TRUE, ...),
  class = "FLMatrix")
}
\description{
initF.default helps to return a list of list.
Can be used for comparing results of R and FL functions which require two objects.
}
