% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLIs.R
\name{is.FLSparseMatrix}
\alias{is.FLSparseMatrix}
\title{Check if the object is an FLSparseMatrix object}
\usage{
is.FLSparseMatrix(object)
}
\description{
Check if the object is an FLSparseMatrix object
}

