% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLJordanDecomp.R
\name{jordan.FLMatrix}
\alias{jordan.FLMatrix}
\title{Jordan Decomposition of a Matrix.}
\usage{
\method{jordan}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{jordan} returns a list of two components:
      \item{J}{FLVector representing J vector obtained from Jordan decomposition}
      \item{P}{FLMatrix representing P matrix obtained from Jordan decomposition}
      \item{PInv}{FLMatrix representing PInv matrix obtained from Jordan decomposition}
}
\description{
\code{jordan} computes the Jordan decomposition for FLMatrix objects.
}
\details{
The wrapper overloads jordan and implicitly calls FLJordanDecompUdt.
}
\section{Constraints}{

Input can only be square matrix with maximum dimension limitations of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultList <- jordan(flmatrix)
resultList$J
resultList$P
resultList$PInv
}

