% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{kmeans}
\alias{kmeans}
\title{K-Means Clustering.
\code{kmeans} performs k-means clustering on FLTable objects.}
\usage{
\method{kmeans}{FLTable}(x, ...)
}
\arguments{
\item{x}{an object of class FLTable, wide or deep}

\item{centers}{the number of clusters}

\item{iter.max}{the maximum number of iterations allowed}

\item{nstart}{the initial number of random sets}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{kmeans} returns a list which replicates equivalent R output
from \code{kmeans} in stats package. The mapping table can be viewed
using \code{object$mapping} if input is wide table.
}
\description{
The DB Lytix function called is FLKMeans.K-Means clusters the training data. 
The relationship of observations to clusters has hard edges. In general, k-means has 
two steps: assigning data points to the nearest cluster and then moving
each cluster’s centroid to the center of the members of the cluster.
}
\section{Constraints}{

If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
widetable  <- FLTable( "FL_DEMO", "tblAbaloneWide", "ObsID")
kmeansobject <- kmeans(widetable,3,20,1,"Rings,SEX")
print(kmeansobject)
plot(kmeansobject)
One can specify ClassSpec and transform categorical variables 
before clustering. This increases the number of variables in the plot
because categorical variable is split into binary numerical variables.
The clusters may not be well-defined as is observed in the case below:-
kmeansobjectnew <- kmeans(widetable,3,20,1,"Rings,SEX",list("DummyCat(D)","SEX(M)"))
plot(kmeansobjectnew)
}
\seealso{
\code{\link[stats]{kmeans}} for R reference function implementation.
}

