% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKNN.R
\name{knn}
\alias{knn}
\title{k-Nearest Neighbour 
k-nearest neighbour classification/Regression for test set from training set. 
For each row of the test set, the k nearest (according to distance metric speicified) 
training set vectors are found, and the classification is decided by 
majority vote, with ties broken at random. If there are ties for the 
kth nearest vector, all candidates are included in the vote.
If classify flag is false, average of k neighbours is returned.}
\usage{
knn(train, test = train, cl = "NULL", k = 1, l = 0, prob = FALSE,
  use.all = FALSE, ...)
}
\arguments{
\item{train}{input deep FLTable}

\item{test}{input deep FLTable}

\item{cl}{ColumnName of true classifications of training set}

\item{k}{number of neighbours considered.}

\item{prob}{If this is true, 
the proportion of the votes for the winning class are returned as attribute prob.}

\item{classify}{logical if classification/regression is solved}

\item{metric}{distance metric to be used. euclidean,manhattan supported.}
}
\value{
FLVector of classifications of test set.
}
\description{
k-Nearest Neighbour 
k-nearest neighbour classification/Regression for test set from training set. 
For each row of the test set, the k nearest (according to distance metric speicified) 
training set vectors are found, and the classification is decided by 
majority vote, with ties broken at random. If there are ties for the 
kth nearest vector, all candidates are included in the vote.
If classify flag is false, average of k neighbours is returned.
}
\examples{
FLdeepTbl <- FLTable(getTestTableName("ARknnDevSmall"),"obsid","varid","num_val")
FLknnOutput <- knn(FLdeepTbl,FLdeepTbl,k=3,prob=TRUE)
FLknnOutput
attributes(FLknnOutput)$prob
}
