% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnx.index.R
\name{knnx.index}
\alias{knnx.index}
\title{knnx.index}
\usage{
knnx.index(data, query, k = 1, algorithm = "kd_tree", ...)
}
\arguments{
\item{data}{input FLTable with pool space to search for}

\item{query}{input search FLTable}

\item{k}{number of neighbours considered.}

\item{algorithm}{currently only brute is supported}

\item{...}{additional parameters like dist Flag to get the distances as well.}
}
\value{
FLMatrix of indices or a list of dist & index if dist FLag is true.
}
\description{
k-Nearest Neighbour Indexes
}
\details{
For each row of the test set, the k nearest (according to Euclidean distance metric) 
training set vectors are found and indices,distances are returned.
}
\examples{
FLdeepTbl <- FLTable(getTestTableName("ARknnDevSmall"),"obsid","varid","num_val")
FLknnOutput <- knnx.index(FLdeepTbl,FLdeepTbl,k=2)
FLknnOutput
}
