% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.test.R
\name{kruskal.test.FLVector}
\alias{kruskal.test.FLVector}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
\method{kruskal.test}{FLVector}(x, g, ...)
}
\arguments{
\item{x}{FLVector with data values}

\item{g}{FLVector giving the group for the 
corresponding elements of y}

\item{...}{The additional arguments used by FL function:
\code{whereconditions} WhereConditions to subset data
\code{GroupBy} Column names defining the different groups in data, if any.}

\item{formula}{a formula of the form response ~ group 
where response gives the data values and group a vector or factor 
of the corresponding groups.
Not applicable if FLVector is input.}

\item{data}{FLTable or FLTableMD objects.}

\item{subset}{Not currently used.}

\item{na.action}{na values omitted always.}
}
\value{
A list with class \code{htest}.
A list of \code{htest} objects if the input is a FLTableMD object.
}
\description{
Performs a Kruskal-Wallis rank sum test.
}
\examples{
x <- c(2.9, 3.0, 2.5, 2.6, 3.2) # normal subjects
y <- c(3.8, 2.7, 4.0, 2.4)      # with obstructive airway disease
z <- c(2.8, 3.4, 3.7, 2.2, 2.0) # with asbestosis
x <- c(x, y, z)
g <- factor(rep(1:3, c(5, 4, 5)),
            labels = c("Normal subjects",
                       "Subjects with obstructive airway disease",
                       "Subjects with asbestosis"))
x <- as.FLVector(x)
g <- as.FLVector(g)
result1 <- kruskal.test(x, g)
print(result1)
FLTableObj <- as.FLTable(airquality,tableName="ARBaseTestTempTable",drop=TRUE)
result2 <- kruskal.test(Ozone ~ Month, data = FLTableObj)
print(result2)
}
