% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKSTest.R
\docType{methods}
\name{ks.test,FLVector,ANY-method}
\alias{ks.test,FLVector,ANY-method}
\title{KS TEst
Perform a one- or two-sample Kolmogorov-Smirnov test.}
\usage{
\S4method{ks.test}{FLVector,ANY}(x, y, mean = NULL, sd = NULL)
}
\arguments{
\item{x}{a FLVector of data values.}

\item{y}{either a FLVector of data values, or a character string naming a cumulative distribution function or 
an actual cumulative distribution function such as pnorm. Only continuous CDFs are valid.}
}
\value{
A list with class "htest".
}
\description{
KS TEst
Perform a one- or two-sample Kolmogorov-Smirnov test.
}
\section{Constraints}{
 As of now only supports normal disctribution, alternative and exact isn't supported for FL objects
}
\examples{
set.seed(100)                                       
p <- as.FLVector(rnorm(50))
q <- as.FLVector(runif(30))
res <- ks.test(p, q)
If y is a FLVector, a two-sample test of the null hypothesis that x and y were drawn from the same continuous distribution is performed.
Alternatively, y can be a character string naming a continuous (cumulative) distribution function, or such a function. 
In this case, a one-sample test is carried out of the null that the distribution function which generated x is distribution y with parameters.
}

