% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLDA.R
\name{lda}
\alias{lda}
\title{Linear discriminant analysis.
\code{lda} performs Linear discriminant analysis on FLTable objects.}
\usage{
lda(formula, data = list(), ...)
}
\arguments{
\item{formula}{A formula of the form groups ~ x1 + x2 + ... That is, the response is the grouping factor and the right hand side specifies the (non-factor) discriminators.}

\item{data}{An object of class FLTable or FLTableMD.}
}
\value{
\code{LDA} returns an object of class \code{FLLDA}
}
\description{
The DB Lytix function called is FLLDA. Performs Linear discriminant analysis and 
stores the results in predefined tables.
}
\examples{
tbl <- FLTable("tblLDA", "OBSID", "VARID", "NUM_VAL")
flmod <- lda(a~. , data = tbl)
flmod$scaling, flmod$means;cof <-coefficients(flmod)
predict(flmod): Not implemented yet.
plot(flmod)
}
\seealso{
\code{\link[stats]{lda}} for R reference implementation.
}

