% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLDA.R
\name{ldaGeneric}
\alias{ldaGeneric}
\title{tbl <- FLTable("tblLDA", "OBSID", "VARID", "NUM_VAL")
flmod <- lda(a~. , data = tbl)
flmod$scaling, flmod$means}
\usage{
ldaGeneric(formula, data, callObject = NULL, familytype = "",
  MaxMARS = MaxMARS, MinRsq = MinRsq, ...)
}
\description{
tbl <- FLTable("tblLDA", "OBSID", "VARID", "NUM_VAL")
flmod <- lda(a~. , data = tbl)
flmod$scaling, flmod$means
}

