% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLength.R
\name{length.FLMatrix}
\alias{length.FLMatrix}
\title{computes the length of FLMatrix object.}
\usage{
\method{length}{FLMatrix}(obj)
}
\arguments{
\item{obj}{is a FLMatrix object.}
}
\value{
\code{length} returns a R Vector giving the length of input object.
}
\description{
computes the length of FLMatrix object.
}

