% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{length.FLSimpleVector}
\alias{length.FLSimpleVector}
\title{computes the length of FLVector object.}
\usage{
\method{length}{FLSimpleVector}(obj)
}
\arguments{
\item{obj}{is a FLVector object.}
}
\value{
\code{length} returns a R Vector giving the length of input object.
}
\description{
computes the length of FLVector object.
}

