% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLength.R
\name{length.FLSparseMatrix}
\alias{length.FLSparseMatrix}
\title{computes the length of FLSparseMatrix object.}
\usage{
\method{length}{FLSparseMatrix}(obj)
}
\arguments{
\item{obj}{is a FLSparseMatrix object.}
}
\value{
\code{length} returns a R Vector giving the length of input object.
}
\description{
computes the length of FLSparseMatrix object.
}

