% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLength.R
\name{length.FLTable}
\alias{length.FLTable}
\title{computes the length of FLTable object.}
\usage{
\method{length}{FLTable}(obj)
}
\arguments{
\item{obj}{is a FLTable object.}
}
\value{
\code{length} returns a R Vector giving the
number of observations or rows in input FLTable object.
}
\description{
computes the length of FLTable object.
}

