% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLength.R
\name{length.FLVector}
\alias{length.FLVector}
\title{computes the length of FLVector object.}
\usage{
\method{length}{FLVector}(obj)
}
\arguments{
\item{obj}{is a FLVector object.}
}
\value{
\code{length} returns a R Vector giving the length of input object.
}
\description{
computes the length of FLVector object.
}

