% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{lm.FLTable}
\alias{lm.FLTable}
\title{Linear Regression.}
\usage{
\method{lm}{FLTable}(formula, data, catToDummy = 0, performNorm = 0,
  performVarReduc = 0, makeDataSparse = 0, minStdDev = 0, maxCorrel = 1,
  classSpec = list(), whereconditions = "", ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable}
}
\value{
\code{lm} performs linear regression and replicates equivalent R output.
}
\description{
\code{lm} performs linear regression on FLTable objects.
}
\details{
The wrapper overloads lm and implicitly calls FLRegrDataPrep and FLLinRegr.
}
\section{Constraints}{

None
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable("FL_DEMO", "tblAbaloneWide", "ObsID")
lmfit <- lm(Rings~Height+Diameter,widetable)
deeptable <- FLTable("FL_DEMO","myLinRegrSmall","ObsID","VarID","Num_Val")
lmfit1 <- lm(NULL,deeptable)
deeptable <- FLTable("FL_DEMO","myLinRegrSmallBroken","ObsID","VarID","Num_Val")
lmfit2 <- lm(NULL,deeptable)
}

