% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/FLMixedmodel.R
\name{lmer.FLTable}
\alias{lmer.FLTable}
\title{\code{lmer} performs mixed model  on FLTable objects.
The DB Lytix function called is FLLinMixedModel. The mixed model extends the linear
model by allowing correlation and heterogeneous variances in the
covariance matrix of the residuals. Fit a linear mixed-effects model (LMM) to data,
FLLinMixedModel estimates the coefficients and covariance matrix of the mixed model
via the expectations.}
\usage{
\method{residuals}{FLMix}(formula, data, fetchID = TRUE, maxiter = 10, ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable.}
}
\value{
\code{lmer} returns an object of class \code{FLMix}
}
\description{
\code{lmer} performs mixed model  on FLTable objects.
The DB Lytix function called is FLLinMixedModel. The mixed model extends the linear
model by allowing correlation and heterogeneous variances in the
covariance matrix of the residuals. Fit a linear mixed-effects model (LMM) to data,
FLLinMixedModel estimates the coefficients and covariance matrix of the mixed model
via the expectations.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

\examples{
## One Random Effect.
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer.FLTable(yVal ~ FixVal + (1 | RanVal), data = fltbl)
flpred <- predict(flmod)
}
\seealso{
\code{\link[lme4]{lmer}} for R reference implementation.
}
=======
% Please edit documentation in R/FLMixed.R, R/FLMixedmodel.R
\name{lmer.FLTable}
\alias{lmer.FLTable}
\alias{lmer.FLTable}
\title{fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)}
\usage{
\method{lmer}{FLTable}(formula, data, fetchID = TRUE, maxiter = 10, ...)

\method{lmer}{FLTable}(formula, data, fetchID = TRUE, maxiter = 10, ...)
}
\description{
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)

fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl)
flpred <- predict(flmod)
}
>>>>>>> 1dd8f5f2fe2269e9973d2b5027be052ad9d9c5fa
