% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMixedmodel.R
\name{lmer}
\alias{lmer}
\title{Fit Linear Mixed-Effects Models}
\usage{
\method{residuals}{FLMix}(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable.}
}
\value{
\code{lmer} returns an object of class \code{FLMix} or \code{FLMixUDT}
}
\description{
\code{lmer} performs mixed model  on FLTable objects.
}
\details{
The DB Lytix function called is FLLinMixedModel. The mixed model extends the linear
model by allowing correlation and heterogeneous variances in the
covariance matrix of the residuals. Fit a linear mixed-effects model (LMM) to data,
FLLinMixedModel estimates the coefficients and covariance matrix of the mixed model
via the expectations.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

\section{Constraints}{

DBLytix currently supports one Fixed and upto 2 Random Effect variables with intercept.
For non-categorical Fixed Effect, only one Random Effect is supported.
Specify if Fixed Effect is categorical using categoricalFixedEffect logical input to the function.
The intercept coefficient returned is the mean value for different Random Effects.
}

\examples{
## One Random Effect.
fltbl  <- FLTable(getTestTableName("tblMixedModel"), "ObsID")
colnames(fltbl) <- tolower(colnames(fltbl))
flmod <- lmer(yval ~ fixval + (1 | ranval), data = fltbl)
flpred <- predict(flmod)
## One Categorical Fixed Effect and Two Random Effects.
fltbl  <- FLTable(getTestTableName("tblMixedModelInt"), "ObsID")
colnames(fltbl) <- tolower(colnames(fltbl))
flmod <- lmer(yval ~ fixval + (1 | ranval1) + (1 | ranval2 ), fltbl, categoricalFixedEffect=TRUE)
flpred <- predict(flmod)
## One Categorical Fixed Effect and One Random Effect.
flmod <- lmer(yval ~ fixval + (1 | ranval1), data = fltbl, categoricalFixedEffect=TRUE)
flpred <- predict(flmod)
}
\seealso{
\code{\link[lme4]{lmer}} for R reference implementation.
}
