% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R
\name{mahalanobis}
\alias{mahalanobis}
\title{Mahalanobis Distance}
\usage{
mahalanobis(x, y, S, ...)
}
\arguments{
\item{x}{An FLVector of data with say, p columns.}

\item{y}{An FLVector of data with say, p columns.}

\item{S}{A covariance matrix of the distribution(p*p)}
}
\description{
Returns the squared Mahalanobis distance of an FLVector x and FLVector y with
respect to a covariance matrix sigma = S.
                         D^2 = (x - y)' Σ^-1 (x - y)
}
\examples{
x<-FLVector(c(0,0))
y<-FLVector(1:2)
ma<-cbind(1:6,1:3)
s<-var(ma)
S<-as.FLMatrix(s)
mahalanobis(x,y,S)
}
