% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar.test.R
\name{mcnemar.test}
\alias{mcnemar.test}
\title{McNemar's Chi-squared Test for Count Data.}
\usage{
mcnemar.test(x, y = NULL, correct = TRUE)
}
\arguments{
\item{x}{FLVector}

\item{y}{FLVector}

\item{correct}{a logical or character indicating whether to apply 
continuity correction when computing the test statistic.
If TRUE,Available methods are EDWARDS(default),YATES.}
}
\value{
A list with class "htest"
}
\description{
Performs McNemar's chi-squared test for 
symmetry of rows and columns in a two-dimensional contingency table.
}
\examples{
flTable <- FLTable("tblMcNemarMulti", "obsid",whereconditions="datasetID=1")
x <- flTable[,"predicted"]
y <- flTable[,"observed"]
ResulthtestObject <- mcnemar.test(x,y)
ResulthtestObject <- mcnemar.test(x,y,FALSE)
ResulthtestObject <- mcnemar.test(x,y,"YATES")
}
