% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLDA.R
\name{mda}
\alias{mda}
\title{library(mda)}
\usage{
mda(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable or FLTableMD.}

\item{subclasses}{Number of subclasses.}

\item{iter}{Maximum number of iterations for expectation maximization.}

\item{init}{Initialization method for each obs' latent variable Prob(x is in
the subclass of class) 1 = Assign weight of 1 to a
random subclass of its class; 0 otherwise.}

\item{hypothesis}{Number of hypotheses to run simultaneously}
}
\value{
\code{mda} returns an object of class \code{FLLDA}
}
\description{
library(mda)
}
\examples{
deeptbl <- FLTable("tblMDA","ObsID", "VarID", "Num_Val")
flmod <- mda(a~., data = deeptbl)
predict(flmod); flmod$N
cof <-coefficients(flmod)
FLMDA performs mixed discriminant analysis. For the training data, MDA divides each
class into a number of artificial subclasses. It calibrates the mixture of Gaussians
and the mixing probability by maximizing the log-likelihood with expectation maximization.
couldnt implement dollar operator: means, precent.explained, values, plot((lack of data of discriminant space).
}
\seealso{
\code{\link[stats]{mda}} for R reference implementation.
}

