% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMixedModelUdt.R
\name{mixUDT.FLTable}
\alias{mixUDT.FLTable}
\title{\code{mixudt} performs mixed model  on FLTable objects.
The DB Lytix function called is FLMixedModelUdt or FLMixedModelIntUdt.
The mixed model extends the linear model by allowing correlation and
heterogeneous variances in the covariance matrix of the residuals.
Fita a linear mixed-effects model (LMM) to data,
FLMixedModelUdt and FLMixedModelIntUdt estimates the coefficients and covariance
matrix of the mixed model via expectation maximization method.}
\usage{
\method{residuals}{FLMixUDT}(formula, data, fetchID = TRUE, ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable.}
}
\value{
\code{lmer} returns an object of class \code{FLMixUDT}
}
\description{
\code{mixudt} performs mixed model  on FLTable objects.
The DB Lytix function called is FLMixedModelUdt or FLMixedModelIntUdt.
The mixed model extends the linear model by allowing correlation and
heterogeneous variances in the covariance matrix of the residuals.
Fita a linear mixed-effects model (LMM) to data,
FLMixedModelUdt and FLMixedModelIntUdt estimates the coefficients and covariance
matrix of the mixed model via expectation maximization method.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

\examples{
One Random Effect.
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
2 Random Effects.
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)
}
\seealso{
\code{\link[lme4]{lmer}} for R reference implementation.
}
