% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMixedModelUdt.R
\name{mixUDT}
\alias{mixUDT}
\title{fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)}
\usage{
mixUDT(formula, data = list(), ...)
}
\description{
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- mixUDT(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)
}
