% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPLSRegr.R
\name{mvr}
\alias{mvr}
\title{PLS Regression.}
\usage{
mvr(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable or FLTableMD.}

\item{ncomp}{Number of components for performing pls & opls}
}
\value{
\code{mvr} returns an object of class \code{FLPLSRegr}
}
\description{
\code{mvr} performs linear regression on FLTable objects.
}
\details{
The DB Lytix function called is FLPLSRegr. Performs PLS Regression and 
stores the results in predefined tables.
}
\examples{
deeptbl  <- FLTable("tblPLSDeep2y", "ObsID", "VarID", "Num_Val")
flmod<- mvr(a~., data =deeptbl, ncomp = 5 )
predict(flmod); residuals(flmod);plot(flmod)
cof <-coefficients(flmod)
Functions to perform partial least squares regression (PLSR),
canonical powered partial least squares (CPPLS) or principal
component regression (PCR), with a formula interface.
Cross-validation can be used.  Prediction, model extraction, plot,
print and summary methods exist.
}
\seealso{
\code{\link[stats]{mvr}} for R reference implementation.
}

