% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{na_mean}
\alias{na_mean}
\title{Compute mean and handle all NA}
\usage{
na_mean(x, na.rm = TRUE, ...)
}
\value{
Mean of non-missing values or,  if all values were missing, NA.
}
\description{
When computing the mean of a vector and removing missing values with na.rm =
TRUE, this results in NaN if all values were NA. This function handles such
occassions by returning NA instead.
}

