% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayes.R
\name{naiveBayes.FLTable}
\alias{naiveBayes.FLTable}
\title{fltbl <- FLTable("tblNBData", "ObsID", "VarID", "NUM_VAL")
flmod <-naiveBayes(object = fltbl, formula = ~.)
rtbl <- iris
rtbl$Species <- sample(x = 2, size = length(rtbl$Species), replace = TRUE)-1
rtbl$Species <- as.numeric(rtbl$Species)
fliris <- as.FL(rtbl)
flmod <-naiveBayes.FLTable(data = fliris, formula = Species~., laplace = 1)
rmod <-naiveBayes(data = rtbl, formula = Species~.)
rtbl <- as.data.frame(Titanic)
rtbl <- rtbl[,-5]
rtbl <- as.data.frame(lapply(rtbl, function(i)as.numeric(i)-1))
colnames(rtbl)[[1]] <- "vcl"
rmod <-  naiveBayes(Survived~.,rtbl)
fltbl <- as.FL(rtbl)
flmod <- naiveBayes.FLTable(formula = Survived~., data = fltbl)}
\usage{
\method{naiveBayes}{FLTable}(formula, data, laplace = 0, ...)
}
\description{
fltbl <- FLTable("tblNBData", "ObsID", "VarID", "NUM_VAL")
flmod <-naiveBayes(object = fltbl, formula = ~.)
rtbl <- iris
rtbl$Species <- sample(x = 2, size = length(rtbl$Species), replace = TRUE)-1
rtbl$Species <- as.numeric(rtbl$Species)
fliris <- as.FL(rtbl)
flmod <-naiveBayes.FLTable(data = fliris, formula = Species~., laplace = 1)
rmod <-naiveBayes(data = rtbl, formula = Species~.)
rtbl <- as.data.frame(Titanic)
rtbl <- rtbl[,-5]
rtbl <- as.data.frame(lapply(rtbl, function(i)as.numeric(i)-1))
colnames(rtbl)[[1]] <- "vcl"
rmod <-  naiveBayes(Survived~.,rtbl)
fltbl <- as.FL(rtbl)
flmod <- naiveBayes.FLTable(formula = Survived~., data = fltbl)
}
