% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{names<-.FLSimpleVector}
\alias{names<-.FLSimpleVector}
\title{Assign names to FLSimpleVectors}
\usage{
\method{names}{FLSimpleVector}(x) <- value
}
\arguments{
\item{x}{FLSimpleVector}

\item{value}{a vector of length same as
FLSimpleVector}
}
\value{
named FLSimpleVector
}
\description{
Assign names to FLSimpleVectors
}
\examples{
flv <- FLSerial(1,5)
names(flv) <- letters[1:5]
print(flv)
}
