% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVector.R
\name{names<-.FLVector}
\alias{names<-.FLVector}
\title{Assign names to FLVectors}
\usage{
\method{names}{FLVector}(x) <- value
}
\arguments{
\item{x}{FLVector}

\item{value}{a vector of length same as
FLVector}
}
\value{
named FLVector
}
\description{
Assign names to FLVectors
}

