% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{names.FLSimpleVector}
\alias{names.FLSimpleVector}
\title{Get names of a FLVector}
\usage{
\method{names}{FLSimpleVector}(x)
}
\arguments{
\item{x}{FLVector}
}
\value{
character vector of names
of FLVector if exists. Else NULL
}
\description{
Get names of a FLVector
}

