% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTable.R
\name{names.FLTable}
\alias{names.FLTable}
\title{Gives the column names of FLTable object}
\usage{
\method{names}{FLTable}(object)
}
\arguments{
\item{object}{}
}
\description{
Gives the column names of FLTable object
}

