% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nest_duplicated}
\alias{nest_duplicated}
\title{Nest duplicated columns}
\usage{
nest_duplicated(df, suffixes = "\\\\.[xy]")
}
\arguments{
\item{df}{Data frame}

\item{suffixes}{Character string to match suffixes. E.g., the default
"\\.[xy]" finds any columns ending with .x or .y}
}
\value{
\code{\link[tibble]{tibble}}
}
\description{
After multiple data frames are joined, there can often a single variable that
is replicated and suffixed with .x, .y, .x.x, and so on. This function
converts any such variables into a single variable with all original values
combined into a list.
}
