% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPLSRegr.R
\name{opls}
\alias{opls}
\title{OPLS Regression.}
\usage{
opls(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable or FLTableMD.}

\item{ncomp}{Number of components for performing pls & opls.}

\item{northo}{Number of orthogonal vectors.}
}
\value{
\code{opls} returns an object of class \code{FLPLSRegr}
}
\description{
\code{mvr} performs linear regression on FLTable objects.
}
\details{
The DB Lytix function called is FLOPLSRegr. Performs OPLS Regression and 
stores the results in predefined tables.
}
\examples{
deeptbl  <- FLTable("tblPLSDeep2y", "ObsID", "VarID", "Num_Val")
flmod<- opls(a~., data =deeptbl, ncomp = 5,northo = 5 )
cof <- coefficients(flmod)
pred <- predict(flmod);res <- residuals(flmod) 
}
\seealso{
\code{\link[stats]{opls}} for R reference implementation.
}

