% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMedoids.R
\name{pam}
\alias{pam}
\title{K-Medoids Clustering.}
\usage{
pam(x, k, ...)
}
\arguments{
\item{x}{an object of class FLTable, can be wide or deep table}

\item{k}{the number of clusters}

\item{diss}{logical if \code{x} is dissimilarity matrix}

\item{metric}{only "euclidean" distance supported currently}

\item{medoids}{initial medoids. Not used currently.}

\item{Stand}{logical indicating if standardization
should be done before calculating diss matrix.}

\item{cluster.only}{logical if only clustering vector is 
needed as output}

\item{do.swap}{logical indicating if swap phase is needed.
currently always TRUE}

\item{keep.diss}{logicals indicating if the 
dissimilarities and/or input data x should be kept in the result}

\item{keep.data}{logicals indicating if the 
dissimilarities and/or input data x should be kept in the result}

\item{pomance}{logical or integer in 0:2 specifying algorithmic short cuts.
currently always FALSE}

\item{trace.lev}{integer specifying a trace level for 
printing diagnostics during the build and swap phase of the algorithm.
currently always 0}

\item{iter.max}{the maximum number of iterations allowed}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}

\item{distTable}{name of the in-database table having dissimilarity
matrix or distance table}
}
\value{
\code{pam} gives a list which replicates equivalent R output
from \code{pam} in cluster package. The mapping table can be viewed
using \code{object$mapping} if input is wide table.
}
\description{
\code{pam} performs k-medoids clustering on FLTable objects.
}
\details{
The DB Lytix function called is FLKMedoids. K-Medoids clusters the training data. 
The algorithm used is PAM (Partitioning Around Medoids).
}
\section{Constraints}{

Plotting time increases with size of data and even fail for large
datasets because data is fetched to R session for plotting.
If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
kmedoidsobject <- pam(widetable,3)
print(kmedoidsobject)
plot(kmedoidsobject)
One can specify ClassSpec and transform categorical variables 
before clustering. This increases the number of variables in the plot
because categorical variable is split into binary numerical variables.
The clusters may not be well-defined as is observed in the case below:-
widetable  <- FLTable( "FL_DEMO", "iris", "rownames")
pamobjectnew <- pam(widetable,3,classSpec=list("Species(setosa)"))
plot(pamobjectnew)
}
\seealso{
\code{\link[cluster]{pam}} for R function reference implementation.
}

