% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrixArithmetic.R, R/M_Addition.R
\name{+}
\alias{+}
\title{Addition of in-database objects.}
\usage{
"+"(x, y)

"+"(x, y)
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{x}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{+} returns an in-database object if there is atleast one in-database object 
as input.Otherwise, the default behavior of R is preserved

\code{+} returns an in-database object if there is atleast one in-database object 
as input.Otherwise, the default behavior of R is preserved
}
\description{
\code{+} does the addition of in-database objects.

\code{+} does the addition of in-database objects.
}
\details{
The addition of in-database objects mimics the normal addition of R data types.
All combinations of operands are possible just like in R and the result is an in-database object.

The addition of in-database objects mimics the normal addition of R data types.
All combinations of operands are possible just like in R and the result is an in-database object.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
Rvector <- 1:5
ResultFLmatrix <- flmatrix + Rvector
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
Rvector <- 1:5
ResultFLmatrix <- flmatrix + Rvector
}

