% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPP.R
\name{PP.test.FLVector}
\alias{PP.test.FLVector}
\title{test case in test_FLdifference.}
\usage{
\method{PP.test}{FLVector}(x, data, trend = 1, lag = 1, ...)
}
\arguments{
\item{data}{An object of class FLVector.}

\item{trend}{Whether to include trend order or not.}

\item{Lag}{Whether to use long or short order for lag of auto-correlation functions.}
}
\value{
\code{PP.test} returns a list of class htest
}
\description{
\code{PP.test} performs Phillips Perron for Stationarity on FLVector objects.
}
\details{
The DB Lytix function called is FLPP. Stored procedure to perform Phillips Perron
for Stationarity.Stationarity : A weak stationarity condition requires the
first moment and the auto-covariance function of the time series to be
finite and time invariant stores the results in data frame.
}
\examples{
vdf <- sqlQuery(connection, "SELECT Num_Val FROM tblsensex")
flv <- as.FL(vdf$Num_Val)
rv <- vdf$Num_Val
flmod <- PP.test(flv)
}
\seealso{
\code{\link[stats]{PP.test}} for R reference implementation.
}
