% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPP.R
\name{PP.test}
\alias{PP.test}
\title{vdf <- sqlQuery(connection, "SELECT Num_Val FROM tblsensex")
flv <- as.FL(vdf$Num_Val)
rv <- vdf$Num_Val
flmod <- PP.test(flv)}
\usage{
PP.test(object, ...)
}
\description{
vdf <- sqlQuery(connection, "SELECT Num_Val FROM tblsensex")
flv <- as.FL(vdf$Num_Val)
rv <- vdf$Num_Val
flmod <- PP.test(flv)
}
