% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPCA.R
\name{prcomp}
\alias{prcomp}
\title{\code{prcomp} Performs principal component analysis on a FLTable.
The function internally computes either the correlation matrix or the
covariance matrix and performs eigenvalue decomposition.
The DB Lytix function which is called id FLPCA.}
\usage{
prcomp(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model variables for which PCA is to be computed.}

\item{data}{An object of class FLTable.}

\item{matrixtype}{Indicates whether a correlation matrix or a covariance matrix
should be used for Eigen value decomposition.Allowed values are "CORREL" and "COVAR".}
}
\value{
\code{prcomp} returns an object of class \code{FLPCA}
}
\description{
\code{prcomp} Performs principal component analysis on a FLTable.
The function internally computes either the correlation matrix or the
covariance matrix and performs eigenvalue decomposition.
The DB Lytix function which is called id FLPCA.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed.}

\item{\code{deeptbl}}{Input data object in deep format.}

\item{\code{otbl}}{output table name.}
}}

\examples{
fltbl <- FLTable(getTestTableName("tblLogRegrMulti"),"OBSID", "VARID", "NUM_VAL")
flmod <- prcomp(data = fltbl, matrixtype = "COVAR",where = "")
rtbl <- iris
rtbl$Species <- as.numeric(rtbl$Species)
fliris <- as.FLTable(rtbl,tableName = getOption("TestTempTableName"),temporary=F, drop = TRUE)
flirispca <- prcomp(Species~., data = fliris)
}
\seealso{
\code{\link[stats]{prcomp}} for R reference implementation.
}
