% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{print.FLLinRegr}
\alias{print.FLLinRegr}
\title{Print FLLinRegr output Info}
\usage{
\method{print}{FLLinRegr}(object)
}
\arguments{
\item{object}{prints results of FLLinRegr on FL objects}

\item{object}{returns coefficient vector of the object}

\item{object}{the residuals, that is response minus fitted values.}

\item{object}{returns the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.}

\item{object}{returns the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.}

\item{object}{plots the results of FLLinRegr on FL objects.}
}
\description{
Print FLLinRegr Object
}
\details{
Printing of output from Linear Regression
}
