% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQRDecomp.R
\name{qr}
\alias{qr}
\title{QR Decomposition.}
\usage{
qr(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}
}
\value{
\code{qr} returns a list of five components:
\item{qr}{a FLMatrix with the same dimensions as \code{object}. The upper triangle contains the R of the decomposition 
and the lower triangle contains information on the Q of the decomposition (stored in compact form)}
\item{qraux}{a FLVector of length ncol(\code{object}) which contains additional information on Q.}
\item{rank}{the FLVector giving rank of \code{object}}
\item{QMatrix}{the resulting Q Matrix stored in-database as FLMatrix}
\item{RMatrix}{the resulting R Matrix stored in-database as FLMatrix}
}
\description{
The QR decomposition involves factorizing a matrix into QMatrix and RMatrix.
}
\details{
\code{qr} replicates the equivalent qr() generic function.\cr
}
\section{Constraints}{

Input can only be with maximum dimension limitations of (700 x 700).
}
\examples{
connection<-odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- qr(flmatrix)
resultList$qr
resultList$qraux
resultList$rank
resultList$pivot
}

