% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{randomForest}
\alias{randomForest}
\title{Classification and Regression with Random Forest}
\usage{
randomForest(data, formula, ...)
}
\arguments{
\item{data}{FLTable}

\item{formula}{formula specifying the independent and dependent variable columns}

\item{ntree}{Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times.}

\item{mtry}{Number of variables randomly sampled as candidates at each split}

\item{nodesize}{Minimum size of terminal nodes.}

\item{maxdepth}{The maximum depth to which the tree can go.
cp: Complexity parameter}
}
\value{
An object of class "FLRandomForest" containing the forest structure details.
}
\description{
Random forest is a technique for reducing
the variance of an estimated prediction function.
It takes multiple random samples(with replacement) from the training
data set, uses each of these samples to construct a separate model and separate predictions for test set, and
then averages them.
}
\examples{
flt<-FLTable("tblDecisionTreeMulti","ObsID","VarID","Num_Val")
flobj<-randomForest(data = flt, formula = -1~., ntree=5)
print(flobj)
pred <- predict(flobj, newdata= flt[1:100,])
pred
plot(flobj)
}
\seealso{
\code{\link[randomForest]{randomForest}} for corresponding R function reference.
}
