% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLRankMatrix.R
\name{rankMatrix.FLMatrix}
\alias{rankMatrix.FLMatrix}
\title{Matrix Rank.}
\usage{
\method{rankMatrix}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{rankMatrix} returns FLVector object of size 1 which replicates the equivalent R output.
}
\description{
\code{rankMatrix} computes the rank of FLMatrix objects.
}
\details{
\code{rankMatrix} computes the rank of input FLMatrix object, stores the result
in-database and returns FLVector object
}
\section{Constraints}{

Input can have maximum dimension limitations of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
resultFLVector <- rankMatrix(flmatrix)
}

