% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRankMatrix.R
\name{rankMatrix}
\alias{rankMatrix}
\title{Matrix Rank.}
\usage{
rankMatrix(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}
}
\value{
\code{rankMatrix} returns R vector object of size 1 which replicates the equivalent R output.
}
\description{
\code{rankMatrix} computes the rank of FLMatrix objects.
}
\section{Constraints}{

Input can have maximum dimension limitations of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- rankMatrix(flmatrix)
}

