% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRbind.R
\name{rbind}
\alias{rbind}
\title{Combine objects by rows.}
\usage{
rbind(x, ...)
}
\arguments{
\item{x...}{can be a sequence of vector, FLVector, matrix, FLMatrix or data frames}
}
\value{
\code{rbind} returns a FLMatrix object which is the row-wise combination of input arguments.
}
\description{
\code{rbind} combines input objects by rows and forms a FLMatrix.
}
\details{
\code{rbind} takes a sequence of vector, FLVector, matrix, FLMatrix or data frames arguments,
combines them by rows and makes a FLMatrix.
}
\section{Constraints}{

Input matrices, FLMatrices and data frames should have same number of columns.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- rbind(flmatrix,1:5,flmatrix)
}

