% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_all}
\alias{read_all}
\title{Read all users logs from a data directory}
\usage{
read_all(data_dir, clean = TRUE, calculate = TRUE)
}
\arguments{
\item{data_dir}{Directory containing all group data  directories. Must be
organised in particular way.}

\item{clean}{Should user list be cleaned as well?}

\item{calculate}{Should derived variables be calculated?}
}
\value{
List of user log-file data lists
}
\description{
Read all users logs from a data directory
}

