% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_logs}
\alias{read_all_streams}
\alias{read_events}
\alias{read_logs}
\alias{read_session}
\alias{read_stream}
\title{Read user's log files}
\usage{
read_logs(user_dir)

read_session(user_dir, file_name = "session.tsv", col_names = c("var",
  "info"))

read_events(user_dir, file_name = "events.tsv", col_names = c("time",
  "event", "detail"))

read_stream(user_dir, file_name, stream_dir = "streams/", is_numeric = TRUE,
  is_vec3 = FALSE)

read_all_streams(user_dir, pattern = "tsv$", stream_dir = "streams/",
  is_numeric = c("input_brake", "input_horizontal", "input_vertical"),
  is_vec3 = c("position", "rotation", "velocity"))
}
\arguments{
\item{user_dir}{Character string defining the user's log-file directory}

\item{file_name}{Character string of the file name. Must include extension such as .tsv}

\item{col_names}{Vector of column names to be used.}

\item{stream_dir}{Character string defining the directory in which stream log
files exist.}

\item{is_numeric}{Indicate whether a stream variable is numeric and,
therefore, should be convereted to numeric. Can be a boolean value
(TRUE/FALSE) or a character vector of variable names to convert if present.}

\item{is_vec3}{Indicate whether a stream variable is a Vector3 value in Unity
and, therefore, should be convereted to a vector of 3 values. Can be a
boolean value (TRUE/FALSE) or a character vector of variable names to
convert if present.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
read_logs with return a list with a "user" S3 class, with three
  tibbles (session, events and streams). All others will return a single
  \code{\link[tibble]{tibble}}.
}
\description{
These functions help to read a user's session, events, or stream log files.
read_logs() is a wrapper function that uses the default values of all other
functions to read all log files into a single list. If this fails, the log
files can be read separately using the other functions and by adjusting the
variables appropriately.
}
\details{
All files that can be read are assumed to be tab-separated values without
variable headers. Uniquely, they are assumed to:

\describe{
 \item{session}{Appear in top-level of user's log file directory; contain
 variable names and values}
 \item{events}{Appear in top-level of user's log file directory; contain a
 timestamp and an event name (tab-separated details about the event can
 follow in some cases).}
 \item{stream}{Appear in streams/ directory of user's log file directory;
 contain a timestamp and a value.}
}

Most functions are helper functions to read in a single file.
read_all_streams, however, reads all the stream files in a directory and
merges them into a single tibble. Also, read_logs will make use of all the
functions to read all log files into a list.
}

