% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_session}
\alias{read_session}
\title{Read participant's session file}
\usage{
read_session(user_dir, session_file = "session.tsv", col_names = c("var",
  "info"), ...)
}
\arguments{
\item{user_dir}{Character string defining the user's log-file directory}

\item{session_file}{Character string of the session file name}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
  of column names.

  If \code{TRUE}, the first row of the input will be used as the column
  names, and will not be included in the data frame. If \code{FALSE}, column
  names will be generated automatically: X1, X2, X3 etc.

  If \code{col_names} is a character vector, the values will be used as the
  names of the columns, and the first row of the input will be read into
  the first row of the output data frame.

  Missing (\code{NA}) column names will generate a warning, and be filled
  in with dummy names \code{X1}, \code{X2} etc. Duplicate column names
  will generate a warning and be made unique with a numeric prefix.}

\item{...}{Additional arguments to pass to \code{\link[readr]{read_tsv}}}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
Read participant's session file
}

