% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{regexpr}
\alias{regexpr}
\title{Pattern Matching}
\usage{
regexpr(startpos = 1, pattern, text, ignore.case = FALSE, perl = FALSE,
  fixed = FALSE, useBytes = FALSE)
}
\arguments{
\item{startpos}{integer(numeric) specifying starting position for each search}

\item{pattern}{string to search for}

\item{text}{FLVector of characters or R vector to be searched.
where matches are sought}

\item{ignore.case}{logical indicating case-sensitivity.
Currently always FALSE for FLVectors}

\item{perl}{logical. Should perl-compatible regexps be used?
Always FALSE for FLVectors}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.
For FLVectors, regualar expressions are not supported}

\item{useBytes}{If TRUE the matching is done byte-by-byte
rather than character-by-character. Always FALSE for FLVector}
}
\value{
FLVector with position of first match
or -1 for no match or R Vector
}
\description{
Match \code{pattern} in each element of \code{text}
}
\details{
The DB Lytix function called is FLInstr.This function returns the position of 
the first occurrence of one string within another starting from the
search position indicated.
}
\section{Constraints}{

row FLVectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- regexpr("A",flv)
}
\seealso{
\code{\link[base]{regexpr}} for R function reference implementation.
}

