% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLGarch.R
\name{regimeshift.FLVector}
\alias{regimeshift.FLVector}
\title{vdf <-  sqlQuery(connection, "SELECT Num_Val FROM tblRegimeShift WHERE Groupid = 1 AND Obsid <500")
flv <- as.FL(rv$Num_Val)
regimeshift(flv)}
\usage{
\method{regimeshift}{FLVector}(data, regimes = 2, ...)
}
\description{
vdf <-  sqlQuery(connection, "SELECT Num_Val FROM tblRegimeShift WHERE Groupid = 1 AND Obsid <500")
flv <- as.FL(rv$Num_Val)
regimeshift(flv)
}
