% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{remoteTable}
\alias{remoteTable}
\title{reference in-database object}
\usage{
remoteTable(object, table)
}
\arguments{
\item{object}{in-database object}

\item{table}{table name. Applicable only if object is the database name.}
}
\value{
character vector giving reference to in-database object
}
\description{
reference in-database object
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

