% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{restrictFLMatrix}
\alias{restrictFLMatrix}
\title{Appends where clauses for subsetting etc.}
\usage{
restrictFLMatrix(object, whereconditions = object@select@whereconditions,
  dimnames = object@Dimnames, conditionDims = c(FALSE, FALSE))
}
\arguments{
\item{object}{An FLMatrix object}

\item{whereconditions}{constraints to be added}

\item{dimnames}{new dimension names}

\item{conditionDims}{vector of 2 LOGICAL values, if first is TRUE,
a inCondition for the rownames is appended, if 2 for the columns respectively.}
}
\description{
Appends where clauses for subsetting etc.
}

