% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{restrictToObsids}
\alias{restrictToObsids}
\title{Select obsids in a larger table with fast EXISTS SQL sub-queries.}
\usage{
restrictToObsids(table, select, exclude = FALSE)
}
\arguments{
\item{table}{an FLTable or FLTableDeep}

\item{select}{a FLSimpleVector containgin all obsids}

\item{exclude}{a flag whether to constrict to the obsids in select or to exclude the ids in select}
}
\description{
Select obsids in a larger table with fast EXISTS SQL sub-queries.
}

