% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{rlm}
\alias{rlm}
\title{Robust Regression.}
\usage{
rlm(formula, data = list(), ...)
}
\description{
performs robust regression
Example for deeptbl:
library(MASS)
options(debugSQL =TRUE)
table  <- FLTable("tblRobustRegr", "ObsID","VarID", "Num_Val")
q <- rlm(a~., data = table)
predict(q)
residuals(q)
q$fitted.values
summary(q)
still to work on plot(q)
Example for widetable:
widetbl <- FLTable("tblautompg", "ObsID")
t <- rlm(Weight~ Acceleration , data = widetbl)
summary(t)
coefficients(t)
residuals(t)
}

