% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RLytix.R
\docType{package}
\name{RLytix}
\alias{RLytix}
\alias{RLytix-package}
\title{RLytix: A package makes it possible to use Fuzzy Logix functionality from R
by providing wrapper functions around the SQL interface of DB Lytix(TM).}
\description{
The RLytix package provides four categories of functions:
Data Mining, Matrix Operations, Hypothesis Testing and Fit Distribution
}
\section{Data Mining functions (17)}{

\describe{
\item{FLDecisionTree}{Fit a Decision Tree Model}
\item{FLFetch}{Fetch a fitted model from the database}
\item{FLKMeans}{Carry out KMeans clustering on a dataset}
\item{FLLDA}{Fit a Linear Discriminant Analysis Model}
\item{FLLinRegr}{Fit a Linear Regression Model}
\item{FLLinRegrStep}{Fit a Linear Regression Model with variable selection.
Variants provided are as follows:}
\describe{
		\item{BW}{Backward Elimination}
		\item{FB}{Fast Backward Elimination}
		\item{UFB}{Ultra Fast Backward Elimination}
		\item{SW}{Stepwise Logistic regression}
}
\item{FLLogRegr}{Fit a Logistic Regression Model}
\item{FLLogRegrStep}{Fit a Linear Regression Model with variable selection.
Variants provided are as follows:}
\describe{
		\item{BW}{Backward Elimination}
		\item{FB}{Fast Backward Elimination}
		\item{UFB}{Ultra Fast Backward Elimination}
		\item{SW}{Stepwise Logistic regression}
}
\item{FLMDA}{Fit a Multiple Discriminant Analysis Model}
\item{FLMatchIt}{Perform Nearest Neighbour Matching}
\item{FLNaiveBayes}{Fit a Naive Bayes Model}
\item{FLPredict}{Use a fitted model to make predictions on a dataset}
\item{FLVIF}{Calculate Variance Inflation Factor}
}
}

\section{Matrix Operations functions(5)}{

\describe{
\item{FLMatrix}{Attach a matrix stored in-database to a local object}
\item{FLCholeskyDecomp}{Cholesky Decomposition of a Matrix}
\item{FLEigen}{Calculates EigenValues and EigenVectors of a matrix}
\item{FLSVD}{Singular Value Decomposition of a Matrix}
\item{FLLUDecomp}{LU Decomposition of a Matrix}
}
}

\section{Hypothesis Testing functions(9)}{

\describe{
\item{FLAncova}{ANalyis Of COVAriance}
\item{FLAnova1Way}{1-way ANalyis Of VAriance}
\item{FLAnova2Way}{2-way ANalyis Of VAriance}
\item{FLtTest}{2-sample and 1-sample tTest}
\item{FLzTest}{2-sample, 1-sample, 2-sample proportion and 1-sample proportion zTest}
}
}

\section{Fit Distribution functions(14)}{

\describe{\item{FLFitDistr}{Fitting a distribution to a dataset. Its Various Variants are included}
\item{MLE Beta}{Maximum Likelihood Estimation of Beta Distribution}
\item{MLE Binomial}{Maximum Likelihood Estimation of Binomial Distribution}
\item{MLE Cauchy}{Maximum Likelihood Estimation of Cauchy Distribution}
\item{MLE ChiSq}{Maximum Likelihood Estimation of Chi-Squared Distribution}
\item{MLE Normal}{Maximum Likelihood Estimation of Normal Distribution}
\item{MLE Poisson}{Maximum Likelihood Estimation of Poisson Distribution}
\item{MLE Weibull}{Maximum Likelihood Estimation of Weibull Distribution}
\item{MDE Beta}{Minimum Distance Estimation of Beta Distribution}
\item{MDE Binomial}{Minimum Distance Estimation of Binomial Distribution}
\item{MDE Cauchy}{Minimum Distance Estimation of Cauchy Distribution}
\item{MDE ChiSq}{Minimum Distance Estimation of Chi-Squared Distribution}
\item{MDE Normal}{Minimum Distance Estimation of Normal Distribution}
\item{MDE Poisson}{Minimum Distance Estimation of Poisson Distribution}
\item{MDE Weibull}{Minimum Distance Estimation of Weibull Distribution}}
}

