% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLROC.R
\name{roc.FLTable}
\alias{roc.FLTable}
\title{TO-DO:- implementation in rocgeneric
roctbl <- FLTable("tblROCcurve", obs_id_colname = "ObsID")
rocmod <- roc.FLTable(ActualVal~ProbVal, data = roctbl)}
\usage{
\method{roc}{FLTable}(formula, data, ...)
}
\description{
TO-DO:- implementation in rocgeneric
roctbl <- FLTable("tblROCcurve", obs_id_colname = "ObsID")
rocmod <- roc.FLTable(ActualVal~ProbVal, data = roctbl)
}
