% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLROC.R
\name{roc}
\alias{roc}
\title{Build a ROC Curve}
\usage{
roc(object, limit = 1000, ...)
}
\arguments{
\item{response}{a FLvector of responses, typically encoded with 0
(controls) and 1 (cases). Only two classes can be used in a ROC curve.}

\item{predictor}{a FLvector of the same length than response, containing the 
predicted value of each observation.}
}
\value{
\code{roc} returns a FLROC class object which replicates equivalent R output
from \code{roc} in pROC package.
}
\description{
\code{roc} is a diagnostic test which plots the true positive rate
against the false positive rate for a given test set.
}
\details{
The wrapper overloads roc from pROC package and implicitly calls DB-Lytix roc function.
}
\examples{
tbl <- FLTable(getTestTableName("tblROCCurve"),
              "ObsID")
mod <- roc(tbl$ActualVal, tbl$ProbVal)
print(mod)
plot(mod)
auc(mod)
mod$levels

Example 2:
data(aSAH)
fltbl <- data.frame(res = aSAH$outcome, pred = aSAH$s100b)
fltbl$res <- as.numeric(fltbl$res)
fltbl$res <- fltbl$res - 1
fltbl <- fltbl[-55, ]
head(fltbl)
fltbl <- as.FLTable(fltbl)
flmod <- roc(fltbl$res, fltbl$pred)
plot(flmod)
summary(flmod)

roctbl <- FLTable("tblROCcurve", obs_id_colname = "ObsID")
rocmod <- roc.FLTable(ActualVal~ProbVal, data = roctbl)
}
\seealso{
\code{\link[pROC]{roc}} for corresponding R function reference.
}
