% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRowMeans.R
\name{rowMeans.FLMatrix}
\alias{rowMeans.FLMatrix}
\title{row means of a FLMatrix.}
\usage{
\method{rowMeans}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix.}
}
\value{
\code{rowMeans} returns a FLVector object representing the row-wise Means.
}
\description{
\code{rowMeans} computes the row-wise average of FLMatrix objects.
}
\details{
The wrapper overloads rowMeans and extends it to FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLVector <- rowMeans(flmatrix)
}

