% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolOps.R
\name{rowMeans}
\alias{rowMeans}
\title{row means of a FLMatrix.}
\usage{
rowMeans(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix.}

\item{...}{any additional arguments}
}
\value{
\code{rowMeans} returns a FLVector object representing the row-wise Means.
}
\description{
\code{rowMeans} computes the row-wise average of FLMatrix objects.
}
\examples{
flmatrix <- FLMatrix(getTestTableName("tblMatrixMulti"), 5,"MATRIX_ID",
                                       "ROW_ID","COL_ID","CELL_VAL",dims= c(5,5))
resultFLVector <- rowMeans(flmatrix)
}
\seealso{
\code{\link[Matrix]{rowMeans}} for corresponding R function reference
}
