% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLRowSums.R
\name{rowSums.FLMatrix}
\alias{rowSums.FLMatrix}
\title{row sums of a FLMatrix.}
\usage{
\method{rowSums}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix.}
}
\value{
\code{rowSums} returns a FLVector object representing the row-wise sums.
}
\description{
\code{rowSums} computes the row-wise sums of FLMatrix objects.
}
\details{
The wrapper overloads rowSums and extends it to FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLVector <- rowSums(flmatrix)
}

