% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcolOps.R
\name{rowSums}
\alias{rowSums}
\title{row sums of a FLMatrix.}
\usage{
rowSums(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix.}

\item{...}{any additional arguments}
}
\value{
\code{rowSums} returns a FLVector object representing the row-wise sums.
}
\description{
\code{rowSums} computes the row-wise sums of FLMatrix objects.
}
\examples{
flmatrix <- FLMatrix(getTestTableName("tblMatrixMulti"), 5,"MATRIX_ID",
                                       "ROW_ID","COL_ID","CELL_VAL",dims= c(5,5))
resultFLVector <- rowSums(flmatrix)
}
\seealso{
\code{\link[Matrix]{rowSums}} for corresponding R function reference
}
