% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{rpart}
\alias{rpart}
\title{Recursive partitioning and Regression Trees}
\usage{
rpart(formula, data = list(), ...)
}
\arguments{
\item{formula}{formula specifying the independent and dependent variable columns}

\item{data}{FLTable}

\item{control}{A list of options that control details of the rpart algorithm.
Minsplit: Minimum number of observations a node should have in order to be splitted.
Maxdepth: The maximum depth to which the tree can go.
cp: Complexity parameter}
}
\value{
An object of class "FLrpart" containing the tree structure details.
}
\description{
Fit a rpart model
}
\examples{
flt<-FLTable("tblDecisionTreeMulti","ObsID","VarID","Num_Val")
flobj<-rpart(data = flt, formula = -1~.)
}
