% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSchurDecomp.R
\name{Schur}
\alias{Schur}
\title{\code{Schur} Calculates the Schur decomposition of a square matrix (n x n) with real eigenvalues.}
\usage{
Schur(object, ...)
}
\arguments{
\item{Objectis}{of class FLMatrix.}
}
\value{
\code{Schur} returns a list of two components:
      \item{Q}{an Orthogonal FLMatrix}
      \item{T}{an Upper Triangle FLMatrix.}
}
\description{
\code{Schur} Calculates the Schur decomposition of a square matrix (n x n) with real eigenvalues.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed.}

\item{\code{data}}{Input data object.}

\item{\code{call}}{call of the function.}
}}

\examples{
flmatrix <- FLMatrix(getTestTableName("tblMatrixMulti"), 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- Schur(flmatrix)
}
\seealso{
\code{\link[Matrix]{Schur}} for R reference implementation.
}
