% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\name{showTable}
\alias{showTable}
\title{Recieves the result of a "show table ..." SQL query as a character string.}
\usage{
showTable(x)
}
\description{
Recieves the result of a "show table ..." SQL query as a character string.
}

