% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLSolve.R
\name{solve}
\alias{solve}
\title{Inverse of a Matrix.}
\usage{
solve(x, ...)
}
\description{
\code{solve} computes the inverse for FLMatrix objects.
}
\details{
The wrapper overloads solve and implicitly calls FLMatrixInvUdt.
}

