% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLSolveExcl.R
\name{solveExcl.FLMatrix}
\alias{solveExcl.FLMatrix}
\title{Inverse of a Matrix excluding a dimension.}
\usage{
\method{solveExcl}{FLMatrix}(object, ExclIdx)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{ExclIdx}{is a positive integer specifying row or column id to be excluded.}
}
\value{
\code{solveExcl} returns a FLMatrix object which is the inverse of input FLMatrix object
after excluding given dimension.
}
\description{
\code{solveExcl} computes the inverse for FLMatrix objects by excluding
the specified row and column from the matrix.
}
\details{
The wrapper overloads solveExcl and implicitly calls FLMatrixInvExclUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 5)
resultFLMatrix <- solveExcl(flmatrix,3)
}

