% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlSendUpdate.JDBCConnection}
\alias{sqlSendUpdate.JDBCConnection}
\title{Send a query to database}
\usage{
\method{sqlSendUpdate}{JDBCConnection}(channel, query)
}
\arguments{
\item{channel}{JDBC connection object}

\item{query}{SQLQuery to be sent}
}
\description{
No result is returned
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

