% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlSendUpdate}
\alias{sqlSendUpdate}
\title{Send a query to database}
\usage{
sqlSendUpdate(channel, query)
}
\arguments{
\item{channel}{JDBC connection object}

\item{query}{SQLQuery to be sent}
}
\description{
No result is returned
}

