% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlSendUpdate.RODBC}
\alias{sqlSendUpdate.RODBC}
\title{Send a query to database}
\usage{
\method{sqlSendUpdate}{RODBC}(connection, query, warn = FALSE, ...)
}
\arguments{
\item{query}{SQLQuery to be sent}

\item{channel}{ODBC connection object}
}
\description{
No result is returned
}
