% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlStoredProc}
\alias{sqlStoredProc}
\title{Send a query to database
Result is returned as data.frame}
\usage{
sqlStoredProc(connection, query, outputParameter, ...)
}
\arguments{
\item{query}{SQLQuery to be sent}

\item{channel}{ODBC/JDBC connection object}
}
\description{
Send a query to database
Result is returned as data.frame
}
