% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStore.R
\name{store}
\alias{store}
\title{stores an object in database.}
\usage{
store(object, pTableName = NULL, ...)
}
\arguments{
\item{object}{the object to store. Can be FLMatrix, FLVector, FLTable, character}

\item{pTableName}{If provided, the object is inserted into this table and 
this storage information is writen into fzzlAdapteRTablesInfo. This allows 
for selective storage of objects}

\item{...}{Additional arguments like pNote to be included in fzzlAdapteRTablesInfo}
}
\value{
in-database object after storing
}
\description{
stores an object in database.
}

