% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{stringdistmatrix}
\alias{stringdistmatrix}
\title{stringdistmatrix}
\usage{
stringdistmatrix(a, b, method = "dl", caseFlag = 0, p = 0, vlength = 3,
  asMatrix = TRUE, ...)
}
\arguments{
\item{a}{character or FLVector of characters}

\item{b}{character or FLVector of characters}

\item{method}{can be \code{c("lv","dl","hamming","jaccard","jw")}
where lv - levenshtein, dl - levenshtein.damerau
jw - jaro-winkler. Default is "lv"}

\item{p}{penality factor for jaro-winkler
if p==0 jaro distance is computed}

\item{vlength}{optional, length of strings to compare
used for hamming}

\item{caseFLag}{logical or 0/1 indicating
if comparision should be case sensitive}
}
\value{
FLMatrix of string distances
}
\description{
compute distance metrics between strings.
}
\details{
stringdistmatrix computes the string distance matrix with rows
according to a and columns according to b.
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflmatrix <- stringdistmatrix("xyz",flv)
resultflmatrix <- stringdistmatrix(c("xyz","abc"),flv,method="lv",caseFLag=1)
resultflmatrix <- stringdistmatrix("xyz",flv,method="hamming",vlength=4)
resultflmatrix <- stringdistmatrix(flv,flv,method="jw",p=1)
resultflmatrix <- stringdistmatrix(flv,c("xyz","abc"),method="jw")
}
\seealso{
\code{\link[stringdist]{stringdist}} for R function reference
implementation.
}

