% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{[.FLVector}
\alias{[.FLVector}
\title{Extract part of FLVector object.}
\usage{
\method{[}{FLVector}(object, pSet = 1:length(object))
}
\arguments{
\item{object}{is a FLVector object}

\item{pSet}{is a vector representing the indices of elements to extract}
}
\value{
\code{[]} returns FLVector object after extraction
which replicates the equivalent R extraction.
}
\description{
\code{[]} acts on FLVector objects and extracts parts of them.
}
\examples{
WideTable <- FLTable("tblAbaloneWide","ObsID")
flvector <- FLVector[,"Diameter"]
resultFLVector <- flvector[10:1]
}
