% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{survdiff}
\alias{survdiff}
\title{Test Survival Curve Differences}
\usage{
survdiff(formula, data, subset = TRUE, na.action = getOption("na.action"),
  rho = 0, ...)
}
\arguments{
\item{formula}{a formula expression as for other 
survival models, of the form Surv(time, status) ~ predictor.}

\item{data}{Wide FLTable or FLTableMD objects}

\item{subset}{Not currently used}

\item{na.action}{na values omitted always}

\item{rho}{a scalar parameter that controls the type of test.
For FL objects, supported values are 0(LogRank),1(Peto-Prentice),
LogRank,Peto-Prentice,Modified Peto-Prentice,Wilcoxon,Taron-Ware.}

\item{...}{The additional arguments used by FL function:
\code{conf.int} the level for a two-sided confidence interval on the survival curve(s). Default is 0.95.
\code{whereconditions} WhereConditions to subset data
\code{GroupBy} Column names defining the different groups in data, if any.}
}
\value{
A list with class \code{survdiff}.
A list of \code{survdiff} objects if the input is a FLTableMD object.
}
\description{
Tests if there is a difference between two or more 
survival curves using the G-rho family of tests, 
or for a single curve against a known alternative.
}
\section{Constraints}{

Only single predictor is supported in formula.
\code{var} is not available in the output.
rho values supported are mentioned above.
}

\examples{
fltMD <- FLTableMD("vwWHAS100","DataSetID","ObsID")
resultList <- survdiff(Surv(TIME_VAL,STATUS)~Gender,data=fltMD)
print(resultList[[1]])
}
