% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{survfit.formula}
\alias{survfit.formula}
\title{Compute a Survival Curve for Censored Data}
\usage{
\method{survfit}{formula}(formula, data, weights, subset, na.action, etype, id,
  istate, ...)
}
\arguments{
\item{formula}{a formula object, which must have a Surv object 
as the response on the left of the ~ operator and 1 on right side(Single survival curve)}

\item{data}{Wide FLTable or FLTableMD objects}

\item{weights}{Not currently used}

\item{subset}{Not currently used}

\item{na.action}{na values omitted always}

\item{etype}{Not currently used}

\item{id}{Not currently used}

\item{istate}{Not currently used}

\item{...}{The additional arguments used by FL function:
\code{conf.int} the level for a two-sided confidence interval on the survival curve(s). Default is 0.95.
\code{whereconditions} WhereConditions to subset data
\code{GroupBy} Column names defining the different groups in data, if any.}
}
\value{
A list with class \code{survfit}.
A list of \code{survfit} objects if the input is a FLTableMD object.
}
\description{
Computes an estimate of a survival curve for censored data
}
\section{Constraints}{

Only Single survival curve supoorted currently.(RHS of formula is 1)
Only conf.type='plain' supported.
}
\examples{
fltMD <- FLTableMD("vwWHAS100","DataSetID","ObsID")
resultList <- survfit(Surv(TIME_VAL,STATUS)~1,data=fltMD)
print(resultList[[1]])
summary(resultList[[1]])
plot(resultList[[1]])
}

