% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSVDecomp.R
\name{svd}
\alias{svd}
\title{Singular Value Decomposition of a Matrix.}
\usage{
svd(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}

\item{nu}{number of left singular vectors to be computed.This must between 0 and nrow(object).}

\item{nv}{number of right singular vectors to be computed.This must between 0 and ncol(object).}
}
\value{
\code{svd} returns a list of three components:
      \item{d}{a FLVector containing the singular values of x, of size min(n, p).}
      \item{u}{a FLMatrix whose columns contain the left singular vectors of x, present if nu > 0. Dimension c(n, nu).}
      \item{v}{a FLMatrix whose columns contain the right singular vectors of x, present if nv > 0. Dimension c(p, nv).}
}
\description{
\code{svd} computes the singular value decomposition for FLMatrix objects.
}
\section{Constraints}{

Input can only be with maximum dimension limitations of (550 x 550).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- svd(flmatrix)
resultList$d
resultList$u
resultList$v
}

