% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSVM.R
\name{svm}
\alias{svm}
\title{Support Vector machine with Linear Kernel}
\usage{
svm(formula, data = list(), ...)
}
\description{
library(e1071)
tbl  <- FLTable("tblSVMLinSepMultiDim", "OBSID", whereconditions= "OBSID>307")
flmod <- svm(DEP~., data = tbl, fetchID = TRUE, kernel = "linear")
predict(flmod)
}
\details{
polynomial Kernel
tbl <- FLTable("tblSVMDense", "OBSID", whereconditions = "OBSID>307")
flmod <- svm(DEP~., data = tbl, fetchID = TRUE, kernel = "polynomial")
}

