% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTranspose.R
\name{t.FLMatrix}
\alias{t.FLMatrix}
\title{Matrix Transpose.}
\usage{
\method{t}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{t} returns a FLMatrix object which is the transpose of input FLMatrix object
and replicates the equivalent R output.
}
\description{
\code{t} returns the transpose of FLMatrix objects.
}
\details{
The wrapper overloads t such that given a matrix of class FLMatrix, t returns the transpose of that object
}
\section{Constraints}{

Input can be a matrix of dimensions (m x n) where m > n, m < n or m = n.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
resultFLMatrix <- t(flmatrix)
}

