% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTranspose.R
\name{t}
\alias{t}
\title{Matrix Transpose.}
\usage{
t(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix}
}
\value{
\code{t} returns a FLMatrix object which is the transpose of input FLMatrix object
and replicates the equivalent R output.
}
\description{
\code{t} returns the transpose of FLMatrix objects.
}
\section{Constraints}{

Input can be a matrix of dimensions (m x n) where m > n, m < n or m = n.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- t(flmatrix)
}
\author{
Gregor Kappler <g.kappler@gmx.net>, phani srikar <phanisrikar93ume@gmail.com>
}

