% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test.R
\name{t.test}
\alias{t.test}
\title{Student's t-Test}
\usage{
t.test(x, y = NULL, alternative = c("two.sided"), mu = 0,
  paired = FALSE, var.equal = FALSE, conf.level = 0.95, tails = 2, ...)
}
\arguments{
\item{x}{FLVector}

\item{y}{FLVector}

\item{mu}{The value of hypothesized mean}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.
If TRUE then the pooled variance is used to estimate the
variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}
}
\value{
A list with class "htest" containing the statistic and p-values.
}
\description{
Performs one and two sample t-tests on vectors of data.
}
\examples{
flx<-as.FLVector(rnorm(100))
fly<-as.FLVector(rnorm(100))
t.test(flx)
t.test(flx,fly)
t.test(flx,fly,var.equal=F)
}
