% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{tableAndAlias}
\alias{tableAndAlias}
\title{reference in-database object}
\usage{
tableAndAlias(object)
}
\arguments{
\item{object}{in-database object}

\item{table}{table name. Applicable only if object is the database name.}
}
\value{
character vector giving reference to in-database object
}
\description{
reference in-database object
}

