% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{teammate_data}
\alias{teammate_data}
\title{Find a teammate's data}
\usage{
teammate_data(user, full_list)
}
\arguments{
\item{user}{A user data object}

\item{full_list}{The full user list to be passed on to clean drone data only.}
}
\value{
A user data object or NULL.
}
\description{
Given a particular user and a full list of all users' data, return the data
for the particular user's teammate.
}
