% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{test_equal_FLMatrix_RMatrix}
\alias{test_equal_FLMatrix_RMatrix}
\title{converts FLMatrix to r matrix and checks if
recursive identical subsetting results in identical
matrices.}
\usage{
test_equal_FLMatrix_RMatrix(b)
}
\arguments{
\item{b}{FLMatrix}
}
\description{
converts FLMatrix to r matrix and checks if
recursive identical subsetting results in identical
matrices.
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}
