% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{test_Matrix_Subsetting}
\alias{test_Matrix_Subsetting}
\title{tests matrix subsetting by names and by index recursively.}
\usage{
test_Matrix_Subsetting(a, b, desc = "")
}
\arguments{
\item{a}{}

\item{b}{}

\item{desc}{}
}
\description{
tests matrix subsetting by names and by index recursively.
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

