% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLtestLib.R
\name{test_Vector_Subsetting}
\alias{test_Vector_Subsetting}
\title{tests vector subsetting by names and by index recursively.}
\usage{
test_Vector_Subsetting(a, b, desc = "", index = TRUE)
}
\arguments{
\item{a}{}

\item{b}{}

\item{desc}{}
}
\description{
tests vector subsetting by names and by index recursively.
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

