% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/M_Multiplication.R
\name{*}
\alias{*}
\title{Element-Wise Multiplication of in-database objects.}
\usage{
"*"(x, y)
}
\arguments{
\item{x}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}

\item{y}{can be an in-database object like FLMatrix,FLSparseMatrix,FLVector or
a normal R object like matrix,sparseMatrix,vector}
}
\value{
\code{*} returns an in-database object if there is atleast one in-database object
as input.Otherwise, the default behavior of R is preserved
}
\description{
\code{*} does the Element-wise Multiplication of in-database objects.
}
\details{
The Element-wise Multiplication of in-database objects mimics the normal Element-wise Multiplication of R data types.
All combinations of operands are possible just like in R and the result is an in-database object.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
Rvector <- 1:5
ResultFLmatrix <- flmatrix * Rvector
}

