% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFTest.R
\name{var.test.FLVector}
\alias{var.test.FLVector}
\title{F Test.
Performs an F test to compare the variances of two samples from normal populations.}
\usage{
\method{var}{test.FLVector}(x, y, ratio = 1, alternative = "two.sided",
  conf.level = 0.95)
}
\arguments{
\item{x:}{FLVectors of data values}

\item{y:}{FLVectors of data values}

\item{alternative:}{a character string specifying the alternative hypothesis,
 must be one of '"two.sided"' (default), '"greater"' or
'"less"'.  You can specify just the initial letter.}
}
\value{
A list with class "htest".
}
\description{
F Test.
Performs an F test to compare the variances of two samples from normal populations.
}
\section{Constraints}{
 conf.level, ratio, formula, data aren't supported currently for FL objects.
}
\examples{
set.seed(450)
x <- as.FLVector(rnorm(50, mean = 0, sd = 2))
y <- as.FLVector(rnorm(30, mean = 1, sd = 1))
var.test(x, y, "two.sided")

The null hypothesis is that the ratio of the variances of the populations from which
x and y were drawn, or in the data to which the linear models x and y were fitted,
is equal to ratio.


}

