% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLFTest.R
\name{var.test}
\alias{var.test}
\title{F Test.}
\usage{
var.test(x, ...)
}
\arguments{
\item{x:}{FLVector of data values}

\item{y:}{FLVector of data values}

\item{ratio:}{the hypothesized ratio of the population variances of x and y.}

\item{alternative:}{a character string specifying the alternative hypothesis,
 must be one of '"two.sided"' (default), '"greater"' or
'"less"'.}
}
\value{
A list with class "htest".
}
\description{
Performs an F test to compare the variances of two samples from normal populations.
The null hypothesis is that the ratio of the variances of the populations from which
x and y were drawn, or in the data to which the linear models x and y were fitted,
is equal to ratio.
}
\section{Constraints}{
 conf.level, ratio, formula, data aren't supported currently for FL objects.
}

\examples{
set.seed(450)
x <- as.FLVector(rnorm(50, mean = 0, sd = 2))
y <- as.FLVector(rnorm(30, mean = 1, sd = 1))
var.test(x, y, alternative = "two.sided")
}
\seealso{
\code{\link[stats]{var.test}} for R reference implementation.
}
