% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{vec3_to_vec}
\alias{vec3_to_vec}
\title{Convert Unity Vector3 to (numeric) vector}
\usage{
vec3_to_vec(vec3, is_numeric = TRUE)
}
\arguments{
\item{vec3}{Character string or vector of character strings in Vector3 format: "(x, y, z)"}

\item{numeric}{Boolean to}
}
\value{
list of (numeric) vector(s)
}
\description{
Convert Unity Vector3 to (numeric) vector
}
