% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{methods}
\name{viewSelectMatrix,FLMatrix,character,character-method}
\alias{viewSelectMatrix,FLMatrix,character,character-method}
\title{todo: alias}
\usage{
\S4method{viewSelectMatrix}{FLMatrix,character,character}(object, localName,
  withName = "z")
}
\description{
todo: alias
}

