% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVIF.R
\name{vif.FLTable}
\alias{vif.FLTable}
\title{\code{vif} performs VIF on FLTable and FLTableMD objects.
The DB Lytix function which can be called are
c(FLVIF,FLVIFBW,FLVIFFB,FLVIFMultiDataSet,FLVIFBWMultiDataSet,FLVIFFBMultiDataSet).
Performs variance inflation factor analysis on data to identify
redundant variables in a dataset. The square root of the variance inflation factor
tells you how much larger the standard error is, compared with what it would be if
that variable were uncorrelated with the other predictor variables in the model.}
\usage{
\method{vif}{FLTable}(formula, data, fetchID = TRUE, method = "normal",
  threshold = c(2, 10), ...)
}
\arguments{
\item{formula}{A symbolic description of model variables for which vif is to be calculated.}

\item{data}{An object of class FLTable or FLTableMD.}

\item{method}{three methods are Available 'normal' for VIF , 'bw' for backward selections on independent variable and 'fw' for fast backward selections on independent variables}

\item{threshold}{One variable is dropped at a time till all of the VIF values in the
model are below the VIFThresholdNumber.}
}
\value{
\code{vif} returns an object of class \code{FLVIF}
}
\description{
\code{vif} performs VIF on FLTable and FLTableMD objects.
The DB Lytix function which can be called are
c(FLVIF,FLVIFBW,FLVIFFB,FLVIFMultiDataSet,FLVIFBWMultiDataSet,FLVIFFBMultiDataSet).
Performs variance inflation factor analysis on data to identify
redundant variables in a dataset. The square root of the variance inflation factor
tells you how much larger the standard error is, compared with what it would be if
that variable were uncorrelated with the other predictor variables in the model.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

\examples{
fltbl <- FLTable(table = "tbllinregr", obs_id_colname="OBSID", var_id_colnames="VARID", "NUM_VAL")
flmod <- vif.FLTable(data = fltbl)
flmod <- vif.FLTable(data = fltbl, method = "fb")
flmod <- vif.FLTable(data = fltbl, method = "bw")
For Multi-Dataset:
fltbl <- FLTableMD(table = "tblLogRegrMulti",group_id_colname="DATASETID",obs_id_colname="ObsID",var_id_colname="VarID",cell_val_colname="Num_Val")
flmod <- vif(data= fltbl, method = "bw", threshold = 5)
}
\seealso{
\code{\link[vif]{vif}} for R reference implementation.
}
