% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waypoints}
\alias{waypoints}
\title{Waypoints defining the simulation route}
\format{A data frame (tibble) with 41 rows and 3 variables:
\describe{
  \item{lap}{Lap number that driver is on at this waypoint}
  \item{x}{x-coordinate in the simulated world. Entire world spanned from 0
  to 1200, with intersections and corners being found at 200 to 1000 in
  increments of 200.}
  \item{y}{y-coordinate in the simulated world. Entire world spanned from 0
  to 1200, with intersections and corners being found at 200 to 1000 in
  increments of 200.}
}}
\usage{
waypoints
}
\description{
A dataset containing the waypoints that define the route taken by drivers
during the simulation. The data includes the lap number and x and y
coordinates of each waypoint. Each waypoint refers to a corner in the
simulation at which the driver must make a turn. This can be at t-shaped
intersections or an L-shaped section of the road that appears on the
outer-most sections.
}
\keyword{datasets}

