% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTable.R
\name{wideToDeep}
\alias{wideToDeep}
\title{Convert Wide Table to Deep Table in database.}
\usage{
wideToDeep(object, ...)
}
\arguments{
\item{object}{FLTable object}

\item{ExcludeCols}{character vector specifying columns to be excluded from conversion}

\item{ClassSpec}{list representing Class specification which identifies then value of the categorical variable to be used a reference}

\item{WhereClause}{character vector giving where conditions if any to reference the wide table}

\item{OutDeepTable}{name to be given to the output deep table, possibly with database}

\item{OutObsIDCol}{name to give to the primary key column name of the output deep table}

\item{OutVarIDCol}{name to give to the varibales name column of the output deep table}

\item{OutValueCol}{name to give to the value column of the output deep table}
}
\value{
\code{wideToDeep} returns a list containing components 
\code{table} which is the FLTable referencing the deep table and \code{AnalysisID} giving the AnalysisID of conversion
}
\description{
The DB Lytix function called is FLWideToDeep.  FLWideToDeep is similar to FLRegrDataPrep , except that 
FLWideToDeep does not require dependent variable or intercept, thus it creates deep tables without them.
}
\examples{
widetable  <- FLTable(getTestTableName("tblAbaloneWide"), "ObsID")
deeptable <- wideToDeep(widetable)
analysisID <- deeptable@wideToDeepAnalysisID

## columns may be excluded from deeptable using ExcludeCols
widetable  <- FLTable(getTestTableName("tblAbaloneWide"), "ObsID", , whereconditions= "obsID< 101")
deeptable <- wideToDeep(widetable, ExcludeCols= "Sex")
analysisID <- deeptable@wideToDeepAnalysisID
}
