% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox.test.R
\name{wilcox.test.FLVector}
\alias{wilcox.test.FLVector}
\title{Wilcox Test
An S3 class to represent wilcox signed Rank test
Performs one- and two-sample Wilcoxon tests on vectors of data; the latter is also known as \code{Mann-Whitney} test.}
\usage{
\method{wilcox.test}{FLVector}(x, y = NULL, paired = TRUE, mu = 0, ...)
}
\arguments{
\item{x}{FLvector of data values. Non-finite (e.g., infinite or missing) values will be omitted.}

\item{y}{an optional FLVector of data values: as with x non-finite values will be omitted.}

\item{paired}{a logical indicating whether you want a paired test.}
}
\value{
A list with class "htest".
}
\description{
Wilcox Test
An S3 class to represent wilcox signed Rank test
Performs one- and two-sample Wilcoxon tests on vectors of data; the latter is also known as \code{Mann-Whitney} test.
}
\section{Constraints}{
 conf.level, conf.level is not supported currently for FL objects.
}
\examples{
Wilcoxon Signed Rank test:
a <- as.FLVector(c(85,70,40,65,80,75,55,20))
b <- as.FLVector(c(75,50,50,40,20,65,40,25))
res <- wilcox.test(a, b, paired = TRUE)

Mann-Whitney test:
a <-(6, 8, 2, 4, 4, 5)
b <-  c(7, 10, 4, 3, 5, 6)
res <- wilcox.test(a, b, paired = FALSE)

If only x is given, or if both x and y are given and paired is TRUE, a Wilcoxon signed rank test of the null that the distribution of x (in the one sample case) or of x - y (in the paired two sample case) is symmetric about mu is performed.
Otherwise, if both x and y are given and paired is FALSE, a Wilcoxon rank sum test (equivalent to the Mann-Whitney test: see the Note) is carried out. In this case, the null hypothesis is that the distributions of x and y differ by a location shift of mu and the alternative is that they differ by some other location shift (and the one-sided alternative "greater" is that x is shifted to the right of y).
}

