% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLWOE.R
\name{woe}
\alias{woe}
\title{\code{woe} Performs woe on FLTable objects.It is an aggregate function which
Calculates the statistical Weight of Evidence (WOE). WOE is a measure of how much
an attribute in the data is related to statuses. This analysis can be used to
assess the overall predictive power of the variables being considered.}
\usage{
woe(events, nonevents, n = 2, data = list(), ...)
}
\arguments{
\item{events}{Events or element of good distribution.}

\item{nonevents}{Non-Events or element of bad distribution.}

\item{n}{number of ID's for which information is to be calculated.
fltbl <- FLTable(table = getTestTableName("tblinfoval"), obs_id_colname="BinID")
flmod <- woe(event = "Events", nonevents = "NonEvents", data = fltbl,n = 4)}

\item{data}{An object of class FLTable.}
}
\description{
\code{woe} Performs woe on FLTable objects.It is an aggregate function which
Calculates the statistical Weight of Evidence (WOE). WOE is a measure of how much
an attribute in the data is related to statuses. This analysis can be used to
assess the overall predictive power of the variables being considered.
}
