% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Converts in-database objects to a data frame in R}
\usage{
as.data.frame(x, ...)
}
\arguments{
\item{x}{can be FLTable, FLVector or FLMatrix}

\item{...}{any additional arguments}
}
\value{
R data.frame object
}
\description{
Caution: data is fetched into R session
}

