% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.FLTable}
\alias{as.FLTable}
\title{casting to FLTable}
\usage{
as.FLTable(object, connection, ...)
}
\arguments{
\item{object}{data frame which
needs to be casted to FLTable}

\item{connection}{ODBC/JDBC connection object}

\item{...}{additional arguments like size}
}
\value{
FLTable object after casting.
}
\description{
Converts input \code{obj} to FLVector object
}

