% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.FLVector}
\alias{as.FLVector}
\title{casting to FLVector}
\usage{
as.FLVector(object, connection, ...)
}
\arguments{
\item{connection}{ODBC connection object}

\item{obj}{matrix,vector,data frame,sparseMatrix,FLMatrix or FLSparseMatrix which
needs to be casted to FLVector}

\item{size}{number of elements in resulting FLVector.
size input is not applicable only in case of FLMatrix,FLSparseMatrix}
}
\value{
FLVector object after casting.
}
\description{
Converts input \code{obj} to FLVector object
}

