% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCastFunctions.R
\name{as.matrix.FLMatrix}
\alias{as.matrix.FLMatrix}
\title{Converts FLSparseMatrix object to a matrix in R}
\usage{
\method{as.matrix}{FLMatrix}(object)
}
\description{
Converts FLSparseMatrix object to a matrix in R
}

