% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.test.R
\name{chisq.test}
\alias{chisq.test}
\title{Chi-square and Pearson Chi-square Test}
\usage{
chisq.test(x, ...)
}
\arguments{
\item{x}{FLMatrix}

\item{pear}{The values is 1 if wanna perform pearson chi-square test. Otherwise, 
by default it's 0 and perform chi- square test.}
}
\value{
A list with class "htest" outputting the corresponding expected and P Values.
}
\description{
Produces the expected and the chi-square value for each cell in the contingency
table.
}
\examples{
mat <- rbind(c(762, 327, 468), c(484, 239, 477))
dimnames(mat) <- list(gender = c("F", "M"),
                   party = c("Democrat","Independent", "Republican"))
FLmat <- as.FL(mat)
chisq.test(FLmat)  
      ## by default pear=0 
      ## performs chi-square test
chisq.test(FLmat,pear=1) 
       ## Pearson chi- square test
}
