% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_laps}
\alias{clean_laps}
\title{Clean lap information}
\usage{
clean_laps(user, n_laps = 5, full_list = NULL)
}
\arguments{
\item{user}{A list of "user" list objects created by reading functions like
\code{\link{read_logs}}. Also handles single user list objects too.}

\item{n_laps}{The number of laps that the participant was supposed to do in
the simulation.}

\item{full_list}{The full user list to be passed on to clean drone data only.}
}
\value{
Original user obejct with lap information handled and cleaned.
}
\description{
For driver data, add a `lap` number column to the events and streams tibbles
and remove all data that precedes the start of the first lap, and comes after
the end of the last lap.
}

