% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLColMeans.R
\name{colMeans}
\alias{colMeans}
\title{column means of a FLMatrix.}
\usage{
colMeans(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix.}
}
\value{
\code{colMeans} returns a FLVector object representing the column-wise Means.
}
\description{
\code{colMeans} computes the column-wise average of FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- colMeans(flmatrix)
}

