% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLColMeans.R
\name{colMeans}
\alias{colMeans}
\title{column means of a FLMatrix.}
\usage{
colMeans(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix.}

\item{...}{any additional arguments.}
}
\value{
\code{colMeans} returns a FLVector object representing the column-wise Means.
}
\description{
\code{colMeans} computes the column-wise average of FLMatrix objects.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO", 
"tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- colMeans(flmatrix)
}

