% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTable.R
\name{deepToWide}
\alias{deepToWide}
\title{Convert Deep Table to Wide Table}
\usage{
deepToWide(object, whereconditions, mapTable, mapName, outWideTableDatabase,
  outWideTableName, Analysisid)
}
\arguments{
\item{object}{FLTable object to convert to wide table}

\item{whereconditions}{character vector specifying whereconditions if any to reference the input deep table}

\item{mapTable}{name of the in-database table containing mapping information to be used in conversion if any}

\item{mapName}{unique identifier for the mapping information in mapping table if any}

\item{outWideTableDatabase}{name of database where output wide table is to be stored}

\item{outWideTableName}{name to give to the output wide table}
}
\value{
\code{deepToWide} returns a list containing components 
\code{table} which is the FLTable referencing the wide table and \code{AnalysisID} giving the AnalysisID of conversion
}
\description{
Convert Deep Table to Wide Table
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
deeptable  <- FLTable( "FL_DEMO", "tblUSArrests", "ObsID","VarID","Num_Val")
resultList <- deepToWide(deeptable)
widetable <- resultList$table
analysisID <- resultList$AnalysisID
}

