% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLDet.R
\name{det.FLMatrix}
\alias{det.FLMatrix}
\title{Determinant of a Matrix.}
\usage{
\method{det}{FLMatrix}(object)
}
\arguments{
\item{object}{is a FLMatrix object}
}
\value{
\code{det} returns determinant stored in-database as FLVector
which replicates the equivalent R vector output.
}
\description{
\code{det} computes the determinant of FLMatrix objects.
}
\details{
The wrapper overloads det and implicitly calls FLMatrixDetUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
resultFLvector <- det(flmatrix)
}

