% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLDet.R
\name{det}
\alias{det}
\title{Determinant of a Matrix.}
\usage{
det(x, ...)
}
\description{
\code{det} computes the determinant for FLMatrix objects.
}
\details{
The wrapper overloads det and implicitly calls FLMatrixDetUdt.
}

