% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLDet.R
\name{det}
\alias{det}
\title{Determinant of a Matrix.}
\usage{
det(object, ...)
}
\arguments{
\item{object}{is a FLMatrix object}

\item{...}{any additional arguments}
}
\value{
\code{det} returns determinant as a R vector
which replicates the equivalent R vector output.
}
\description{
\code{det} computes the determinant of FLMatrix objects.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO", 
"tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL",connection)
resultFLvector <- det(flmatrix)
}

