% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLDiag.R
\name{diag}
\alias{diag}
\title{Matrix Diagonals}
\usage{
diag(x, ...)
}
\arguments{
\item{x}{is an object of class FLMatrix or FLVector}
}
\value{
If x is a FLMatrix then diag(x) returns the diagonal of x as FLVector object.
  If x is FLVector, the value is a diagonal square FLMatrix with diagonal elements as given in FLVector.
}
\description{
Extract or replace the diagonal of a matrix, or construct a diagonal matrix.
}
\details{
diag has three distinct usages:
x is a FLMatrix, when it extracts the diagonal.
x is a scalar (length-one FLVector) and the only argument, it returns a square identity matrix of size given by the scalar.
x is a FLVector, either of length at least 2. This returns a square matrix with the given diagonal entries.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO", 
"tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- diag(flmatrix)
DeepTable <- FLTable( "FL_DEMO", "tblUSArrests","ObsID")
flvectorDeep <- DeepTable[1:5,1]
resultFLMatrix <- diag(flvectorDeep)
}

