% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStatFunctions.R
\docType{methods}
\name{dist,FLMatrix-method}
\alias{dist,FLMatrix-method}
\title{Distance Matrix Computation}
\usage{
\S4method{dist}{FLMatrix}(x, method = "euclidean", diag = TRUE,
  upper = TRUE, p = 2)
}
\description{
Distance Matrix Computation
}
\section{Constraints}{

only manhattan and euclidean are supported currently.
}

\examples{
flmatrix <- FLMatrix("FL_DEM.tblMatrixMulti", 1,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
dist(flmatrix)
dist(flmatrix,diag=TRUE)
dist(flmatrix,upper=TRUE)
dist(flmatrix,diag=TRUE,upper=TRUE)
dist(flmatrix,"manhattan",TRUE,TRUE)
}
