% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLEigen.R
\name{eigen}
\alias{eigen}
\title{Spectral Decomposition of a Matrix.}
\usage{
eigen(x, symmetric = FALSE, only.values = FALSE, EISPACK = FALSE)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{eigen} returns a list of FLMatrix object containing the eigen vectors and
a FLVector object containing eigen values which replicates the equivalent R output.
}
\description{
\code{eigen} Computes eigenvalues and eigenvectors of FLMatrices.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
Complex Eigen values and vectors are not Supported.
}

\examples{
flmatrix <- FLMatrix("tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- eigen(flmatrix)
resultList$values
resultList$vectors
}
