% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ensureQuerySize}
\alias{ensureQuerySize}
\title{Ensure sqlQuery constructed meets limits
namely max size and max nestings}
\usage{
ensureQuerySize(pResult, pInput, pOperator, pStoreResult = FALSE, ...)
}
\arguments{
\item{pResult}{object whose constructSelect
needs to be within limits}

\item{pInput}{list of input objects}

\item{pOperator}{function which generated the pResult}

\item{pStoreResult}{Flag whether to store the pResult}
}
\value{
pResult after storing transparently inputs 
and recomputing the operation
}
\description{
Ensure sqlQuery constructed meets limits
namely max size and max nestings
}
\examples{
flm <- FLMatrix("tblmatrixMulti",3,"Matrix_id","ROW_ID","COL_ID","CELL_VAL")
flv <- as.FLVector(rnorm(25))
vexpression <- paste0(rep("flm+flv",13),collapse="")
cat("no.of Nested Queries: ",length(gregexpr("FROM",constructSelect(eval(parse(text=vexpression))))[[1]]))
vResult <- eval(parse(text=paste0(vexpression,"+",vexpression)))
cat("no.of Nested Queries in Result: ",length(gregexpr("FROM",constructSelect(vResult))[[1]]))
}

