% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLCleanStr}
\alias{FLCleanStr}
\title{Clean string}
\usage{
FLCleanStr(object)
}
\arguments{
\item{object}{FLVector of characters}
}
\value{
a clean FLVector
}
\description{
Remove non-printable characters from each
element of FLVector of strings.
}
\details{
The DB lytix function called is FLCleanStr.
The clean string function is a scaler that removes all non-printable
characters from a string(vector of characters) and outputs a formatted string.
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLCleanStr(flv)
}

