% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLConcatString}
\alias{FLConcatString}
\title{Concatenate elements of vector}
\usage{
FLConcatString(object, delimiter)
}
\arguments{
\item{object}{FLVector of characters}

\item{delimiter}{character}
}
\value{
FLVector of length 1 with result string
}
\description{
Concatenate elements of FLVector with
a delimiter as collapse value.
}
\details{
The DB Lytix function called is FLConcatString.
The concat string function is an aggregate that joins the values of a string 
column from a table(or a vector of characters) into an output string using a 
user supplied delimiter to separate the fields.
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLConcatString(flv,",")
}
\seealso{
\code{\link[base]{paste0}} for R function reference
implementation.
}

