% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flConnect}
\alias{flConnect}
\title{Creates either an ODBC connection or an JDBC connection and initializes
the FL session tables.}
\usage{
flConnect(host = NULL, database = NULL, user = NULL, passwd = NULL,
  jdbc.jarsDir = NULL, jdbc.options = "", odbcSource = NULL,
  driverClass = NULL, verbose = FALSE, ...)
}
\arguments{
\item{host}{}

\item{database}{}

\item{user}{}

\item{passwd}{}

\item{jdbc.jarsDir}{if provided, class paths for jars to be loaded.
If one element, a directory, all jar files in that directory will be loaded.
If multiple elements, specific jar files, these will be loaded.
Issues can occur unless you provide the fully qualified path.}

\item{jdbc.options}{}

\item{odbcSource}{}

\item{driverClass}{}

\item{verbose}{print debugging messages}

\item{...}{}

\item{#"TMODE}{}

\item{CHARSET}{}
}
\value{
either an ODBC connection or an JDBC connection
}
\description{
Creates either an ODBC connection or an JDBC connection and initializes
the FL session tables.
}
\examples{
connection <- flConnect("jdbc:teradata://10.200.4.116",
                          "FL_DEMO",
                         "psrikar","fzzlpass",
                        c("C:/Users/phani/Downloads/terajdbc4.jar",
                            "C:/Users/phani/Downloads/tdgssconfig.jar"),
                        driverClass="com.teradata.jdbc.TeraDriver")
connection <- flConnect("jdbc:hive2://192.168.2.179:10000",
                         "dblytix",
                         "hive","",
                        c("C:/Users/phani/Downloads/hive-jdbc-1.2.1-standalone.jar",
                            "C:/Users/phani/Downloads/hadoop-common-2.6.0.jar"),
                        driverClass="org.apache.hive.jdbc.HiveDriver")
connection <- flConnect("jdbc:ncluster://192.168.100.220:2406","fuzzylogix",
                         "db_superuser","db_superuser",
                        c("C:/Users/phani/Downloads/noarch-aster-jdbc-driver.jar",
                          "C:/Users/phani/Downloads/noarch-aster-adfs-client.jar"),
                        driverClass = "com.asterdata.ncluster.Driver")
connection <- flConnect(odbcSource="Gandalf",database="FL_DEMO",platform="TD")
connection <- flConnect(odbcSource="AsterVM",database="fuzzylogix",platform="TDAster")
}

