% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flConnect}
\alias{flConnect}
\title{Creates either an ODBC connection or an JDBC connection and initializes
the FL session tables.}
\usage{
flConnect(host = NULL, database = NULL, user = NULL, passwd = NULL,
  jdbc.jarsDir = NULL, jdbc.options = "", odbcSource = NULL, ...)
}
\arguments{
\item{host}{}

\item{database}{}

\item{user}{}

\item{passwd}{}

\item{jdbc.jarsDir}{if provided, class paths for tdgssconfig.jar and terajdbc4.jar in that dir are loaded.  Issues can occur unless you provide the fully qualified path.}

\item{odbcSource}{}

\item{...}{}
}
\value{
either an ODBC connection or an JDBC connection
}
\description{
Creates either an ODBC connection or an JDBC connection and initializes
the FL session tables.
}

