% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLDataMining-class}
\alias{FLDataMining-class}
\title{An S4 class to represent FLLinRegr}
\description{
An S4 class to represent FLLinRegr
}
\section{Slots}{

\describe{
\item{\code{offset}}{this can be used to specify a priori known component to be included in the 
linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases
An S4 class to represent objects returned by Data-Mining functions}

\item{\code{deeptable}}{A character vector containing a deeptable (either conversion from a 
widetable or input deeptable)}

\item{\code{AnalysisID}}{An output character ID from call to data-mining function}

\item{\code{wideToDeepAnalysisID}}{An output character ID from FLRegrDataPrep}

\item{\code{mapTable}}{A character string name for the mapping table in-database if input is wide-table, generated by FLRegrDataPrep}

\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

