% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLFetchTable}
\alias{FLFetchTable}
\title{Fetch Table}
\usage{
FLFetchTable(table)
}
\arguments{
\item{table}{an object of class FLTable.}
}
\value{
a data frame with each column of the table as one variable
}
\description{
Pulls a Table from a database into R
}
\details{
\code{FLFetchTable} retrieves a table (from the database) to which a FLTable object is
mapped.
}

