% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{FLIsNumeric}
\alias{FLIsNumeric}
\title{Check if Numeric}
\usage{
FLIsNumeric(object)
}
\arguments{
\item{object}{FLVector}
}
\value{
FLVector with 1 for TRUE and 0 for FALSE
}
\description{
Check if an element is numeric
}
\section{Constraints}{

row vectors are not supported currently.
}
\examples{
widetable  <- FLTable("FL_DEMO", "tblstringID", "stringID")
flv <- widetable[1:6,"string"]
resultflvector <- FLIsNumeric(flv)
}

