% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLJordanDecomp.R
\name{FLJordan}
\alias{FLJordan}
\title{Jordan Decomposition of a Matrix.}
\usage{
FLJordan(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{FLJordan} returns a list of two components:
      \item{J}{FLVector representing J vector obtained from Jordan decomposition}
      \item{P}{FLMatrix representing P matrix obtained from Jordan decomposition}
      \item{PInv}{FLMatrix representing PInv matrix obtained from Jordan decomposition}
}
\description{
\code{FLJordan} computes the Jordan decomposition for FLMatrix objects.
}
\section{Constraints}{

Input can only be square matrix with maximum dimension limitations of (700 x 700).
If the input matrix have repeated eigenvalues, then FLJordan function may give incorrect results.
}

\examples{
flmatrix <- FLMatrix("tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultList <- FLJordan(flmatrix)
resultList$J
resultList$P
resultList$PInv
}
\seealso{
Refer to \code{\link{FLMatrix}}, \code{\link{FLVector}} to know more about 
FLMatrix and FLVector.
}
