% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLLinRegr-class}
\alias{FLLinRegr-class}
\title{An S4 class to represent FLLinRegr}
\description{
An S4 class to represent FLLinRegr
}
\section{Slots}{

\describe{
\item{\code{formula}}{an object of class 'formula': Model Formula}

\item{\code{deeptable}}{A character vector containing 
the deeptable on conversion from a widetable}

\item{\code{AnalysisID}}{An output character ID from CALL FLLinRegr}

\item{\code{wideToDeepAnalysisID}}{An output character ID from FLRegrDataPrep}

\item{\code{mapTable}}{name of the mapping table}

\item{\code{scoreTable}}{name of the scoring table}

\item{\code{modelID}}{id of the model with best fit}

\item{\code{table}}{input FLTable object}

\item{\code{results}}{cache list of results computed}
}}

