% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\docType{class}
\name{FLLinRegr-class}
\alias{FLLinRegr-class}
\title{An S4 class to represent output from Linear Regression(lm) on in-database Objects}
\description{
An S4 class to represent output from Linear Regression(lm) on in-database Objects
}
\section{Slots}{

\describe{
\item{\code{offset}}{column name used as offset}

\item{\code{vfcalls}}{information about system tables
and in-database procedures called during execution}
}}

