% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLogRegr.R
\docType{class}
\name{FLLogRegr-class}
\alias{FLLogRegr-class}
\title{An S4 class to represent output from glm on in-database Objects}
\description{
An S4 class to represent output from glm on in-database Objects
}
\section{Slots}{

\describe{
\item{\code{offset}}{column name used as offset}

\item{\code{vfcalls}}{contains names of tables}
}}

