% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLLUDecomp.R
\docType{class}
\name{FLLU-class}
\alias{FLLU-class}
\title{An S4 class to represent LU Decomposition}
\description{
An S4 class to represent LU Decomposition
}
\section{Slots}{

\describe{
\item{\code{x}}{object of class FLVector}

\item{\code{perm}}{object of class FLVector}

\item{\code{Dim}}{object of class FLVector}

\item{\code{lower}}{object of class FLMatrix}

\item{\code{upper}}{object of class FLMatrix}

\item{\code{data_perm}}{object of class FLMatrix}
}}

