% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLMatrix-class}
\alias{FLMatrix-class}
\title{An S4 class to represent FLMatrix}
\description{
An S4 class to represent FLMatrix
}
\section{Slots}{

\describe{
\item{\code{dimnames}}{list dimension names of FLMatrix object}
}}

