% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLMatrix-class}
\alias{FLMatrix-class}
\title{An S4 class to represent FLMatrix}
\description{
An S4 class to represent FLMatrix
}
\section{Slots}{

\describe{
\item{\code{odbc_connection}}{ODBC connectivity for R}

\item{\code{db_name}}{character}

\item{\code{matrix_table}}{character}

\item{\code{matrix_id_value}}{numeric id value}

\item{\code{matrix_id_colname}}{character}

\item{\code{row_id_colname}}{character}

\item{\code{col_id_colname}}{character}

\item{\code{cell_val_colname}}{character}

\item{\code{nrow}}{numeric number of rows of FLMatrix object}

\item{\code{ncol}}{numeric number of cols of FLMatrix object}

\item{\code{dimnames}}{list dimension names of FLMatrix object}
}}

