% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{FLMatrix}
\alias{FLMatrix}
\title{Constructor function for FLMatrix.}
\usage{
FLMatrix(connection, databaseName = result_db_name, matrix_table,
  matrix_id_value = "", matrix_id_colname = "", row_id_colname = "ROW_ID",
  col_id_colname = "COL_ID", cell_val_colname = "CELL_VAL", nrow = 0,
  ncol = 0, dimnames = NULL, conditionDims = c(FALSE, FALSE),
  whereconditions = c(""))
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{matrix_table}{name of the matrix table}

\item{matrix_id_value}{identifier for the input matrix}

\item{matrix_id_colname}{matrix id value in \code{matrix_table}}

\item{row_id_colname}{column name in \code{matrix_table} where row numbers are stored}

\item{col_id_colname}{column name in \code{matrix_table} where column numbers are stored}

\item{cell_val_colname}{column name in \code{matrix_table} where matrix elements are stored}

\item{database}{name of the database}
}
\value{
\code{FLMatrix} returns an object of class FLMatrix mapped
to an in-database matrix.
}
\description{
\code{FLMatrix} constructs an object of class \code{FLMatrix}.
}
\details{
\code{FLMatrix} object is an in-database equivalent to matrix object.
This object is used as input for matrix operation functions.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 2)
}

