% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FLStartSession}
\alias{FLStartSession}
\title{Starts Session and Creates temp Tables for result storage}
\usage{
FLStartSession(connection, database = "FL_DEMO", persistent = "test",
  drop = TRUE,
  tableoptions = ", FALLBACK ,\\n\\t\\t\\t\\t     NO BEFORE JOURNAL,\\n\\t\\t\\t\\t     NO AFTER JOURNAL,\\n\\t\\t\\t\\t     CHECKSUM = DEFAULT,\\n\\t\\t\\t\\t     DEFAULT MERGEBLOCKRATIO")
}
\arguments{
\item{connection}{ODBC/JDBC connection object}

\item{database}{name of current database}

\item{persistent}{NULL if result tables are to be created as volatile tables}

\item{drop}{logical to specify to drop result tables if already existing}

\item{tableoptions}{options used to create result tables}
}
\description{
Strongly recommended to run before beginning a new R session
use options to specify the following:- ResultDatabaseFL, ResultVectorTableFL, ResultMatrixTableFL, MatrixNameMapTableFL, ResultSparseMatrixTableFL
}

