% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FLStartSession}
\alias{FLStartSession}
\title{Starts Session and Creates temp Tables for result storage}
\usage{
FLStartSession(connection, database = getOption("ResultDatabaseFL"),
  temporaryTables = TRUE, drop = FALSE, debug = FALSE,
  tableoptions = NULL, ...)
}
\arguments{
\item{connection}{ODBC/JDBC connection object}

\item{database}{name of current database}

\item{temporaryTables}{TRUE if result tables are to be created as volatile tables}

\item{drop}{logical to specify to drop result tables if already existing}

\item{tableoptions}{options used to create result tables}
}
\description{
Strongly recommended to run before beginning a new R session
use options to specify the following:- 
ResultDatabaseFL, ResultVectorTableFL, ResultMatrixTableFL, 
NameMapTableFL, ResultSparseMatrixTableFL
}

