% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSV.R
\name{FLSV}
\alias{FLSV}
\title{Singular Values of a FLMatrix.}
\usage{
FLSV(x, ...)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{FLSV} returns a FLVector object representing the singular values.
}
\description{
\code{FLSV} computes the singular values for FLMatrix objects.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (700 x 700).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- FLSV(flmatrix)
}
\author{
Phani Srikar <phanisrikar93ume@gmail.com>
}

