% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R
\docType{class}
\name{FLTable-class}
\alias{FLTable-class}
\title{An S4 class to represent FLTable, an in-database data.frame.}
\arguments{
\item{object}{retrieves the column names of FLTable object}
}
\description{
An S4 class to represent FLTable, an in-database data.frame.
}
\section{Slots}{

\describe{
\item{\code{select}}{FLTableQuery the select statement for the table.}

\item{\code{dimnames}}{the observation id and column names}

\item{\code{isDeep}}{logical (currently ignored)}
}}

