% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\docType{class}
\name{FLTable-class}
\alias{FLTable-class}
\title{An S4 class to represent FLTable}
\arguments{
\item{object}{retrieves the column names of FLTable object}
}
\description{
An S4 class to represent FLTable
}
\section{Slots}{

\describe{
\item{\code{db_name}}{character}

\item{\code{table_name}}{character}

\item{\code{obs_id_colname}}{character}

\item{\code{var_id_colnames}}{character}

\item{\code{cell_val_colname}}{character}

\item{\code{isDeep}}{logical}
}}

