% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLTable.R
\name{FLTable}
\alias{FLTable}
\title{Constructor function for FLTable}
\usage{
FLTable(connection, database, table)
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{database}{name of the database in \code{Teradata} which contains the table}

\item{table}{name of the  table}
}
\value{
\code{FLTable} returns an object of class FLTable mapped to a table
in Teradata.
}
\description{
\code{FLTable} constructs an object of class \code{FLTable}. This object is
used as the input for analytic functions
}
\examples{
\dontrun{

		connection <- odbcConnect("Gandalf")
		database   <- "FL_R_WRAP"
		table_name <- "tblAutoMpg"
		tbl        <- FLTable(connection, database, table_name)

}
}

