% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLTable.R
\name{FLTable}
\alias{FLTable}
\title{Constructor function for FLTable.}
\usage{
FLTable(connection, database, table, primary_key = character(0))
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{database}{name of the database}

\item{table}{name of the table}

\item{primary_key}{column name set as primary key}
}
\value{
\code{FLTable} returns an object of class FLTable mapped to a table
in Teradata.
}
\description{
\code{FLTable} constructs an object of class \code{FLTable}.
}
\details{
This object is used as input for data mining functions.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable(connection, "FL_TRAIN", "tblAbaloneWide", "ObsID")
names(widetable)
}

