% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTriDiag.R
\name{FLTriDiag}
\alias{FLTriDiag}
\title{TriDiagonal or Upper Hessenberg matrix of a FLMatrix.}
\usage{
FLTriDiag(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{FLTriDiag} returns a FLMatrix object representing the upper Hessenberg or TriDiagonal matrix.
}
\description{
\code{FLTriDiag} computes the TriDiagonal or Upper Hessenberg matrix of FLMatrix object.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (700 x 700).
}

\examples{
flmatrix <- FLMatrix("tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- FLTriDiag(flmatrix)
}
