% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLogRegr.R
\name{glm.FLTable}
\alias{glm.FLTable}
\title{Linear Regression.}
\usage{
\method{glm}{FLTable}(formula, data, rushil = "binomial", iter = 25,
  threshold = 0.1, ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable}
}
\value{
\code{glm} performs linear regression and replicates equivalent R output.
}
\description{
\code{glm} performs linear regression on FLTable objects.
}
\details{
The wrapper overloads glm and implicitly calls FLRegrDataPrep and FLLogRegr.
}
\section{Constraints}{

None
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable(connection, "FL_REV4546", "tblAbaloneWide", "ObsID")
glmfit <- glm(Rings~Height+Diameter,widetable)
}

