% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLogRegr.R, R/FLPoissonRegr.R
\name{glm.FLTable}
\alias{glm.FLTable}
\title{Logistic and Poisson Regression.}
\usage{
\method{glm}{FLTable}(formula, family = stats::poisson, data, offset = "",
  maxiter = 25, catToDummy = 0, performNorm = 0, performVarReduc = 0,
  makeDataSparse = 0, minStdDev = 0, maxCorrel = 1, classSpec = list(),
  whereconditions = "", ...)

\method{glm}{FLTable}(formula, family = stats::poisson, data, offset = "",
  maxiter = 25, catToDummy = 0, performNorm = 0, performVarReduc = 0,
  makeDataSparse = 0, minStdDev = 0, maxCorrel = 1, classSpec = list(),
  whereconditions = "", ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{family}{Can be one of poisson,binomial,linear or multinomial.
Can be family functions like stats::poisson wherever possible.}

\item{data}{An object of class FLTable}

\item{maxiter}{maximum number of iterations.}

\item{catToDummy}{Transform categorical variables to numerical values
either using dummy variables or by using Empirical
Logit. If the value is 1, transformation is done using
dummy variables, else if the value is 0,
transformation is done using Empirical Logit.}

\item{performNorm}{0/1 indicating whether to perform standardization of data.}

\item{performVarReduc}{0/1. If the value is 1,
the stored procedure eliminates variables based on standard deviation and
correlation.}

\item{makeDataSparse}{If 0,Retains zeroes and NULL values
from the input table. If 1, Removes zeroes and NULL. If 2,Removes zeroes 
but retains NULL values.}

\item{minStdDev}{Minimum acceptable standard deviation for
elimination of variables. Any variable that has a
standard deviation below this threshold is
eliminated. This parameter is only consequential if
the parameter PerformVarReduc = 1. Must be >0.}

\item{maxCorrel}{Maximum acceptable absolute correlation between
a pair of columns for eliminating variables. If the
absolute value of the correlation exceeds this
threshold, one of the columns is not transformed.
Again, this parameter is only consequential if the
parameter PerformVarReduc = 1. Must be >0 and <=1.}

\item{classSpec}{list describing the categorical dummy variables.}

\item{whereconditions}{takes the where_clause as a string.}

\item{pThreshold}{The threshold for False positive value 
that a user can specify to calculate the false positives 
and false negatives. Must be between 0 and 1.}

\item{pRefLevel}{Reference value for dependent variable
in case of multinomial family.}
}
\value{
\code{glm} performs logistic 
or poisson regression and replicates equivalent R output.
}
\description{
\code{glm} performs logistic and poisson regression on FLTable objects.
}
\section{Constraints}{

The anova method is not yet available for FLLogRegr.
In case of multinomial family, residuals,fitted.values
properties are not available.plot,influence methods are
also not available.
}
\examples{
library(RODBC)
connection <- flConnect("Gandalf")
deeptable <- FLTable("FL_DEMO","tblLogRegr","ObsID","VarID","Num_Val",
               whereconditions="ObsID<7001")
glmfit <- glm(NULL,data=deeptable)
summary(glmfit)
plot(glmfit)
glmfit <- glm(NULL,data=deeptable,family="logisticwt",eventweight=0.8,noneventweight=1)
summary(glmfit)
plot(glmfit)
connection <- flConnect(odbcSource = "Gandalf",database = "FL_DEV")
widetable  <- FLTable("FL_DEV", "siemenswidetoday1", "ObsID")
poissonfit <- glm(event ~ meanTemp, family=poisson, data=widetable,offset="age")
summary(poissonfit)
plot(poissonfit)
predData <- FLTable("FL_DEV","preddata1","ObsID")
mu <- predict(poissonfit,newdata=predData)
deeptable <- FLTable("FL_DEMO","tblLogRegrMN10000","ObsID","VarID","Num_Val",
             whereconditions="ObsID<7001")
glmfit <- glm(NULL,data=deeptable,family="multinomial")
glmfit$coefficients
glmfit$FLLogRegrStats
glmfit$FLCoeffStdErr
summary(glmfit)
print(glmfit)
library(RODBC)
connection <- flConnect(odbcSource = "Gandalf",database = "FL_DEV")
widetable  <- FLTable("FL_DEV", "siemenswidetoday1", "ObsID")
options(debugSQL=T)
pfitToffA <- glm(event ~ meanTemp, family=poisson, data=widetable,offset="age")
predData <- FLTable("FL_DEV","preddata1","ObsID")
mu <- predict(pfitToffA,newdata=predData)
}

