% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{hamming.distance}
\alias{hamming.distance}
\title{hamming.distance}
\usage{
hamming.distance(xsource, targets, caseFlag = 0, vlength = 3, ...)
}
\arguments{
\item{xsource}{character or FLVector of characters}

\item{targets}{character or FLVector of characters}

\item{vlength}{optional, length of strings to compare}

\item{caseFLag}{logical or 0/1 indicating
if comparision should be case sensitive}
}
\value{
FLVector if any \code{xsource} or \code{targets}
is R character of length 1. Otherwise returns a FLMatrix.
}
\description{
computes the hamming distance between strings.
}
\details{
The DB Lytix function called is FLHammingDist.
This function computes the hamming distance between the 
two char string arguments(the number of non-overlapping characters). 
The DB Lytix function called is FLHammingDist.
}
\section{Constraints}{

row vectors are not supported currently.
Output is slightly different from vwr::hamming.
Refer to \code{@return} section.
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflvector <- hamming.distance("xyz",flv)
resultflmatrix <- hamming.distance(flv,flv,caseFLag=1)
resultflmatrix <- hamming.distance(flv,c("xyz","poli"),caseFLag=1)
}
\seealso{
\code{\link[vwr]{hamming.distance}} for R function reference
implementation.
}

