% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{kmeans.FLTable}
\alias{kmeans.FLTable}
\title{K-Means Clustering.}
\usage{
\method{kmeans}{FLTable}(x, centers, iter.max = 10, nstart = 1,
  excludeCols = as.character(c()), classSpec = list(),
  whereconditions = "")
}
\arguments{
\item{x}{an object of class FLTable, wide or deep}

\item{centers}{the number of clusters}

\item{iter.max}{the maximum number of iterations allowed}

\item{nstart}{the initial number of random sets}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{kmeans} returns a list which replicates equivalent R output
from \code{kmeans} in stats package
}
\description{
\code{kmeans} performs k-means clustering on FLTable objects.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
widetable  <- FLTable( "FL_DEMO", "tblAbaloneWide", "ObsID")
kmeansobject <- kmeans(widetable,3,20,1,"Rings,SEX",list("DummyCat(D)","SEX(M)"))
print(kmeansobject)
plot(kmeansobject)
}

