% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{kmeans.FLTable}
\alias{kmeans.FLTable}
\title{K-Means Clustering.}
\usage{
\method{kmeans}{FLTable}(table, centers, max.iter = 10, nstart = 1,
  exclude = as.character(c()), class_spec = list(), where_clause = "")
}
\arguments{
\item{table}{an object of class FLTable}

\item{centers}{the number of clusters}

\item{max.iter}{the maximum number of iterations allowed}

\item{nstart}{the initial number of random sets}

\item{exclude}{the comma separated character string of columns to be excluded}

\item{class_spec}{list describing the categorical dummy variables}

\item{where_clause}{takes the where_clause as a string}
}
\value{
\code{kmeans} performs k-means clustering and replicates equivalent R output.
}
\description{
\code{kmeans} performs k-means clustering on FLTable objects.
}
\details{
The wrapper overloads kmeans and implicitly calls FLKMeans.
}
\section{Constraints}{

None
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable(connection, "FL_TRAIN", "tblAbaloneWide", "ObsID")
kmeansobject <- kmeans(widetable,3,20,2,"Rings,SEX",list("DummyCat(D)","SEX(M)"))
print(kmeansobject)
}

