% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{kmeans.FLTable}
\alias{kmeans.FLTable}
\title{K-Means Clustering.}
\usage{
\method{kmeans}{FLTable}(x, centers, iter.max = 10, nstart = 1,
  excludeCols = as.character(c()), classSpec = list(),
  whereconditions = "")
}
\arguments{
\item{x}{an object of class FLTable, wide or deep}

\item{centers}{the number of clusters}

\item{iter.max}{the maximum number of iterations allowed}

\item{nstart}{the initial number of random sets}

\item{excludeCols}{the comma separated character string of columns to be excluded}

\item{classSpec}{list describing the categorical dummy variables}

\item{whereconditions}{takes the where_clause as a string}
}
\value{
\code{kmeans} returns a list which replicates equivalent R output
from \code{kmeans} in stats package
}
\description{
\code{kmeans} performs k-means clustering on FLTable objects.
}
\section{Constraints}{

If classSpec is not specified, the categorical variables are excluded
from analysis by default.
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
widetable  <- FLTable( "FL_DEMO", "tblAbaloneWide", "ObsID")
kmeansobject <- kmeans(widetable,3,20,1,"Rings,SEX")
print(kmeansobject)
plot(kmeansobject)
One can specify ClassSpec and transform categorical variables 
before clustering. This increases the number of variables in the plot
because categorical variable is split into binary numerical variables.
The clusters may not be well-defined as is observed in the case below:-
kmeansobjectnew <- kmeans(widetable,3,20,1,"Rings,SEX",list("DummyCat(D)","SEX(M)"))
plot(kmeansobjectnew)
}

