% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLKMeans.R
\name{kmeans.FLTable}
\alias{kmeans.FLTable}
\title{K-Means Clustering.}
\usage{
\method{kmeans}{FLTable}(table, centers, max.iter, isDeep = FALSE)
}
\arguments{
\item{table}{an object of class FLTable}

\item{centers}{the number of clusters}

\item{max.iter}{the maximum number of iterations allowed}

\item{isDeep}{optional for widetable}
}
\value{
\code{kmeans} performs k-means clustering and replicates equivalent R output.
}
\description{
\code{kmeans} performs k-means clustering on FLMatrix objects.
}
\details{
The wrapper overloads kmeans and implicitly calls FLKMeans.
}
\section{Constraints}{

None
}
\examples{
\dontrun{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable(connection, "FL_TRAIN", "tblAbaloneWide", "ObsID")
kmeans(widetable,3,20)
}
}

