% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLKNN.R
\name{knnx.index}
\alias{knnx.index}
\title{k-Nearest Neighbour Indexes
For each row of the test set, the k nearest (according to distance metric speicified) 
training set vectors are found and indices,distances are returned.}
\usage{
knnx.index(data, query, k = 1, algorithm = "kd_tree", ...)
}
\arguments{
\item{data}{input deep FLTable}

\item{query}{input deep FLTable}

\item{k}{number of neighbours considered.}

\item{...}{additional parameters like dist Flag to get the distances as well.}
}
\value{
FLVector of indices or a list of dist & index if dist FLag is true.
}
\description{
k-Nearest Neighbour Indexes
For each row of the test set, the k nearest (according to distance metric speicified) 
training set vectors are found and indices,distances are returned.
}
\examples{
FLdeepTbl <- FLTable(getTestTableName("ARknnDevSmall"),"obsid","varid","num_val")
FLknnOutput <- knnx.index(FLdeepTbl,FLdeepTbl,k=2)
FLknnOutput
}
