% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{levenshtein.distance}
\alias{levenshtein.distance}
\title{levenshtein.distance}
\usage{
levenshtein.distance(xsource, targets, caseFlag = 0, ...)
}
\arguments{
\item{xsource}{character or FLVector of characters}

\item{targets}{character or FLVector of characters}

\item{caseFLag}{logical or 0/1 indicating 
if comparision should be case sensitive}
}
\value{
FLMatrix if both \code{xsource} and \code{targets}
are FLVectors. Otherwise returns a FLVector
}
\description{
computes the levenshtein distance between strings
}
\section{Constraints}{

row vectors are not supported currently
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflvector <- levenshtein.distance("xyz",flv)
resultflmatrix <- levenshtein.distance(flv,flv,caseFLag=1)
resultflvector <- levenshtein.distance(flv,"xyz",caseFLag=1)
}

