% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{lm.FLTable}
\alias{lm.FLTable}
\title{Linear Regression.}
\usage{
\method{lm}{FLTable}(formula, data, ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable}
}
\value{
\code{lm} performs linear regression and replicates equivalent R output.
}
\description{
\code{lm} performs linear regression on FLTable objects.
}
\details{
The wrapper overloads lm and implicitly calls FLRegrDataPrep and FLLinRegr.
}
\section{Constraints}{

None
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
widetable  <- FLTable(connection, "FL_REV4546", "tblAbaloneWide", "ObsID")
lmfit <- lm(Rings~Height+Diameter,widetable)
}

