% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{lm}
\alias{lm}
\title{Linear Regression.}
\usage{
lm(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable or FLTableMD}

\item{catToDummy}{Transform categorical variables to numerical values
either using dummy variables or by using Empirical
Logit. If the value is 1, transformation is done using
dummy variables, else if the value is 0,
transformation is done using Empirical Logit.}

\item{performNorm}{0/1 indicating whether to perform standardization of data.}

\item{performVarReduc}{0/1. If the value is 1,
the stored procedure eliminates variables based on standard deviation and
correlation.}

\item{makeDataSparse}{If 0,Retains zeroes and NULL values
from the input table. If 1, Removes zeroes and NULL. If 2,Removes zeroes 
but retains NULL values.}

\item{minStdDev}{Minimum acceptable standard deviation for
elimination of variables. Any variable that has a
standard deviation below this threshold is
eliminated. This parameter is only consequential if
the parameter PerformVarReduc = 1. Must be >0.}

\item{maxCorrel}{Maximum acceptable absolute correlation between
a pair of columns for eliminating variables. If the
absolute value of the correlation exceeds this
threshold, one of the columns is not transformed.
Again, this parameter is only consequential if the
parameter PerformVarReduc = 1. Must be >0 and <=1.}

\item{classSpec}{list describing the categorical dummy variables.}

\item{whereconditions}{takes the where_clause as a string.}
}
\value{
\code{lm} returns an object of class \code{FLLinRegr}
}
\description{
\code{lm} performs linear regression on FLTable objects.
}
\details{
The DB Lytix function called is FLLinRegr. Performs Linear Regression and 
stores the results in predefined tables.
}
\section{Constraints}{

The anova method is not yet available for FLLinRegr
If \code{data} is FLTableMD, only single formula is accepted.
So input deeptable or deeptable produced after data preparation
should have same VarIDs'.
For FLTableMD data object, only coefficients and summary
methods are defined.Predict method on \code{FLTableMD}
\code{newdata} is not supported.
Properties like \code{print(x),model,plot} might take time as they
have to fetch data
}
\examples{
widetable  <- FLTable("tblAbaloneWide", "ObsID")
lmfit <- lm(Rings~Height+Diameter,widetable)
lmfit$coefficients
lmfit$fitted.values
plot(lmfit)
mu <- predict(lmfit,newdata=widetable)
deeptable <- FLTable("FL_DEMO","myLinRegrSmall","ObsID","VarID","Num_Val")
lmfit <- lm(NULL,deeptable)
summary(lmfit)
flMDObject <- FLTableMD(table="tblAutoMPGMD",
                      group_id_colname="GroupID",
                      obs_id_colname="ObsID",group_id = c(2,4))
vformula <- MPG~HorsePower+Displacement+Weight+Acceleration
lmfit <- lm(vformula,
           data=flMDObject)
coeffList <- coef(lmfit)
summaryList <- summary(lmfit)
}
\seealso{
\code{\link[stats]{lm}} for R reference implementation.
}

