% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMixed.R, R/FLMixedmodel.R
\name{lmer.FLTable}
\alias{lmer.FLTable}
\alias{lmer.FLTable}
\title{fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)}
\usage{
\method{lmer}{FLTable}(formula, data, fetchID = TRUE, maxiter = 10, ...)

\method{lmer}{FLTable}(formula, data, fetchID = TRUE, maxiter = 10, ...)
}
\description{
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl, pIntercept = 1)
flpred <- predict(flmod)
tbl  <- FLTable("tblMixedModelInt", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 |   RanVal1) + (1 | RanVal2 ), tbl)
flpred <- predict(flmod)

fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer(yVal ~ FixVal + (1 | RanVal), data = fltbl)
flpred <- predict(flmod)
}
