% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMixedmodel.R
\name{lmer}
\alias{lmer}
\title{Fit Linear Mixed-Effects Models}
\usage{
\method{residuals}{FLMix}(formula, data = list(), ...)
}
\arguments{
\item{formula}{A symbolic description of model to be fitted}

\item{data}{An object of class FLTable.}
}
\value{
\code{lmer} returns an object of class \code{FLMix}
}
\description{
\code{lmer} performs mixed model  on FLTable objects.
}
\details{
The DB Lytix function called is FLLinMixedModel. The mixed model extends the linear
model by allowing correlation and heterogeneous variances in the
covariance matrix of the residuals. Fit a linear mixed-effects model (LMM) to data,
FLLinMixedModel estimates the coefficients and covariance matrix of the mixed model
via the expectations.
}
\section{Slots}{

\describe{
\item{\code{results}}{cache list of results computed}

\item{\code{table}}{Input data object}
}}

\examples{
## One Random Effect.
fltbl  <- FLTable("tblMixedModel", "ObsID")
flmod <- lmer.FLTable(yVal ~ FixVal + (1 | RanVal), data = fltbl)
flpred <- predict(flmod)
}
\seealso{
\code{\link[lme4]{lmer}} for R reference implementation.
}
