% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLconstructSQL.R, R/FLSubsetting.R
\name{names.FLSimpleVector}
\alias{names.FLSimpleVector}
\alias{names.FLSimpleVector}
\title{Get names of a FLVector}
\usage{
\method{names}{FLSimpleVector}(x)

\method{names}{FLSimpleVector}(x)
}
\arguments{
\item{x}{FLVector}

\item{x}{FLSimpleVector}
}
\value{
character vector of names
of FLVector if exists. Else NULL

character vector of names
of FLSimpleVector if exists. Else NULL
}
\description{
Get names of a FLVector

Get names of a FLSimpleVector
}
