% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_all}
\alias{read_all}
\title{Read all users logs from a data directory}
\usage{
read_all(data_dir)
}
\arguments{
\item{data_dir}{Directory containing all group data  directories. Must be
organised in particular way.}
}
\value{
List of user log-file data lists
}
\description{
Read all users logs from a data directory
}

