% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLLinRegr.R
\name{rlm}
\alias{rlm}
\title{Robust Regression.}
\usage{
rlm(formula, data = list(), psi, ...)
}
\description{
performs robust regression
}
\examples{
Example for deeptbl:
library(MASS)
options(debugSQL =TRUE)
table  <- FLTable(getTestTableName("tblRobustRegr"), "ObsID","VarID", "Num_Val")
flmod <- rlm(a~., data = table)
predict(flmod)
residuals(flmod)
flmod$fitted.values
summary(flmod)
}
