% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLROC.R
\name{roc.FLVector}
\alias{roc.FLVector}
\title{tbl <- FLTable("tblROCCurve", "ObsID")
mod <- roc(tbl$ActualVal, tbl$ProbVal)
Example 2:
data(aSAH)
fltbl <- data.frame(res = aSAH$outcome, pred = aSAH$s100b)
fltbl$res <- as.numeric(fltbl$res)
fltbl$res <- fltbl$res - 1
fltbl <- fltbl[-55, ]
head(fltbl)
fltbl <- as.FLTable(fltbl)
flmod <- roc(fltbl$res, fltbl$pred)}
\usage{
\method{roc}{FLVector}(response, predictor, ...)
}
\description{
tbl <- FLTable("tblROCCurve", "ObsID")
mod <- roc(tbl$ActualVal, tbl$ProbVal)
Example 2:
data(aSAH)
fltbl <- data.frame(res = aSAH$outcome, pred = aSAH$s100b)
fltbl$res <- as.numeric(fltbl$res)
fltbl$res <- fltbl$res - 1
fltbl <- fltbl[-55, ]
head(fltbl)
fltbl <- as.FLTable(fltbl)
flmod <- roc(fltbl$res, fltbl$pred)
}
