% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRowMeans.R
\name{rowMeans}
\alias{rowMeans}
\title{row means of a FLMatrix.}
\usage{
rowMeans(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix.}
}
\value{
\code{rowMeans} returns a FLVector object representing the row-wise Means.
}
\description{
\code{rowMeans} computes the row-wise average of FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- rowMeans(flmatrix)
}

