% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRowSums.R
\name{rowSums}
\alias{rowSums}
\title{row sums of a FLMatrix.}
\usage{
rowSums(object, ...)
}
\arguments{
\item{object}{is of class FLMatrix.}

\item{...}{any additional arguments}
}
\value{
\code{rowSums} returns a FLVector object representing the row-wise sums.
}
\description{
\code{rowSums} computes the row-wise sums of FLMatrix objects.
}
\examples{
connection <- RODBC::odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", 
resultFLVector <- rowSums(flmatrix)
}

