% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRowSums.R
\name{rowSums}
\alias{rowSums}
\title{row sums of a FLMatrix.}
\usage{
rowSums(x, ...)
}
\arguments{
\item{x}{is of class FLMatrix.}
}
\value{
\code{rowSums} returns a FLVector object representing the row-wise sums.
}
\description{
\code{rowSums} computes the row-wise sums of FLMatrix objects.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- rowSums(flmatrix)
}
\author{
Phani Srikar <phanisrikar93ume@gmail.com>
}

