% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLSolve.R
\name{solve.FLMatrix}
\alias{solve.FLMatrix}
\title{Inverse of a Matrix.}
\usage{
\method{solve}{FLMatrix}(object)
}
\arguments{
\item{object}{is of class FLMatrix}
}
\value{
\code{solve} returns a FLMatrix object which is the inverse of input FLMatrix object
and replicates the equivalent R output.
}
\description{
\code{solve} computes the inverse for FLMatrix objects.
}
\details{
The wrapper overloads solve and implicitly calls FLMatrixInvUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_TRAIN", "tblMatrixMulti", 1)
resultFLMatrix <- solve(flmatrix)
}

