% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSolve.R
\name{solve}
\alias{solve}
\title{Inverse of a Matrix.}
\usage{
solve(a, b = NULL, ...)
}
\arguments{
\item{...}{any additional arguments}

\item{x}{is of class FLMatrix}
}
\value{
\code{solve} returns a FLMatrix object which is the inverse of input FLMatrix object
and replicates the equivalent R output.
}
\description{
\code{solve} computes the inverse for FLMatrix objects.
}
\details{
The wrapper overloads solve and implicitly calls FLMatrixInvUdt.
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (1000 x 1000).
}

\examples{
flmatrix <- FLMatrix("tblMatrixMulti", 2,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLMatrix <- solve(flmatrix)
}
