% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlQuery}
\alias{sqlQuery}
\title{Send a query to database}
\usage{
sqlQuery(connection, query)
}
\arguments{
\item{query}{SQLQuery to be sent}

\item{channel}{ODBC/JDBC connection object}
}
\description{
Result is returned as data.frame
}

