% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlSendUpdate.JDBCConnection}
\alias{sqlSendUpdate.JDBCConnection}
\title{Send a query to database}
\usage{
\method{sqlSendUpdate}{JDBCConnection}(connection, query, warn = TRUE, ...)
}
\arguments{
\item{query}{SQLQuery to be sent}

\item{channel}{JDBC connection object}
}
\description{
No result is returned
}
