% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sqlSendUpdate}
\alias{sqlSendUpdate}
\title{Send a query to database}
\usage{
sqlSendUpdate(connection, query, ...)
}
\arguments{
\item{query}{SQLQuery to be sent}

\item{channel}{JDBC connection object}
}
\description{
No result is returned
}

