% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{store}
\alias{store}
\title{stores a matrix in a table.
TODO:  define when data is stored (automatic caching, user requests...)}
\usage{
store(object, returnType, connection, ...)
}
\value{
A FLMatrix based on a stored table of the executed
}
\description{
stores a matrix in a table.
TODO:  define when data is stored (automatic caching, user requests...)
}
\author{
Gregor Kappler <g.kappler@gmx.net>
}

