% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStore.R
\name{store}
\alias{store}
\title{stores an object in database.}
\usage{
store(object, returnType, ...)
}
\arguments{
\item{object}{the object to store. Can be FLMatrix, FLVector, FLTable, character}

\item{returnType}{return type of the stored data. Applicable only when object is a character representing a SQL Query}

\item{connection}{ODBC/JDBC connection  object. Applicable only when object is a character representing a SQL Query}
}
\value{
in-database object after storing
}
\description{
stores an object in database.
}

