% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStringFunctions.R
\name{stringdist}
\alias{stringdist}
\title{stringdist}
\usage{
stringdist(a, b, method = "dl", caseFlag = 0, p = 0, vlength = 3, ...)
}
\arguments{
\item{a}{character or FLVector of characters}

\item{b}{character or FLVector of characters}

\item{method}{can be \code{c("lv","dl","hamming","jaccard","jw")}
where lv - levenshtein, dl - levenshtein.damerau
jw - jaro-winkler. Default is "lv"}

\item{p}{penality factor for jaro-winkler
if p==0 jaro distance is computed}

\item{vlength}{optional, length of strings to compare
used for hamming}

\item{caseFLag}{logical or 0/1 indicating
if comparision should be case sensitive}
}
\value{
FLVector if any \code{a} or \code{b}
is R character of length 1. Otherwise returns a FLMatrix.
}
\description{
compute distance metrics between strings
}
\section{Constraints}{

row vectors are not supported currently.
Output is slightly different from stringdist::stringdist.
Refer to \code{@return} section.
}
\examples{
widetable  <- FLTable("FL_DEMO", "iris", "rownames")
flv <- widetable[1:10,"Species"]
resultflvector <- stringdist("xyz",flv)
resultflvector <- stringdist("xyz",flv,method="lv",caseFLag=1)
resultflvector <- stringdist("xyz",flv,method="hamming",vlength=4)
resultflmatrix <- stringdist(flv,flv,method="jw",p=1)
resultflmatrix <- stringdist(c("xyz","poli"),flv,method="jw")
}

