% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{[.FLMatrix}
\alias{[.FLMatrix}
\title{Extract part of FLMatrix object.}
\usage{
\method{[}{FLMatrix}(object, rows = 1, cols = 1, drop = TRUE)
}
\arguments{
\item{object}{is a FLMatrix object}

\item{rows}{is a vector input corresponding to rows to be extracted}

\item{cols}{is a vector input corresponding to columns to be extracted}
}
\value{
\code{[]} returns FLMatrix object after extraction
which replicates the equivalent R extraction.
}
\description{
\code{[]} acts on FLMatrix objects and extracts parts of them.
}
\section{Constraints}{

Applying UDT functions on subsetted matrices with discontinuous row and col ids' 
may result in error
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 2,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLmatrix <- flmatrix[1,]
}

