% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{[.FLTable}
\alias{[.FLTable}
\title{Extract part of FLTable object.}
\usage{
\method{[}{FLTable}(object, rows = 1, cols = 1, drop = TRUE)
}
\arguments{
\item{object}{is a FLTable object}

\item{rows}{is a vector input corresponding to rows to be extracted}

\item{cols}{is a vector input corresponding to columns to be extracted}
}
\value{
\code{[]} returns FLMatrix object after extraction
which replicates the equivalent R extraction.
}
\description{
\code{[]} acts on FLMatrix objects and extracts parts of them.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
fltable <- FLTable(connection, "FL_DEMO", "tblAbaloneWide", "ObsID")
resultFLtable <- fltable[1:10,4:6]
}

