% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSubsetting.R
\name{[.FLVector}
\alias{[.FLVector}
\title{Extract part of FLVector object.}
\usage{
\method{[}{FLVector}(object, pSet = 1:length(object))
}
\arguments{
\item{pSet}{is a vector representing the indices of elements to extract}

\item{pObj}{is a FLVector object}
}
\value{
\code{[]} returns FLVector object after extraction
which replicates the equivalent R extraction.
}
\description{
\code{[]} acts on FLVector objects and extracts parts of them.
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
WideTable <- FLTable(connection, "FL_TRAIN", "tblVectorWide","vector_key")
flvectorWide <- FLVector(WideTable,"vector_value")
resultFLVector <- flvectorWide[1:2]
DeepTable <- FLTable(connection, "FL_TRAIN", "tblVectorDeep","vector_id","vector_key","vector_value")
flvectorDeep <- FLVector(DeepTable,"vector_value",1)
resultFLVector <- flvectorDeep[1:2]
}

