% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTrace.R
\name{tr}
\alias{tr}
\title{Matrix Trace.}
\usage{
tr(x, ...)
}
\arguments{
\item{x}{an object of class FLMatrix}
}
\value{
\code{tr} returns R Vector object of size 1 which replicates the equivalent R output.
}
\description{
\code{tr} computes the trace of FLMatrix objects.
}
\details{
\code{tr} computes the trace of input FLMatrix object, stores the result
in-database and returns R vector object
}
\section{Constraints}{

Input can only be with maximum dimension limitations
of (1000 x 1000).
}
\examples{
library(RODBC)
connection <- odbcConnect("Gandalf")
flmatrix <- FLMatrix(connection, "FL_DEMO", "tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- tr(flmatrix)
}

