% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTrace.R
\name{tr}
\alias{tr}
\title{Matrix Trace.}
\usage{
tr(object, ...)
}
\arguments{
\item{object}{an object of class FLMatrix}

\item{...}{any additional arguments}
}
\value{
\code{tr} returns R Vector object of size 1 which replicates the equivalent R output.
}
\description{
\code{tr} computes the trace of FLMatrix objects.
}
\details{
\code{tr} computes the trace of input FLMatrix object, stores the result
in-database and returns R vector object
}
\section{Constraints}{

Input can only be with maximum dimension limitations
of (1000 x 1000).
}
\examples{
connection <- flConnect(odbcSource="Gandalf")
flmatrix <- FLMatrix("FL_DEMO.tblMatrixMulti", 5,"MATRIX_ID","ROW_ID","COL_ID","CELL_VAL")
resultFLVector <- tr(flmatrix)
}

