% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLTable.R
\name{wideToDeep}
\alias{wideToDeep}
\title{Convert Wide Table to Deep Table in database.}
\usage{
wideToDeep(object, ...)
}
\arguments{
\item{object}{FLTable object}

\item{excludeCols}{character vector specifying columns to be excluded from conversion}

\item{classSpec}{list representing Class specification which identifies then value of the categorical variable to be used a reference}

\item{whereconditions}{character vector giving where conditions if any to reference the wide table}

\item{outDeepTableName}{name to be given to the output deep table, possibly with database}

\item{outObsIDCol}{name to give to the primary key column name of the output deep table}

\item{outVarIDCol}{name to give to the varibales name column of the output deep table}

\item{outValueCol}{name to give to the value column of the output deep table}
}
\value{
\code{wideToDeep} returns a list containing components 
\code{table} which is the FLTable referencing the deep table and \code{AnalysisID} giving the AnalysisID of conversion
}
\description{
The DB Lytix function called is FLWideToDeep.  FLWideToDeep is similar to FLRegrDataPrep , except that 
FLWideToDeep does not require dependent variable or intercept, thus it creates deep tables without them.
}
\examples{
widetable  <- FLTable("tblAbaloneWide", "ObsID")
deeptable <- wideToDeep(widetable)
analysisID <- deeptable@wideToDeepAnalysisID
}

