% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z.test}
\alias{z.test}
\title{Z test for 1 and 2 samples}
\usage{
z.test(x, y = NULL, test_val = 0, tails = 2, conf.level = 0.95,
  prob = 0, ...)
}
\arguments{
\item{x}{FLVector}

\item{y}{FLVector}

\item{test_val}{The mean value for which the z test has to be performed.
The default value is 0.}

\item{prob}{The values is 1 if the sample set is for proportions. Otherwise, by default it's 0.}
}
\value{
A list with class "htest" outputting the corresponding Z Stat and P Values.
}
\description{
Performs z test for samples and sample proportions on vector values
}
\examples{
flx<-as.FLVector(rnorm(100))
fly<-as.FLVector(rnorm(100))
z.test(flx)
    ## Two- sided one sample z-test.
    ## mean for 'x' is zero. The alternative hypothesis states 
    ## that it is either greater or less than zero.

z.test(flx,fly,2)
    ## Two- sided two sample z-test.
    ## The null hypothesis is that the population mean for 'x'
    ## less that for 'y' is 2. 
    ## The alternative hypothesis is that this difference is not 2.

## z- test for population proportion.
fla<-as.FLVector(sample(c(0,1),replace=T,size=100))
flb<-as.FLVector(sample(c(0,1),replace=T,size=100))
 ##test_val= 0.4 implies that null hypothesis is
 ##population proportion is 0.4
z.test(fla,prob=1, test_val= 0.4)
z.test(fla,flb,prob=1)
}
